@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'اذافة :entity',
   tooLong: 'قيمة الحقل طويلة جداً يجب أن تتضمن :max من الحروف والأرقام',
   invalidFormat: 'تحتوي قيمة الحقل على تنسيق غير صالح أو أحرف غير مسموح بها.',
-  confirmChangeLibrary: 'من خلال القيام بذلك سوف تفقد كل العمل المنجز مع نوع المحتوى الحالي. هل تريد بالتأكيد تغيير نوع المحتوى؟',
+  confirmChangeLibrary:
+    'من خلال القيام بذلك سوف تفقد كل العمل المنجز مع نوع المحتوى الحالي. هل تريد بالتأكيد تغيير نوع المحتوى؟',
   commonFields: 'تجاوزات النصوص والترجمة',
   commonFieldsDescription: 'يمكنك هنا تعديل الإعدادات أو ترجمة النصوص المستخدمة في هذا المحتوى.',
   uploading: 'جاري التحميل .. الرجاء الانتظار...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'ألصق رابط اليوتيوب أو عنوان مصدر الفيديو الاخر',
   uploadAudioTitle: 'رفع ملف الصوت',
   uploadVideoTitle: 'رفع ملف الفيديو',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'ادراج',
   cancel: 'الغاء',
   height: 'ارتفاع',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'رفع',
   uploadPlaceholder: 'لم يتم اختيار ملف',
   uploadInstructionsTitle: 'رفع ملف',
-  uploadInstructionsContent: 'يمكن أن تجد أمثلة في <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'يمكن أن تجد أمثلة في <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'رفع ملف',
   uploadFileButtonChangeLabel: 'تغيير الملف',
   uploadingThrobber: 'يتم الرفع الان...',
@@ -84,7 +87,8 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'تعذر تحميل الملف المحدد',
   h5pFileWrongExtensionContent: 'فقط الملفات بالامتداد .h5p مسموحة.',
   h5pFileValidationFailedTitle: 'تعذر التحقق من الملف.',
-  h5pFileValidationFailedContent: 'تأكد أن الملف الذي تم رفعه يحوي على محتوى صالح. H5P يجب تحميل الملفات التي تحتوي على مكتبات فقط من خلال صفحة مكتبات H5P.',
+  h5pFileValidationFailedContent:
+    'تأكد أن الملف الذي تم رفعه يحوي على محتوى صالح. H5P يجب تحميل الملفات التي تحتوي على مكتبات فقط من خلال صفحة مكتبات H5P.',
   h5pFileUploadServerErrorTitle: 'الملف لا يمكن تحميله',
   h5pFileUploadServerErrorContent: 'حدث خطأ غير متوقع. تحقق من سجل أخطاء السيرفر  more details.',
   contentTypeSectionAll: 'الكل',
@@ -153,9 +157,11 @@ H5PEditor.language.core = {
   videoQualityDescription: 'يساعد هذا التصنيف المستخدم في تحديد الجودة الحالية للفيديو. مثلا 1080p، 720p، HD أو موبايل',
   videoQualityDefaultLabel: 'الجودة :index',
   noContentTypesAvailable: 'No content types are available',
-  noContentTypesAvailableDesc: 'Your site is having difficulties connecting to H5P.org and listing the available content types.',
+  noContentTypesAvailableDesc:
+    'Your site is having difficulties connecting to H5P.org and listing the available content types.',
   contentTypeCacheOutdated: 'Content type list outdated',
-  contentTypeCacheOutdatedDesc: 'Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.',
+  contentTypeCacheOutdatedDesc:
+    'Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.',
   tryAgain: 'Try again',
   getHelp: 'Get help',
   untitled: 'Untitled :libraryTitle',
@@ -170,20 +176,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
   addNewChange: 'Add new change',
   confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
   logThisChange: 'Log this change',
   newChangeHasBeenLogged: 'New change has been logged',
   loggedChanges: 'Logged changes',
   noChangesHaveBeenLogged: 'No changes have been logged',
   errorHeader: 'An error occured',
-  errorCalculatingMaxScore: 'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
+  errorCalculatingMaxScore:
+    'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
   maxScoreSemanticsMissing: 'Could not find the expected semantics in the content.',
   copyButton: 'Copy',
   copiedButton: 'Copied',
   pasteButton: 'Paste',
   pasteAndReplaceButton: 'Paste & Replace',
   pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
+  confirmPasteContent:
+    'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
   confirmPasteButtonText: 'Replace content',
   copyToClipboard: 'Copy H5P content to the clipboard',
   copiedToClipboard: 'Content is copied to the clipboard',
@@ -193,8 +202,10 @@ H5PEditor.language.core = {
   pasteError: 'Cannot paste from clipboard',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insert using',
   tabTitleBasicFileUpload: 'File Upload',
@@ -207,14 +218,86 @@ H5PEditor.language.core = {
   language: 'Language',
   noLanguagesSupported: 'No languages supported',
   changeLanguage: 'Change language to :language?',
-  thisWillPotentially: "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
+  thisWillPotentially:
+    "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
   notAllTextsChanged: 'Not all texts were changed, there is only partial coverage for :language.',
-  contributeTranslations: 'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Unfortunately, the selected content type \'%lib\' isn\'t installed on this system.',
+  contributeTranslations:
+    'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Unfortunately, the selected content type '%lib' isn't installed on this system.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
