@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Dodati :entity',
   tooLong: 'Sadržaj polja je previše dug. Može sadržavati :max slova ili manje od toga.',
   invalidFormat: 'Sadržaj polja sadrži netačan format ili znakove koji nisu dozvoljeni.',
-  confirmChangeLibrary: 'Ako ovo uradite ovaj interaktivni sadržaj će biti izgubljen. Jeste li sigurni da želite promijeniti vrstu sadržaja?',
+  confirmChangeLibrary:
+    'Ako ovo uradite ovaj interaktivni sadržaj će biti izgubljen. Jeste li sigurni da želite promijeniti vrstu sadržaja?',
   commonFields: 'Prijevod teksta',
   commonFieldsDescription: 'Ovdje možete urediti podešavanja ili ili tekst korišten u ovoj sadržaju.',
   uploading: 'Učitava se, molimo sačekajte...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Prenesite YouTube link ili izvorni URL audio fajla',
   uploadAudioTitle: 'Upload audio fajla',
   uploadVideoTitle: 'Upload video fajla',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Ubaciti',
   cancel: 'Otkazati',
   height: 'visina',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Upload',
   uploadPlaceholder: 'Nijedan fajl nije odabran',
   uploadInstructionsTitle: 'Upload H5P fajla.',
-  uploadInstructionsContent: 'Možete početi koristeći se sa primjerima na <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Možete početi koristeći se sa primjerima na <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Uploadujte fajla',
   uploadFileButtonChangeLabel: 'Izmjenite fajl',
   uploadingThrobber: 'Sada se uploaduje...',
@@ -84,7 +87,8 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Označeni fajl nije moguće uploadovati',
   h5pFileWrongExtensionContent: 'Samo fajlovi sa .h5p ekstenzijom su dozvoljeni.',
   h5pFileValidationFailedTitle: 'Nije moguće verificirati H5P fajl.',
-  h5pFileValidationFailedContent: 'Budite sigurni da uploadovani H5P sadrži ispravni H5P sadržaj. H5P fajlovi koji sadrže biblioteke trebaju biti uploadovani samo preko stranice za H5P biblioteke.',
+  h5pFileValidationFailedContent:
+    'Budite sigurni da uploadovani H5P sadrži ispravni H5P sadržaj. H5P fajlovi koji sadrže biblioteke trebaju biti uploadovani samo preko stranice za H5P biblioteke.',
   h5pFileUploadServerErrorTitle: 'H5P fajl nije moguće uploadovati',
   h5pFileUploadServerErrorContent: 'Neočekivana greška se pojavila. Provjerite vaš server za greške više detalja.',
   contentTypeSectionAll: 'Sve',
@@ -127,10 +131,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Nije u stanju komunicirati sa hubom.',
   errorCommunicatingHubContent: 'Javila se greška. Molimo pokušajte ponovo.',
   warningNoContentTypesInstalled: 'Nemate instaliranu nijednu vrstu sadržaja.',
-  warningChangeBrowsingToSeeResults: 'Kliknite na <em>Sve</em> da dobijete listu svih vrsta sadržaja koje možete instalirati',
+  warningChangeBrowsingToSeeResults:
+    'Kliknite na <em>Sve</em> da dobijete listu svih vrsta sadržaja koje možete instalirati',
   warningUpdateAvailableTitle: 'Dostupna je nova verzija za sadržaj :contentType.',
   warningUpdateAvailableBody: 'Instalirajte najnoviju verziju za bolje iskustvo sadržaja.',
-  licenseDescription: 'Neke osobine licence su naznačene ispod. Kliknite na info-sličicu iznad da pročitate tekst originalne licence.',
+  licenseDescription:
+    'Neke osobine licence su naznačene ispod. Kliknite na info-sličicu iznad da pročitate tekst originalne licence.',
   licenseModalTitle: 'Detalji o licenci',
   licenseModalSubtitle: 'Označite licencu da vidite odgovarajuću upotrebu',
   licenseUnspecified: 'Neodređeno',
@@ -150,12 +156,15 @@ H5PEditor.language.core = {
   screenshots: 'Snimci ekrana',
   reloadButtonLabel: 'Ponovno učitavanje',
   videoQuality: 'Oznaka kvalitete videa',
-  videoQualityDescription: 'Ova oznaka pomaže korisniku da zna trenutnu kvalitetu videa. Npr. 1080p, 720p, HD or Mobile',
+  videoQualityDescription:
+    'Ova oznaka pomaže korisniku da zna trenutnu kvalitetu videa. Npr. 1080p, 720p, HD or Mobile',
   videoQualityDefaultLabel: 'Kvalitet :index',
   noContentTypesAvailable: 'No content types are available',
-  noContentTypesAvailableDesc: 'Your site is having difficulties connecting to H5P.org and listing the available content types.',
+  noContentTypesAvailableDesc:
+    'Your site is having difficulties connecting to H5P.org and listing the available content types.',
   contentTypeCacheOutdated: 'Content type list outdated',
-  contentTypeCacheOutdatedDesc: 'Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.',
+  contentTypeCacheOutdatedDesc:
+    'Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.',
   tryAgain: 'Try again',
   getHelp: 'Get help',
   untitled: 'Untitled :libraryTitle',
@@ -170,20 +179,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
   addNewChange: 'Add new change',
   confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
   logThisChange: 'Log this change',
   newChangeHasBeenLogged: 'New change has been logged',
   loggedChanges: 'Logged changes',
   noChangesHaveBeenLogged: 'No changes have been logged',
   errorHeader: 'An error occured',
-  errorCalculatingMaxScore: 'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
+  errorCalculatingMaxScore:
+    'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
   maxScoreSemanticsMissing: 'Could not find the expected semantics in the content.',
   copyButton: 'Copy',
   copiedButton: 'Copied',
   pasteButton: 'Paste',
   pasteAndReplaceButton: 'Paste & Replace',
   pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
+  confirmPasteContent:
+    'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
   confirmPasteButtonText: 'Replace content',
   copyToClipboard: 'Copy H5P content to the clipboard',
   copiedToClipboard: 'Content is copied to the clipboard',
@@ -193,8 +205,10 @@ H5PEditor.language.core = {
   pasteError: 'Cannot paste from clipboard',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insert using',
   tabTitleBasicFileUpload: 'File Upload',
@@ -207,14 +221,86 @@ H5PEditor.language.core = {
   language: 'Language',
   noLanguagesSupported: 'No languages supported',
   changeLanguage: 'Change language to :language?',
-  thisWillPotentially: "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
+  thisWillPotentially:
+    "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
   notAllTextsChanged: 'Not all texts were changed, there is only partial coverage for :language.',
-  contributeTranslations: 'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Unfortunately, the selected content type \'%lib\' isn\'t installed on this system.',
+  contributeTranslations:
+    'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Unfortunately, the selected content type '%lib' isn't installed on this system.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
