@@ -1,5 +1,5 @@
 H5PEditor.language.core = {
-   missingTranslation: '[Chybí překlad :key]',
+  missingTranslation: '[Chybí překlad :key]',
   loading: 'Načítám, čekejte prosím ...',
   selectLibrary: 'Vyberte knihovnu, kterou chcete pro svůj obsah použít.',
   unknownFieldPath: 'Nelze najít ":path".',
@@ -54,7 +54,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Vložte odkaz na YouTube nebo jinou adresu URL zdroje videa',
   uploadAudioTitle: 'Nahrajte zvukový soubor',
   uploadVideoTitle: 'Nahrajte soubor videa',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Vložit',
   cancel: 'Zrušit',
   height: 'výška',
@@ -74,7 +75,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Nahrát',
   uploadPlaceholder: 'Nebyl vybrán soubor',
   uploadInstructionsTitle: 'Nahrát H5P soubor.',
-  uploadInstructionsContent: 'Můžete začít příklady z <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Můžete začít příklady z <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Nahrát soubor',
   uploadFileButtonChangeLabel: 'Změnit soubor',
   uploadingThrobber: 'Nyní se nahrává...',
@@ -84,11 +86,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Vybraný soubor nelze nahrát',
   h5pFileWrongExtensionContent: 'Povoleny jsou pouze soubory s příponou .h5p.',
   h5pFileValidationFailedTitle: 'Nelze ověřit soubor H5P.',
-  h5pFileValidationFailedContent: 'Ujistěte se, že nahraný H5P obsahuje platný obsah H5P. H5P' +
-  ' soubory obsahující pouze knihovny by měly být nahrány přes stránku Knihovny H5P.',
+  h5pFileValidationFailedContent:
+    'Ujistěte se, že nahraný H5P obsahuje platný obsah H5P. H5P soubory obsahující pouze knihovny by měly být nahrány přes stránku Knihovny H5P.',
   h5pFileUploadServerErrorTitle: 'The H5P file could not be uploaded',
-  h5pFileUploadServerErrorContent: 'Došlo k neočekávané chybě. Zkontrolujte protokol chyb serveru' +
-  ' kde je více informací.',
+  h5pFileUploadServerErrorContent:
+    'Došlo k neočekávané chybě. Zkontrolujte protokol chyb serveru kde je více informací.',
   contentTypeSectionAll: 'Všechny typy obsahu',
   searchResults: 'Search Results',
   contentTypeSearchFieldPlaceholder: 'Vyhledejte typy obsahu',
@@ -128,11 +130,13 @@ H5PEditor.language.core = {
   currentMenuSelected: 'aktuální výběr',
   errorCommunicatingHubTitle: 'Nelze komunikovat s hubem.',
   errorCommunicatingHubContent: 'Nastala chyba. Prosím zkuste to znov.',
-  warningNoContentTypesInstalled: "YNemáte nainstalován žádný typ obsahu.",
-  warningChangeBrowsingToSeeResults: 'Kliknutím na <em> Vše </em> získáte seznam všech typů obsahu, které můžete nainstalovat.',
+  warningNoContentTypesInstalled: 'YNemáte nainstalován žádný typ obsahu.',
+  warningChangeBrowsingToSeeResults:
+    'Kliknutím na <em> Vše </em> získáte seznam všech typů obsahu, které můžete nainstalovat.',
   warningUpdateAvailableTitle: 'K dispozici je nová verze :contentType .',
   warningUpdateAvailableBody: 'Aktualizace na nejnovější verzi pro lepší zážitek.',
-  licenseDescription: 'Některé funkce této licence jsou uvedeny níže. Kliknutím na ikonu informace výše zobrazíte původní text licence.',
+  licenseDescription:
+    'Některé funkce této licence jsou uvedeny níže. Kliknutím na ikonu informace výše zobrazíte původní text licence.',
   licenseModalTitle: 'Detail licences',
   licenseModalSubtitle: 'Vyberte licenci pro zobrazení informací o správném použití',
   licenseUnspecified: 'Nespecifikováno',
@@ -152,12 +156,14 @@ H5PEditor.language.core = {
   screenshots: 'Ukázky',
   reloadButtonLabel: 'Znovu načíst',
   videoQuality: 'Popisek kvality videa',
-  videoQualityDescription: 'Tento popisek pomáhá uživateli zjistit aktuální kvalitu videa. Např. 1080p, 720p, HD nebo Mobile',
+  videoQualityDescription:
+    'Tento popisek pomáhá uživateli zjistit aktuální kvalitu videa. Např. 1080p, 720p, HD nebo Mobile',
   videoQualityDefaultLabel: 'Kvalita :index',
   noContentTypesAvailable: 'Nejsou k dispozici žádné typy obsahu',
   noContentTypesAvailableDesc: 'Váš web má potíže s připojením k H5P.org a se seznamem dostupných typů obsahu.',
   contentTypeCacheOutdated: 'Seznam typů obsahu je zastaralý',
-  contentTypeCacheOutdatedDesc: 'Váš web má potíže s připojením k serveru H5P.org, aby mohl zkontrolovat aktualizace typu obsahu. Možná nebudete moci aktualizovat nebo instalovat nové typy obsahu.',
+  contentTypeCacheOutdatedDesc:
+    'Váš web má potíže s připojením k serveru H5P.org, aby mohl zkontrolovat aktualizace typu obsahu. Možná nebudete moci aktualizovat nebo instalovat nové typy obsahu.',
   tryAgain: 'Zkusit znovu',
   getHelp: 'Získat pomoc',
   untitled: 'Nepojmenovaná :libraryTitle',
@@ -166,26 +172,29 @@ H5PEditor.language.core = {
   addTitle: 'Přidat název',
   usedForSearchingReportsAndCopyrightInformation: 'Používá se pro vyhledávání, zprávy a informace o autorských právech',
   metadataSharingAndLicensingInfo: 'Metadata (informace o sdílení a licenci)',
-  fillInTheFieldsBelow : 'Vyplnit v poli níže',
+  fillInTheFieldsBelow: 'Vyplnit v poli níže',
   saveMetadata: 'Uložit metadata',
   addAuthor: 'Uložit autora',
   confirmRemoveAuthor: 'Opravdu chcete tohoto autora odebrat?',
   addNewChange: 'Přidat novou změnu',
   confirmDeleteChangeLog: 'Opravdu chcete odstranit tuto položku protokolu změn?',
-  changelogDescription: 'Některé licence vyžadují, aby změny provedené v původním díle nebo derivátech byly zaznamenány a zobrazeny. Můžete zde protokolovat své změny z licenčních důvodů nebo jen proto, abyste sami a ostatní mohli sledovat změny provedené v tomto obsahu.',
+  changelogDescription:
+    'Některé licence vyžadují, aby změny provedené v původním díle nebo derivátech byly zaznamenány a zobrazeny. Můžete zde protokolovat své změny z licenčních důvodů nebo jen proto, abyste sami a ostatní mohli sledovat změny provedené v tomto obsahu.',
   logThisChange: 'Protokolovat tuto změnu',
   newChangeHasBeenLogged: 'Byla zaznamenána nová změna',
   loggedChanges: 'Protokolované změny',
   noChangesHaveBeenLogged: 'Nebyly zaznamenány žádné změny',
   errorHeader: 'Nastala chyba',
-  errorCalculatingMaxScore: 'Nelze vypočítat maximální skóre pro tento obsah. Předpokládá se, že maximální skóre je 0. Pokud to není správné, kontaktujte svého správce.',
+  errorCalculatingMaxScore:
+    'Nelze vypočítat maximální skóre pro tento obsah. Předpokládá se, že maximální skóre je 0. Pokud to není správné, kontaktujte svého správce.',
   maxScoreSemanticsMissing: 'Nelze najít očekávanou sémantiku v obsahu.',
   copyButton: 'Kopírovat',
   copiedButton: 'Zkopírováno',
   pasteButton: 'Vložit',
   pasteAndReplaceButton: 'Vložit a nahradit',
   pasteContent: 'Nahradit obsah',
-  confirmPasteContent: 'Tímto způsobem nahradíte aktuální obsah obsahem ze své schránky. Aktuální obsah bude ztracen. Opravdu chcete pokračovat?',
+  confirmPasteContent:
+    'Tímto způsobem nahradíte aktuální obsah obsahem ze své schránky. Aktuální obsah bude ztracen. Opravdu chcete pokračovat?',
   confirmPasteButtonText: 'Nahradit obsah',
   copyToClipboard: 'Zkopírovat obsah H5P do schránky',
   copiedToClipboard: 'Obsah je zkopírován do schránky',
@@ -195,8 +204,10 @@ H5PEditor.language.core = {
   pasteError: 'Nelze vložit ze schránky',
   pasteContentNotSupported: 'Obsah schránky H5P není v tomto kontextu podporován',
   pasteContentRestricted: 'Obsah schránky byl na tomto webu omezen',
-  pasteTooOld: 'Obsah ve schránce H5P je nižší verze (: clip), než co je v tomto kontextu podporováno (:local), pokud je to možné, zkuste aktualizovat obsah, který chcete vložit, zkopírujte znovu a zkuste jej vložit sem.',
-  pasteTooNew: 'Obsah ve schránce H5P má vyšší verzi (: clip), než je v tomto kontextu podporováno (:local), pokud je to možné, zkuste nejprve aktualizovat tento obsah a potom zkuste vložit obsah sem znovu.',
+  pasteTooOld:
+    'Obsah ve schránce H5P je nižší verze (: clip), než co je v tomto kontextu podporováno (:local), pokud je to možné, zkuste aktualizovat obsah, který chcete vložit, zkopírujte znovu a zkuste jej vložit sem.',
+  pasteTooNew:
+    'Obsah ve schránce H5P má vyšší verzi (: clip), než je v tomto kontextu podporováno (:local), pokud je to možné, zkuste nejprve aktualizovat tento obsah a potom zkuste vložit obsah sem znovu.',
   ok: 'OK',
   avTablistLabel: 'Vložit pomocí',
   tabTitleBasicFileUpload: 'Nahrát souboru',
@@ -209,12 +220,86 @@ H5PEditor.language.core = {
   language: 'Jazyk',
   noLanguagesSupported: 'Nejsou podporovány žádné jazyky',
   changeLanguage: 'Změnit jazyk na :language?',
-  thisWillPotentially: "Tím se potenciálně resetuje veškerý text a překlady. Toto nemůžete vrátit zpět. Samotný obsah se nezmění. Chcete pokračovat?",
+  thisWillPotentially:
+    'Tím se potenciálně resetuje veškerý text a překlady. Toto nemůžete vrátit zpět. Samotný obsah se nezmění. Chcete pokračovat?',
   notAllTextsChanged: 'Ne všechny texty byly změněny, pouze částečné pokrytí :language.',
-  contributeTranslations: 'Pokud chcete dokončit překlad pro :language, můžete se dozvědět o <a href=":url" target="_new"> přispívání překladů do H5P</a>',
-  unknownLibrary: 'Bohužel vybraný typ obsahu  \'%lib\' není v tomto systému nainstalován.',
+  contributeTranslations:
+    'Pokud chcete dokončit překlad pro :language, můžete se dozvědět o <a href=":url" target="_new"> přispívání překladů do H5P</a>',
+  unknownLibrary: "Bohužel vybraný typ obsahu  '%lib' není v tomto systému nainstalován.",
   proceedButtonLabel: 'Pokračovat k uložení',
   enterFullscreenButtonLabel: 'Přejít na celou obrazovku',
   exitFullscreenButtonLabel: 'Ukončit celou obrazovku',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
+  a11yTitleShowLabel: 'Show label for AT',
+  a11yTitleHideLabel: 'Hide label for AT',
 };
