@@ -27,12 +27,14 @@ H5PEditor.language.core = {
   addEntity: ':entity hinzufügen',
   tooLong: 'Der Text des Feldes ist zu lang. Es sollte :max Buchstaben oder weniger beinhalten.',
   invalidFormat: 'Das Feld beinhaltet ein ungültiges Format oder verbotene Zeichen.',
-  confirmChangeLibrary: 'Wenn dies ausgeführt wird, dann geht alles verloren, was mit dem aktuellen Inhaltstyp erstellt wurde. Ganz sicher den Inhaltstyp wechseln?',
+  confirmChangeLibrary:
+    'Wenn dies ausgeführt wird, dann geht alles verloren, was mit dem aktuellen Inhaltstyp erstellt wurde. Ganz sicher den Inhaltstyp wechseln?',
   commonFields: 'Bezeichnungen und Beschriftungen',
-  commonFieldsDescription: 'Hier können Bezeichnungen bearbeitet oder Texte übersetzt werden, die in diesem Inhalt verwendet werden.',
+  commonFieldsDescription:
+    'Hier können Bezeichnungen bearbeitet oder Texte übersetzt werden, die in diesem Inhalt verwendet werden.',
   uploading: 'Lade hoch, bitte warten...',
   noFollow: 'Dem Feld ":path" kann nicht gefolgt werden.',
-  editCopyright: 'Urheberrecht bearbeiten',
+  editCopyright: 'Rechteangaben bearbeiten',
   close: 'Schließen',
   tutorial: 'Tutorial',
   editMode: 'Bearbeitungsmodus',
@@ -49,12 +51,13 @@ H5PEditor.language.core = {
   selectFiletoUpload: 'Datei zum Hochladen auswählen',
   or: 'oder',
   enterAudioUrl: 'URL der Audiodatei eingeben',
-  enterVideoUrl: 'Enter video URL',
+  enterVideoUrl: 'Video-URL eingeben',
   enterAudioTitle: 'Link oder URL zur Audiodatei einfügen',
   enterVideoTitle: 'YouTube-Link oder andere Video-URL einfügen',
   uploadAudioTitle: 'Audio-Datei hochladen',
   uploadVideoTitle: 'Video-Datei hochladen',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P unterstützt alle externen Videoquellen, die das mp4-, webm- or ogv-Format haben, wie Vimeo Pro, und unterstützt YouTube- und Panopto-Links.',
   insert: 'Einfügen',
   cancel: 'Abbrechen',
   height: 'Höhe',
@@ -74,7 +77,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Hochladen',
   uploadPlaceholder: 'Keine Datei ausgewählt',
   uploadInstructionsTitle: 'Eine H5P-Datei hochladen.',
-  uploadInstructionsContent: 'Du kannst mit Beispielen von <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a> starten.',
+  uploadInstructionsContent:
+    'Du kannst mit Beispielen von <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a> starten.',
   uploadFileButtonLabel: 'Datei hochladen',
   uploadFileButtonChangeLabel: 'Datei ändern',
   uploadingThrobber: 'Lade hoch ...',
@@ -84,9 +88,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Die ausgewählte Datei konnte nicht hochgeladen werden',
   h5pFileWrongExtensionContent: 'Nur Dateien mit der Endung .h5p sind erlaubt.',
   h5pFileValidationFailedTitle: 'Die H5P-Datei ist ungültig oder fehlerhaft.',
-  h5pFileValidationFailedContent: 'Stelle sicher, dass die hochgeladene Datei gültigen H5P-Inhalt enthält. H5P-Dateien, die nur Bibliotheken enthalten, sollten über die H5P-Bibliothekenseite hochgeladen werden.',
+  h5pFileValidationFailedContent:
+    'Stelle sicher, dass die hochgeladene Datei gültigen H5P-Inhalt enthält. H5P-Dateien, die nur Bibliotheken enthalten, sollten über die H5P-Bibliothekenseite hochgeladen werden.',
   h5pFileUploadServerErrorTitle: 'Die H5P-Datei konnte nicht hochgeladen werden',
-  h5pFileUploadServerErrorContent: 'Ein unerwarteter Fehler ist aufgetreten. Bitte prüfe die Fehlerlogdatei des Servers für mehr Details.',
+  h5pFileUploadServerErrorContent:
+    'Ein unerwarteter Fehler ist aufgetreten. Bitte prüfe die Fehlerlogdatei des Servers für mehr Details.',
   contentTypeSectionAll: 'Alle Inhaltstypen',
   searchResults: 'Suchergebnisse',
   contentTypeSearchFieldPlaceholder: 'Nach Inhaltstypen suchen',
@@ -96,7 +102,7 @@ H5PEditor.language.core = {
   contentTypeDetailButtonLabel: 'Details',
   contentTypeUpdateButtonLabel: 'Aktualisieren',
   contentTypeUpdatingButtonLabel: 'Aktualisiere',
-  contentTypeGetButtonLabel: 'Übernehmen',
+  contentTypeGetButtonLabel: 'Installieren',
   contentTypeBackButtonLabel: 'Zurück',
   contentTypeIconAltText: 'Icon',
   contentTypeInstallSuccess: ':contentType erfolgreich installiert!',
@@ -116,7 +122,8 @@ H5PEditor.language.core = {
   readLess: 'weniger anzeigen',
   contentTypeOwner: 'Von :owner',
   contentTypeUnsupportedApiVersionTitle: 'Dieser Inhaltstyp erfordert eine neuere Version von H5P.',
-  contentTypeUnsupportedApiVersionContent: 'Kontaktiere bitte deinen Systemadministrator, um die notwendigen Aktualisierungen zu erhalten.',
+  contentTypeUnsupportedApiVersionContent:
+    'Kontaktiere bitte deinen Systemadministrator, um die notwendigen Aktualisierungen zu erhalten.',
   contentTypeUpdateAvailable: 'Aktualisierung verfügbar',
   contentTypeRestricted: 'Beschränkter Inhaltstyp',
   contentTypeRestrictedDesc: 'Die Verwendung dieses Inhaltstyps wurde durch einen Administrator beschränkt.',
@@ -127,10 +134,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Es ist keine Verbindung zum Hub möglich.',
   errorCommunicatingHubContent: 'Ein Fehler ist aufgetreten. Bitte versuche es noch einmal.',
   warningNoContentTypesInstalled: 'Es sind keine Inhaltstypen installiert.',
-  warningChangeBrowsingToSeeResults: 'Klicke <em>Alle Inhaltstypen</em> an, um eine Liste aller installierbaren Inhaltstypen zu erhalten.',
+  warningChangeBrowsingToSeeResults:
+    'Klicke <em>Alle Inhaltstypen</em> an, um eine Liste aller installierbaren Inhaltstypen zu erhalten.',
   warningUpdateAvailableTitle: 'Eine neuere Version von :contentType ist verfügbar.',
   warningUpdateAvailableBody: 'Aktualisiere auf die neueste Version, um von allen Verbesserungen zu profitieren.',
-  licenseDescription: 'Einige der Lizenzmerkmale sind unten aufgeführt. Klicke auf das Info-Icon, um den Originallizenztext zu lesen.',
+  licenseDescription:
+    'Einige der Eigenschaften der Lizenz sind unten aufgeführt. Klicke auf das Info-Icon, um den Originallizenztext zu lesen.',
   licenseModalTitle: 'Lizenzdetails',
   licenseModalSubtitle: 'Wähle eine Lizenz aus, um zu erfahren, welche Auflagen sie umfasst.',
   licenseUnspecified: 'Nicht näher angegeben',
@@ -150,12 +159,15 @@ H5PEditor.language.core = {
   screenshots: 'Bildschirmfotos',
   reloadButtonLabel: 'Neu laden',
   videoQuality: 'Beschreibung der Videoqualität',
-  videoQualityDescription: 'An dieser Beschreibung kann der Benutzer die Qualität des hochgeladenen Videos erkennen. Beispiele: 1080p, 720p, HD oder Mobil',
+  videoQualityDescription:
+    'An dieser Beschreibung kann der Benutzer die Qualität des hochgeladenen Videos erkennen. Beispiele: 1080p, 720p, HD oder mobil',
   videoQualityDefaultLabel: 'Videoqualität :index',
   noContentTypesAvailable: 'Keine Inhaltstypen verfügbar',
-  noContentTypesAvailableDesc: 'Es können keine Inhaltstypen aufgelistet werden, da Verbindungsprobleme zu H5P.org bestehen.',
+  noContentTypesAvailableDesc:
+    'Es können keine Inhaltstypen aufgelistet werden, da Verbindungsprobleme zu H5P.org bestehen.',
   contentTypeCacheOutdated: 'Die Liste der Inhaltstypen ist veraltet',
-  contentTypeCacheOutdatedDesc: 'Es kann nicht nach Aktualisierungen für Inhaltstypen gesucht werden, da Verbindungsprobleme zu H5P.org bestehen. Es ist möglich, dass keine neuen Inhaltstypen installiert oder bestehende aktualisiert werden können.',
+  contentTypeCacheOutdatedDesc:
+    'Es kann nicht nach Aktualisierungen für Inhaltstypen gesucht werden, da Verbindungsprobleme zu H5P.org bestehen. Es ist möglich, dass keine neuen Inhaltstypen installiert oder bestehende aktualisiert werden können.',
   tryAgain: 'Bitte versuche es erneut',
   getHelp: 'Hilfe',
   untitled: 'Unbenannt: :libraryTitle',
@@ -170,20 +182,24 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Bist du sicher, dass du diesen Autor aus der Liste entfernen möchtest?',
   addNewChange: 'Änderung hinzufügen',
   confirmDeleteChangeLog: 'Bist du dir sicher, dass du diesen Eintrag des Änderungsprotokolls löschen möchtest?',
-  changelogDescription: 'Einige Lizenzen schreiben vor, dass Änderungen am ursprünglichen Werk oder Derivaten gekennzeichnet werden. Hier kannst du solche Änderungen protokollieren - aus Lizenzgründen oder um einen Überblick über die vorgenommen Änderungen zu haben.',
+  changelogDescription:
+    'Einige Lizenzen schreiben vor, dass Änderungen am ursprünglichen Werk oder an Derivaten gekennzeichnet werden. Hier kannst du solche Änderungen protokollieren - aus Lizenzgründen oder um einen Überblick über die vorgenommen Änderungen zu haben.',
   logThisChange: 'Änderung festhalten',
   newChangeHasBeenLogged: 'Neue Änderung wurde festgehalten',
   loggedChanges: 'Änderungen festgehalten',
   noChangesHaveBeenLogged: 'Keine Änderungen festgehalten',
   errorHeader: 'Ein Fehler ist aufgetreten',
-  errorCalculatingMaxScore: 'Die maximale Punktzahl für diesen Inhalt konnte nicht berechnet werden. Es wird angenommen, dass die maximale Punktzahl 0 ist. Kontaktiere bitte deinen Administrator, falls das nicht korrekt ist.',
-  maxScoreSemanticsMissing: 'Die Datei, in der die Struktur des Inhalts beschrieben wird, konnte nicht gefunden werden (semantics.json).',
+  errorCalculatingMaxScore:
+    'Die maximale Punktzahl für diesen Inhalt konnte nicht berechnet werden. Es wird angenommen, dass die maximale Punktzahl 0 ist. Kontaktiere bitte deinen Administrator, falls das nicht korrekt ist.',
+  maxScoreSemanticsMissing:
+    'Die Datei, in der die Struktur des Inhalts beschrieben wird, konnte nicht gefunden werden (semantics.json).',
   copyButton: 'Kopieren',
   copiedButton: 'Kopiert',
   pasteButton: 'Einfügen',
   pasteAndReplaceButton: 'Einfügen & Ersetzen',
   pasteContent: 'Inhalte ersetzen',
-  confirmPasteContent: 'Diese Aktion ersetzt den derzeitigen Inhalt mit dem aus der Zwischenablage. Der aktuelle Inhalt geht dadurch verloren. Bist du dir sicher, dass du weitermachen willst?',
+  confirmPasteContent:
+    'Diese Aktion ersetzt den derzeitigen Inhalt mit dem aus der Zwischenablage. Der aktuelle Inhalt geht dadurch verloren. Bist du dir sicher, dass du weitermachen willst?',
   confirmPasteButtonText: 'Inhalt ersetzen',
   copyToClipboard: 'Kopiere H5P-Inhalt in den Zwischenspeicher',
   copiedToClipboard: 'Inhalt wurde in die Zwischenablage kopiert',
@@ -192,9 +208,12 @@ H5PEditor.language.core = {
   pasteNoContent: 'Kein H5P-Inhalt in der Zwischenablage',
   pasteError: 'Kann nicht aus der Zwischenablage einfügen',
   pasteContentNotSupported: 'Der Inhalt der Zwischenablage kann an dieser Stelle nicht eingefügt werden',
-  pasteContentRestricted: 'Der Inhalt der Zwischenlage kann auf dieser Seite nicht eingefügt werden, da die Nutzung des Inhaltstyps von den Administratoren eingeschränkt wurde',
-  pasteTooOld: 'Der Inhalt in der Zwischenablage liegt in einer zu alten Version vor (:clip), die nicht an dieser Stelle eingefügt werden kann (erfordert :local). Aktualisiere bitte den Inhalt, den du einfügen möchtest, kopiere ihn erneut und füge ihn dann hier wieder ein.',
-  pasteTooNew: 'Der Inhalt in der Zwischenablage liegt in einer zu neuen Version vor (:clip), die nicht an dieser Stelle eingefügt werden kann (erfordert: local). Aktualisiere bitte den Inhalt, in den du hineinkopieren möchtest und versuche es dann erneut. ',
+  pasteContentRestricted:
+    'Der Inhalt der Zwischenlage kann auf dieser Seite nicht eingefügt werden, da die Nutzung des Inhaltstyps von den Administratoren eingeschränkt wurde',
+  pasteTooOld:
+    'Der Inhalt in der Zwischenablage liegt in einer zu alten Version vor (:clip), die nicht an dieser Stelle eingefügt werden kann (erfordert :local). Aktualisiere bitte den Inhalt, den du einfügen möchtest, kopiere ihn erneut und füge ihn dann hier wieder ein.',
+  pasteTooNew:
+    'Der Inhalt in der Zwischenablage liegt in einer zu neuen Version vor (:clip), die nicht an dieser Stelle eingefügt werden kann (erfordert: local). Aktualisiere bitte den Inhalt, in den du hineinkopieren möchtest und versuche es dann erneut. ',
   ok: 'OK',
   avTablistLabel: 'Einfügen über',
   tabTitleBasicFileUpload: 'Datei hochladen',
@@ -207,14 +226,86 @@ H5PEditor.language.core = {
   language: 'Sprache',
   noLanguagesSupported: 'Keine Sprachen unterstützt',
   changeLanguage: 'Sprache auf :language ändern?',
-  thisWillPotentially: "Hierdurch werden möglicherweise alle Bezeichnungen und Beschriftungen zurückgesetzt. Dies kann nicht rückgängig gemacht werden. Der Inhalt selbst wird nicht geändert. Möchtest du fortfahren?",
+  thisWillPotentially:
+    'Hierdurch werden möglicherweise alle Bezeichnungen und Beschriftungen zurückgesetzt. Dies kann nicht rückgängig gemacht werden. Der Inhalt selbst wird nicht geändert. Möchtest du fortfahren?',
   notAllTextsChanged: 'Nicht alle Texte wurden geändert, da die Übersetzungen in :language nicht vollständig ist.',
-  contributeTranslations: 'Wenn du die Übersetzung in :language vervollständigen möchtest, kannst du <a href=":url" target="_new">Übersetzungen zu H5P beisteuern</a>',
-  unknownLibrary: 'Leider ist der ausgewählte Inhaltstyp \'%lib\' auf dieser Seite nicht installiert.',
+  contributeTranslations:
+    'Wenn du die Übersetzung in :language vervollständigen möchtest, kannst du <a href=":url" target="_new">Übersetzungen zu H5P beisteuern</a>',
+  unknownLibrary: "Leider ist der ausgewählte Inhaltstyp '%lib' auf dieser Seite nicht installiert.",
   proceedButtonLabel: 'Weiter zum Speichern',
   enterFullscreenButtonLabel: 'Vollbild',
   exitFullscreenButtonLabel: 'Vollbild beenden',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
-  reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  reuseSuccess: ':title wurde Erfolgreich vom H5P Hub importiert.',
+  noContentHeader: 'Kein passender Inhalt?',
+  noContentSuggestion: 'Erstelle ihn selber!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'Gesamter geteilter Inhalt',
+  popularContent: 'Beliebter Inhalt',
+  allPopular: 'Alles beliebte',
+  newOnTheHub: 'Neu im Hub',
+  allNew: 'Alle neuen Inhalte',
+  filterBy: 'Filtern nach',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Niveau',
+      dialogHeader: 'Wähle den Schwierigkeitsgrad',
+      dialogButtonLabel: 'Nach Schwierigkeitsgrad filtern',
+    },
+    language: {
+      dropdownLabel: 'Sprache',
+      dialogHeader: 'Wähle die Sprache(n)',
+      dialogButtonLabel: 'Nach Sprachen filtern',
+      searchPlaceholder: 'Tippe, um nach Sprachen zu suchen',
+    },
+    reviewed: {
+      dropdownLabel: 'Überprüft',
+      dialogHeader: 'Überprüfter Inhalt',
+      dialogButtonLabel: 'Filtern',
+      optionLabel: 'Nur überprüften Inhalt anzeigen',
+    },
+    contentTypes: {
+      dropdownLabel: 'Inhaltstypen',
+      dialogHeader: 'Wähle Inhaltstyp(en) aus',
+      dialogButtonLabel: 'Nach Inhaltstypen filtern',
+      searchPlaceholder: 'Tippe, um nach Inhaltstypen zu suchen',
+    },
+    disciplines: {
+      dropdownLabel: 'Fach',
+      dialogHeader: 'Wähle das Fach / die Fächer aus',
+      dialogButtonLabel: 'Nach Fächern filtern',
+      searchPlaceholder: 'Tippe, um nach Fächern zu suchen',
+    },
+    licenses: {
+      dropdownLabel: 'Lizenz',
+      dialogHeader: 'Wähle die von dir bevorzugten Nutzungsrechte aus',
+      dialogButtonLabel: 'Nach Lizenzen filtern',
+      options: { modified: 'Darf verändert werden', commercial: 'Erlaubt kommerzielle Nutzung' },
+    },
+  },
+  clearFilters: 'Alle Filter zurücksetzen',
+  contentSearchFieldPlaceholder: 'Nach Inhalt suchen',
+  loadingContentTitle: 'Wir laden den Inhalt für dich...',
+  loadingContentSubtitle: 'Bitte warten',
+  by: 'Von',
+  dropdownButton: 'Dropdown öffnen',
+  paginationNavigation: 'Seitennavigation',
+  page: 'Seite',
+  currentPage: 'Aktuelle Seite',
+  nextPage: 'Zur nächsten Seite gehen',
+  previousPage: 'Zur vorherigen Seite gehen',
+  contentPreviewButtonLabel: 'Vorschau',
+  contentDownloadButtonLabel: 'Inhalt verwenden',
+  reuseContentTabLabel: 'Geteilten Inhalt verwenden',
+  contentPublisherPanelHeader: 'Informationen zum Herausgeber',
+  noContentFoundDesc: 'Kein Inhalt passt zu deinen Suchkriterien.',
+  h5pType: 'Inhaltstyp',
+  level: 'Niveau',
+  size: 'Größe',
+  failedFetchingData: 'Konnte Daten nicht herunterladen',
+  filterErrorMessage: 'Etwas ist schiefgegangen. Biete lade die Seite neu.',
+  in: 'in',
+  navigateToParent: 'Zum übergeordneten Objekt navigieren',
 };
