@@ -27,9 +27,11 @@ H5PEditor.language.core = {
   addEntity: 'Προσθήκη :entity',
   tooLong: 'Η τιμή του πεδίου είναι πολύ μεγάλη, πρέπει να περιέχει :max γράμματα ή λιγότερα.',
   invalidFormat: 'Η τιμή του πεδίου περιέχει μια μη έγκυρη μορφή ή χαρακτήρες οι οποίοι απαγορεύονται.',
-  confirmChangeLibrary: 'Κάνοντας αυτό θα χάσετε όλη την εργασία που έχει γίνει με τον τρέχοντα τύπο περιεχομένου. Είστε βέβαιοι ότι θέλετε να αλλάξετε τύπο περιεχομένου;',
+  confirmChangeLibrary:
+    'Κάνοντας αυτό θα χάσετε όλη την εργασία που έχει γίνει με τον τρέχοντα τύπο περιεχομένου. Είστε βέβαιοι ότι θέλετε να αλλάξετε τύπο περιεχομένου;',
   commonFields: 'Παραλλαγές κειμένου και μεταφράσεις',
-  commonFieldsDescription: 'Εδώ μπορείτε να επεξεργαστείτε ρυθμίσεις ή να μεταφράσετε κείμενα τα οποία χρησιμοποιούνται σε αυτό το περιεχόμενο.',
+  commonFieldsDescription:
+    'Εδώ μπορείτε να επεξεργαστείτε ρυθμίσεις ή να μεταφράσετε κείμενα τα οποία χρησιμοποιούνται σε αυτό το περιεχόμενο.',
   uploading: 'Μεταφόρτωση, παρακαλούμε περιμένετε...',
   noFollow: 'Δεν είναι δυνατή η παρακολούθηση του πεδίου ":path".',
   editCopyright: 'Επεξεργασία πνευματικών δικαιωμάτων',
@@ -54,7 +56,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Επικολλήστε σύνδεσμο YouTube ή άλλη διεύθυνση URL βίντεο',
   uploadAudioTitle: 'Μεταφορτώστε αρχείο ήχου',
   uploadVideoTitle: 'Μεταφορτώστε αρχείο video',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Εισαγωγή',
   cancel: 'Ακύρωση',
   height: 'ύψος',
@@ -74,7 +77,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Μεταφόρτωση',
   uploadPlaceholder: 'Δεν έχει επιλεγεί αρχείο',
   uploadInstructionsTitle: 'Μεταφορτώστε ένα αρχείο H5P.',
-  uploadInstructionsContent: 'Μπορείτε να ξεκινήσετε με παραδείγματα από το <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Μπορείτε να ξεκινήσετε με παραδείγματα από το <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Μεταφόρτωση αρχείου',
   uploadFileButtonChangeLabel: 'Αλλαγή αρχείου',
   uploadingThrobber: 'Γίνεται μεταφόρτωση...',
@@ -84,9 +88,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Δεν ήταν δυνατή η μεταφόρτωση του επιλεγμένου αρχείου',
   h5pFileWrongExtensionContent: 'Μόνο αρχεία με επέκταση .h5p επιτρέπονται.',
   h5pFileValidationFailedTitle: 'Δεν ήταν δυνατή η επικύρωση του αρχείου H5P.',
-  h5pFileValidationFailedContent: 'Βεβαιωθείτε ότι το φορτωμένο H5P περιέχει έγκυρο περιεχόμενο H5P. Τα αρχεία H5P τα οποία περιέχουν μόνο βιβλιοθήκες, θα πρέπει να μεταφορτωθούν μέσω της σελίδας Βιβλιοθήκες H5P.',
+  h5pFileValidationFailedContent:
+    'Βεβαιωθείτε ότι το φορτωμένο H5P περιέχει έγκυρο περιεχόμενο H5P. Τα αρχεία H5P τα οποία περιέχουν μόνο βιβλιοθήκες, θα πρέπει να μεταφορτωθούν μέσω της σελίδας Βιβλιοθήκες H5P.',
   h5pFileUploadServerErrorTitle: 'Δεν ήταν δυνατή η μεταφόρτωση του αρχείου H5P',
-  h5pFileUploadServerErrorContent: 'Προέκυψε ένα μη αναμενόμενο σφάλμα. Ελέγξτε το αρχείο σφαλμάτων του διακομιστή σας για περισσότερες λεπτομέρειες.',
+  h5pFileUploadServerErrorContent:
+    'Προέκυψε ένα μη αναμενόμενο σφάλμα. Ελέγξτε το αρχείο σφαλμάτων του διακομιστή σας για περισσότερες λεπτομέρειες.',
   contentTypeSectionAll: 'All Content Types',
   searchResults: 'Search Results',
   contentTypeSearchFieldPlaceholder: 'Αναζητήστε Τύπους Περιεχομένου',
@@ -116,7 +122,8 @@ H5PEditor.language.core = {
   readLess: 'Διαβάστε λιγότερα',
   contentTypeOwner: 'Από :owner',
   contentTypeUnsupportedApiVersionTitle: 'Αυτός ο τύπος περιεχομένου απαιτεί νεότερη έκδοση',
-  contentTypeUnsupportedApiVersionContent: 'Επικοινωνήστε με το διαχειριστή του συστήματός σας, για την παροχή των απαραίτητων ενημερώσεων',
+  contentTypeUnsupportedApiVersionContent:
+    'Επικοινωνήστε με το διαχειριστή του συστήματός σας, για την παροχή των απαραίτητων ενημερώσεων',
   contentTypeUpdateAvailable: 'Υπάρχει διαθέσιμη ενημέρωση',
   contentTypeRestricted: 'Περιορισμένος τύπος περιεχομένου',
   contentTypeRestrictedDesc: 'Η χρήση αυτού του τύπου περιεχομένου έχει περιοριστεί από τον διαχειριστή.',
@@ -127,10 +134,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Δεν είναι σε θέση να επικοινωνήσει με το διανομέα.',
   errorCommunicatingHubContent: 'Παρουσιάστηκε σφάλμα. Παρακαλούμε προσπαθήστε ξανα.',
   warningNoContentTypesInstalled: 'Δεν έχετε εγκατεστημένους τύπους περιεχομένου.',
-  warningChangeBrowsingToSeeResults: 'Κάντε κλικ στο κουμπί <em>Όλα</em> για να λάβετε τη λίστα με όλους τους τύπους περιεχομένου τους οποίους μπορείτε να εγκαταστήσετε.',
+  warningChangeBrowsingToSeeResults:
+    'Κάντε κλικ στο κουμπί <em>Όλα</em> για να λάβετε τη λίστα με όλους τους τύπους περιεχομένου τους οποίους μπορείτε να εγκαταστήσετε.',
   warningUpdateAvailableTitle: 'Μια νέα έκδοση του :contentType είναι διαθέσιμη.',
   warningUpdateAvailableBody: 'Ενημερώστε στην τελευταία έκδοση για βελτιωμένη εμπειρία χρήσης.',
-  licenseDescription: 'Ορισμένα από τα χαρακτηριστικά αυτής της άδειας αναφέρονται παρακάτω. Κάντε κλικ στο παραπάνω εικονίδιο πληροφοριών για να διαβάσετε το πρωτότυπο κείμενο της άδειας χρήσης.',
+  licenseDescription:
+    'Ορισμένα από τα χαρακτηριστικά αυτής της άδειας αναφέρονται παρακάτω. Κάντε κλικ στο παραπάνω εικονίδιο πληροφοριών για να διαβάσετε το πρωτότυπο κείμενο της άδειας χρήσης.',
   licenseModalTitle: 'Λεπτομέρειες Άδειας Χρήσης',
   licenseModalSubtitle: 'Επιλέξτε μια άδεια χρήσης για να δείτε πληροφορίες σχετικά με τη σωστή χρήση',
   licenseUnspecified: 'Απροσδιόριστο',
@@ -150,12 +159,15 @@ H5PEditor.language.core = {
   screenshots: 'Στιγμιότυπα',
   reloadButtonLabel: 'Επαναφόρτωση',
   videoQuality: 'Ετικέτα ποιότητας βίντεο',
-  videoQualityDescription: 'Αυτή η ετικέτα βοηθά τον χρήστη να αναγνωρίσει την τρέχουσα ποιότητα του βίντεο. Για παράδειγμα 1080p, 720p, HD ή Mobile',
+  videoQualityDescription:
+    'Αυτή η ετικέτα βοηθά τον χρήστη να αναγνωρίσει την τρέχουσα ποιότητα του βίντεο. Για παράδειγμα 1080p, 720p, HD ή Mobile',
   videoQualityDefaultLabel: 'Ποιότητα :index',
   noContentTypesAvailable: 'No content types are available',
-  noContentTypesAvailableDesc: 'Your site is having difficulties connecting to H5P.org and listing the available content types.',
+  noContentTypesAvailableDesc:
+    'Your site is having difficulties connecting to H5P.org and listing the available content types.',
   contentTypeCacheOutdated: 'Content type list outdated',
-  contentTypeCacheOutdatedDesc: 'Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.',
+  contentTypeCacheOutdatedDesc:
+    'Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.',
   tryAgain: 'Try again',
   getHelp: 'Get help',
   untitled: 'Untitled :libraryTitle',
@@ -170,20 +182,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
   addNewChange: 'Add new change',
   confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
   logThisChange: 'Log this change',
   newChangeHasBeenLogged: 'New change has been logged',
   loggedChanges: 'Logged changes',
   noChangesHaveBeenLogged: 'No changes have been logged',
   errorHeader: 'An error occured',
-  errorCalculatingMaxScore: 'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
+  errorCalculatingMaxScore:
+    'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
   maxScoreSemanticsMissing: 'Could not find the expected semantics in the content.',
   copyButton: 'Copy',
   copiedButton: 'Copied',
   pasteButton: 'Paste',
   pasteAndReplaceButton: 'Paste & Replace',
   pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
+  confirmPasteContent:
+    'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
   confirmPasteButtonText: 'Replace content',
   copyToClipboard: 'Copy H5P content to the clipboard',
   copiedToClipboard: 'Content is copied to the clipboard',
@@ -193,8 +208,10 @@ H5PEditor.language.core = {
   pasteError: 'Cannot paste from clipboard',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insert using',
   tabTitleBasicFileUpload: 'File Upload',
@@ -207,14 +224,86 @@ H5PEditor.language.core = {
   language: 'Language',
   noLanguagesSupported: 'No languages supported',
   changeLanguage: 'Change language to :language?',
-  thisWillPotentially: "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
+  thisWillPotentially:
+    "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
   notAllTextsChanged: 'Not all texts were changed, there is only partial coverage for :language.',
-  contributeTranslations: 'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Unfortunately, the selected content type \'%lib\' isn\'t installed on this system.',
+  contributeTranslations:
+    'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Unfortunately, the selected content type '%lib' isn't installed on this system.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
