@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Añadir :entity',
   tooLong: 'El valor del campo es demasiado largo, debería contener un máximo de :max lettras.',
   invalidFormat: 'El valor del campo contiene un formato no válido o caracteres prohibidos.',
-  confirmChangeLibrary: 'Al hacer esto perderá todo el trabajo realizado con el tipo de contenido actual. ¿Está seguro de que quiere cambiar el tipo de contenido?',
+  confirmChangeLibrary:
+    'Al hacer esto perderá todo el trabajo realizado con el tipo de contenido actual. ¿Está seguro de que quiere cambiar el tipo de contenido?',
   commonFields: 'Parámetros y textos',
   commonFieldsDescription: 'Aquí puede editar parámetros o traducir textos utilizados en este contenido.',
   uploading: 'Subiendo, por favor espera...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Peque el link de Youtube o la URL de la fuente del video',
   uploadAudioTitle: 'Subir fichero de audio',
   uploadVideoTitle: 'Subir fichero de video',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Insertar',
   cancel: 'Cancelar',
   height: 'altura',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Subir',
   uploadPlaceholder: 'Ningún fichero seleccionado',
   uploadInstructionsTitle: 'Subir un fichero H5P.',
-  uploadInstructionsContent: 'Usted puede comenzar con ejemplos de <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Usted puede comenzar con ejemplos de <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Subir un fichero',
   uploadFileButtonChangeLabel: 'Cambiar fichero',
   uploadingThrobber: 'Subiendo ahora...',
@@ -84,9 +87,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'El fichero elegido no se puede subir',
   h5pFileWrongExtensionContent: 'Solo se permiten ficheros con la extensión .h5p.',
   h5pFileValidationFailedTitle: 'No se puede validar el fichero H5P.',
-  h5pFileValidationFailedContent: 'Asegúrese de que el H5P cargado contiene contenido H5P válido. H5P Los archivos que contienen sólo bibliotecas deben cargarse a través de la página Bibliotecas H5P.',
+  h5pFileValidationFailedContent:
+    'Asegúrese de que el H5P cargado contiene contenido H5P válido. H5P Los archivos que contienen sólo bibliotecas deben cargarse a través de la página Bibliotecas H5P.',
   h5pFileUploadServerErrorTitle: 'El fichero H5P no se ha podido cargar',
-  h5pFileUploadServerErrorContent: 'Ocurrió un error inesperado. Compruebe el registro de errores del servidor para obtener más detalles.',
+  h5pFileUploadServerErrorContent:
+    'Ocurrió un error inesperado. Compruebe el registro de errores del servidor para obtener más detalles.',
   contentTypeSectionAll: 'Todos los tipos de contenidos',
   searchResults: 'Resultados de búsqueda',
   contentTypeSearchFieldPlaceholder: 'Buscar tipos de contenido',
@@ -116,7 +121,8 @@ H5PEditor.language.core = {
   readLess: 'Leer menos',
   contentTypeOwner: 'Por :owner',
   contentTypeUnsupportedApiVersionTitle: 'Este tipo de contenido requiere una nueva versión del core',
-  contentTypeUnsupportedApiVersionContent: 'Póngase en contacto con el responsable del sistema para obtener las actualizaciones necesarias.',
+  contentTypeUnsupportedApiVersionContent:
+    'Póngase en contacto con el responsable del sistema para obtener las actualizaciones necesarias.',
   contentTypeUpdateAvailable: 'Actualización disponible',
   contentTypeRestricted: 'Tipo de contenido restringido',
   contentTypeRestrictedDesc: 'El uso de este tipo de contenido ha sido restringido por un administrador.',
@@ -127,10 +133,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'No se puede comunicar con el hub.',
   errorCommunicatingHubContent: 'Ha ocurrido un error. Por favor, inténtelo de nuevo.',
   warningNoContentTypesInstalled: 'No tiene ningún tipo de contenido instalado.',
-  warningChangeBrowsingToSeeResults: 'Click <em>Todos</em> para obtener la lista de todos los tipos de contenido que puede instalar .',
+  warningChangeBrowsingToSeeResults:
+    'Click <em>Todos</em> para obtener la lista de todos los tipos de contenido que puede instalar .',
   warningUpdateAvailableTitle: 'Una nueva versión de :contentType está disponible.',
   warningUpdateAvailableBody: 'Actualización a la última versión para una experiencia mejorada.',
-  licenseDescription: 'Algunas de las características de esta licencia se indican a continuación. Haga clic en el icono de información de arriba para leer el texto original de la licencia.',
+  licenseDescription:
+    'Algunas de las características de esta licencia se indican a continuación. Haga clic en el icono de información de arriba para leer el texto original de la licencia.',
   licenseModalTitle: 'Detalles de la Licencia',
   licenseModalSubtitle: 'Seleccione una licencia para ver información sobre el uso adecuado',
   licenseUnspecified: 'Sin especificar',
@@ -150,19 +158,23 @@ H5PEditor.language.core = {
   screenshots: 'Capturas de pantalla',
   reloadButtonLabel: 'Recargar',
   videoQuality: 'Etiqueta de calidad de vídeo',
-  videoQualityDescription: 'Esta etiqueta ayuda al usuario a identificar la calidad actual del vídeo. Por ejemplo, 1080p, 720p, HD o móvil',
+  videoQualityDescription:
+    'Esta etiqueta ayuda al usuario a identificar la calidad actual del vídeo. Por ejemplo, 1080p, 720p, HD o móvil',
   videoQualityDefaultLabel: 'Calidad :index',
   noContentTypesAvailable: 'No hay tipos de contenido disponibles',
-  noContentTypesAvailableDesc: 'Su sitio tiene dificultades para conectarse a H5P.org y para listar los tipos de contenido disponibles.',
+  noContentTypesAvailableDesc:
+    'Su sitio tiene dificultades para conectarse a H5P.org y para listar los tipos de contenido disponibles.',
   contentTypeCacheOutdated: 'Lista de tipos de contenido obsoleta',
-  contentTypeCacheOutdatedDesc: 'Su sitio tiene dificultades para conectarse a H5P.org para comprobar si hay actualizaciones del tipo de contenido. Es posible que no pueda actualizar o instalar nuevos tipos de contenido.',
+  contentTypeCacheOutdatedDesc:
+    'Su sitio tiene dificultades para conectarse a H5P.org para comprobar si hay actualizaciones del tipo de contenido. Es posible que no pueda actualizar o instalar nuevos tipos de contenido.',
   tryAgain: 'Inténtelo de nuevo',
   getHelp: 'Obtener ayuda',
   untitled: 'Sin título :libraryTitle',
   title: 'Título',
   metadata: 'Metadata',
   addTitle: 'Añadir título',
-  usedForSearchingReportsAndCopyrightInformation: 'Utilizado para búsquedas, informes e información de derechos de autor',
+  usedForSearchingReportsAndCopyrightInformation:
+    'Utilizado para búsquedas, informes e información de derechos de autor',
   metadataSharingAndLicensingInfo: 'Metadata (información sobre uso compartido y licencias)',
   fillInTheFieldsBelow: 'Rellene los siguientes campos',
   saveMetadata: 'Guardar metadata',
@@ -170,20 +182,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: '¿Está seguro de que desea eliminar a este autor?',
   addNewChange: 'Añadir nuevo cambio',
   confirmDeleteChangeLog: '¿Está seguro de que desea borrar esta entrada del registro de cambios?',
-  changelogDescription: 'Algunas licencias requieren que los cambios realizados en la obra original o los derivados se registren y visualicen. Usted puede registrar sus cambios aquí por razones de licencia o simplemente para permitirse a usted y a otros mantener un registro de los cambios realizados en este contenido.',
+  changelogDescription:
+    'Algunas licencias requieren que los cambios realizados en la obra original o los derivados se registren y visualicen. Usted puede registrar sus cambios aquí por razones de licencia o simplemente para permitirse a usted y a otros mantener un registro de los cambios realizados en este contenido.',
   logThisChange: 'Registrar este cambio',
   newChangeHasBeenLogged: 'Se ha registrado un nuevo cambio',
   loggedChanges: 'Cambios registrados',
   noChangesHaveBeenLogged: 'No se ha registrado ningún cambio',
   errorHeader: 'Se ha producido un error',
-  errorCalculatingMaxScore: 'No se pudo calcular la puntuación máxima para este contenido. La puntuación máxima se supone que es 0. Contacte con su administrador si esto no es correcto.',
+  errorCalculatingMaxScore:
+    'No se pudo calcular la puntuación máxima para este contenido. La puntuación máxima se supone que es 0. Contacte con su administrador si esto no es correcto.',
   maxScoreSemanticsMissing: 'No se pudo encontrar la semántica esperada en el contenido.',
   copyButton: 'Copiar',
   copiedButton: 'Copiado',
   pasteButton: 'Pegar',
   pasteAndReplaceButton: 'Pegar y Reemplazar',
   pasteContent: 'Reemplazar contenido',
-  confirmPasteContent: 'Al hacer esto usted reemplazará el contenido actual con el contenido de su portapapeles. El contenido actual se perderá. ¿Está seguro de que quiere continuar?',
+  confirmPasteContent:
+    'Al hacer esto usted reemplazará el contenido actual con el contenido de su portapapeles. El contenido actual se perderá. ¿Está seguro de que quiere continuar?',
   confirmPasteButtonText: 'Reemplazar contenido',
   copyToClipboard: 'Copiar el contenido H5P al portapapeles',
   copiedToClipboard: 'El contenido se ha copiado al portapapeles',
@@ -193,13 +208,16 @@ H5PEditor.language.core = {
   pasteError: 'No se puede pegar desde el portapapeles',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insertar usando',
   tabTitleBasicFileUpload: 'Subir fichero',
   tabTitleInputLinkURL: 'Link/URL',
-  errorTooHighVersion: 'Los parámetros contienen %used mientras que solamente %supported o anteriores están soportados.',
+  errorTooHighVersion:
+    'Los parámetros contienen %used mientras que solamente %supported o anteriores están soportados.',
   errorNotSupported: 'Los parámetros contienen %used que no está soportado.',
   errorParamsBroken: 'Los parámetros están rotos.',
   libraryMissing: 'Falta la biblioteca requerida %lib.',
@@ -207,14 +225,86 @@ H5PEditor.language.core = {
   language: 'Idioma',
   noLanguagesSupported: 'No hay idiomas soportados',
   changeLanguage: '¿Cambiar el idioma a :language?',
-  thisWillPotentially: 'Esto potencialmente restablecerá todo el texto y las traducciones. No puedes deshacer esto. El contenido en sí no será cambiado. ¿Desea proceder?',
+  thisWillPotentially:
+    'Esto potencialmente restablecerá todo el texto y las traducciones. No puedes deshacer esto. El contenido en sí no será cambiado. ¿Desea proceder?',
   notAllTextsChanged: 'No todos los textos fueron cambiados, solo hay una cobertura parcial para :language.',
-  contributeTranslations: 'Si desea completar la traducción para :language, puede obtener información sobre <a href=":url" target="_new"> contribuir con traducciones al H5P </a>',
-  unknownLibrary: 'Desafortunadamente, el tipo de contenido seleccionado \'%lib\' no está instalado en este sistema.',
+  contributeTranslations:
+    'Si desea completar la traducción para :language, puede obtener información sobre <a href=":url" target="_new"> contribuir con traducciones al H5P </a>',
+  unknownLibrary: "Desafortunadamente, el tipo de contenido seleccionado '%lib' no está instalado en este sistema.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
