@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Lisa :entity',
   tooLong: 'Välja väärtus on liiga pikk, lubatud on :max tähte või vähem.',
   invalidFormat: 'Välja väärtus sisaldab sobimatut formaati või keelatud märke.',
-  confirmChangeLibrary: 'Seda tehes kaotad kogu töö, mis on tehtud jooksva sisu-tüübiga. Kas oled kindel, et soovid muuta sisu tüüpi?',
+  confirmChangeLibrary:
+    'Seda tehes kaotad kogu töö, mis on tehtud jooksva sisu-tüübiga. Kas oled kindel, et soovid muuta sisu tüüpi?',
   commonFields: 'Teksti asendused ja tõlge',
   commonFieldsDescription: 'Siin saad muuta seadeid või tõlkida selle sisutüübi juures kasutatud tekste.',
   uploading: 'Laadib üles, palun oota...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Kleebi lähtevideo link',
   uploadAudioTitle: 'Laadi helifail üles',
   uploadVideoTitle: 'Laadi videofail üles',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Aseta',
   cancel: 'Tühista',
   height: 'kõrgus',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Laadi üles',
   uploadPlaceholder: 'Faili ei ole valitud',
   uploadInstructionsTitle: 'Laadi H5P fail üles.',
-  uploadInstructionsContent: 'Võid alustada näidetega lehelt <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Võid alustada näidetega lehelt <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Laadi fail üles',
   uploadFileButtonChangeLabel: 'Muuda faili',
   uploadingThrobber: 'Laadib üles...',
@@ -84,11 +87,10 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Valitud faili ei saanud üles laadida',
   h5pFileWrongExtensionContent: 'Ainult .h5p laiendiga failid on lubatud.',
   h5pFileValidationFailedTitle: 'Ei saanud H5P faili valideerida.',
-  h5pFileValidationFailedContent: 'Veendu, et üles laetud H5P sisu on valiidne. H5P' +
-  ' failid, mis sisaldab vaid teeke, tuleks laadida läbi H5P teekide lehe.',
+  h5pFileValidationFailedContent:
+    'Veendu, et üles laetud H5P sisu on valiidne. H5P failid, mis sisaldab vaid teeke, tuleks laadida läbi H5P teekide lehe.',
   h5pFileUploadServerErrorTitle: 'H5P faili ei saanud üles laadida',
-  h5pFileUploadServerErrorContent: 'Ootamatu viga. Vaata täpsemalt oma vea-logist' +
-  ' täiendavaid detaile.',
+  h5pFileUploadServerErrorContent: 'Ootamatu viga. Vaata täpsemalt oma vea-logist täiendavaid detaile.',
   contentTypeSectionAll: 'Kõik sisutüübid',
   searchResults: 'Otsi tulemusi',
   contentTypeSearchFieldPlaceholder: 'Otsi sisutüüpe',
@@ -128,11 +130,13 @@ H5PEditor.language.core = {
   currentMenuSelected: 'jooksvalt valitud',
   errorCommunicatingHubTitle: 'Ei saa keskusega ühendust.',
   errorCommunicatingHubContent: 'Juhtus viga. Palun proovi uuesti.',
-  warningNoContentTypesInstalled: "Sul ei ole ühtki sisutüüpi paigaldatud.",
-  warningChangeBrowsingToSeeResults: 'Kliki <em>Kõik</em> et näha loetelu kõikidest sisutüüpidest, mida saad paigaldada.',
+  warningNoContentTypesInstalled: 'Sul ei ole ühtki sisutüüpi paigaldatud.',
+  warningChangeBrowsingToSeeResults:
+    'Kliki <em>Kõik</em> et näha loetelu kõikidest sisutüüpidest, mida saad paigaldada.',
   warningUpdateAvailableTitle: 'Sisutüübist :contentType on uus versioon saadaval.',
   warningUpdateAvailableBody: 'Parema kasutajakogemuse jaoks uuenda viimasesse versiooni.',
-  licenseDescription: 'Mõned selle litsentsi iseärasustest on näidatud allpool. Kliki ülal olevat info-ikooni, et lugeda litsentsilepingu teksti.',
+  licenseDescription:
+    'Mõned selle litsentsi iseärasustest on näidatud allpool. Kliki ülal olevat info-ikooni, et lugeda litsentsilepingu teksti.',
   licenseModalTitle: 'Litsentsi detailid',
   licenseModalSubtitle: 'Vali litsents, et näha teavet sobilikust kasutamisest',
   licenseUnspecified: 'Määratlemata',
@@ -152,12 +156,15 @@ H5PEditor.language.core = {
   screenshots: 'Ekraanipildid',
   reloadButtonLabel: 'Laadi uuesti',
   videoQuality: 'Videokvaliteedi silt',
-  videoQualityDescription: 'See silt aitab kasutajal määratleda videomängimise kvaliteeti. Näiteks 1080p, 720p, HD või mobiilne',
+  videoQualityDescription:
+    'See silt aitab kasutajal määratleda videomängimise kvaliteeti. Näiteks 1080p, 720p, HD või mobiilne',
   videoQualityDefaultLabel: 'Kvaliteet :index',
   noContentTypesAvailable: 'Saadaval ei ole sisutüüpe',
-  noContentTypesAvailableDesc: 'Sinu lehekülg ei saa ühenduda H5P.org serveriga ja näidata võimalikke sisutüüpide loetlemisega.',
+  noContentTypesAvailableDesc:
+    'Sinu lehekülg ei saa ühenduda H5P.org serveriga ja näidata võimalikke sisutüüpide loetlemisega.',
   contentTypeCacheOutdated: 'Sisutüüpide loetelu on aegunud',
-  contentTypeCacheOutdatedDesc: 'Sinu lehekülg ei saa ühenduda H5P.org serveriga, kontrollimaks sisutüüpide uuendusi. Sa ei saa sisutüüpe uuendada ega paigaldada.',
+  contentTypeCacheOutdatedDesc:
+    'Sinu lehekülg ei saa ühenduda H5P.org serveriga, kontrollimaks sisutüüpide uuendusi. Sa ei saa sisutüüpe uuendada ega paigaldada.',
   tryAgain: 'Proovi uuesti',
   getHelp: 'Vajan abi',
   untitled: 'Ilma pealkirjata :libraryTitle',
@@ -166,26 +173,29 @@ H5PEditor.language.core = {
   addTitle: 'Lisa pealkiri',
   usedForSearchingReportsAndCopyrightInformation: 'Kasutatakse otsimise, raportite ja autoriõiguse teabe jaoks',
   metadataSharingAndLicensingInfo: 'Metaandmed (jagamine ja litsentsiteave)',
-  fillInTheFieldsBelow : 'Täida allolevad väljad',
+  fillInTheFieldsBelow: 'Täida allolevad väljad',
   saveMetadata: 'Salvesta metaandmed',
   addAuthor: 'Salvesta autor',
   confirmRemoveAuthor: 'Oled kindel, et tahad eemaldada selle autori?',
   addNewChange: 'Lisa uus muudatus',
   confirmDeleteChangeLog: 'Oled kindel, et tahad selle muutus-logi sissekande kustutada?',
-  changelogDescription: 'Mõnede litsentside jaoks on vaja algsesse töösse ja selle tuletistesse tehtud muudatusi logida ja näidata. Tehtud muutused saad logida siia litsenseerimispõhjustel või lihtsalt selleks, et aidata sellesse sisusse tehtud muutuste üle järge pidada.',
+  changelogDescription:
+    'Mõnede litsentside jaoks on vaja algsesse töösse ja selle tuletistesse tehtud muudatusi logida ja näidata. Tehtud muutused saad logida siia litsenseerimispõhjustel või lihtsalt selleks, et aidata sellesse sisusse tehtud muutuste üle järge pidada.',
   logThisChange: 'Logi see muudatus',
   newChangeHasBeenLogged: 'Uus muudatus on logitud',
   loggedChanges: 'Logitud muudatused',
   noChangesHaveBeenLogged: 'Ei leidu logitud muudatusi',
   errorHeader: 'Juhtus viga',
-  errorCalculatingMaxScore: 'Ei saa selle sisu jaoks arvutada maksimaalseid punkte. Eeldatakse, et maksimaalsed punktid on 0. Kui see ei ole nii, siis võta ühendust oma administraatoriga.',
+  errorCalculatingMaxScore:
+    'Ei saa selle sisu jaoks arvutada maksimaalseid punkte. Eeldatakse, et maksimaalsed punktid on 0. Kui see ei ole nii, siis võta ühendust oma administraatoriga.',
   maxScoreSemanticsMissing: 'Ei saanud mahutada vajalikku semantikat sisusse mahutada.',
   copyButton: 'Kopeeri',
   copiedButton: 'Kopeeritud',
   pasteButton: 'Kleebi',
   pasteAndReplaceButton: 'Kleebi ja asenda',
   pasteContent: 'Asenda sisu',
-  confirmPasteContent: 'Seda tehes asendad olemasoleva sisu kleebitava sisuga. Senine sisu läheb kaotsi. Kas oled kindel, et soovid jätkata?',
+  confirmPasteContent:
+    'Seda tehes asendad olemasoleva sisu kleebitava sisuga. Senine sisu läheb kaotsi. Kas oled kindel, et soovid jätkata?',
   confirmPasteButtonText: 'Asenda sisu',
   copyToClipboard: 'Kopeeri H5P sisu mällu',
   copiedToClipboard: 'Sisu on mällu kopeeritud',
@@ -195,8 +205,10 @@ H5PEditor.language.core = {
   pasteError: 'Ei saa mälust kleepida',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'Hästi',
   avTablistLabel: 'Aseta, kasutades',
   tabTitleBasicFileUpload: 'Faili üleslaadimine',
@@ -209,14 +221,86 @@ H5PEditor.language.core = {
   language: 'Keel',
   noLanguagesSupported: 'Keeli ei toetata',
   changeLanguage: 'Vali uueks keeleks :language?',
-  thisWillPotentially: "Tõenäoliselt lähtestad sellega kogu teksti ja tõlked. Neid ei saa seejärel taastada. Sisu ennast ei muudeta. Kas tahad jätkata?",
+  thisWillPotentially:
+    'Tõenäoliselt lähtestad sellega kogu teksti ja tõlked. Neid ei saa seejärel taastada. Sisu ennast ei muudeta. Kas tahad jätkata?',
   notAllTextsChanged: 'Mitte kõiki tekste ei muudetud. Ainult osaliselt on kaetud - :language.',
-  contributeTranslations: 'Kui tahad lõpetada tõlge keelele :language, siis vaata lähemalt <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Kahjuks ei ole valitud sisutüüp \'%lib\' sellesse süsteemi paigaldatud.',
+  contributeTranslations:
+    'Kui tahad lõpetada tõlge keelele :language, siis vaata lähemalt <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Kahjuks ei ole valitud sisutüüp '%lib' sellesse süsteemi paigaldatud.",
   proceedButtonLabel: 'Edasi salvestama',
   enterFullscreenButtonLabel: 'Ava täisekraanis',
   exitFullscreenButtonLabel: 'Välju täisekraanist',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
