@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Gehitu :entity',
   tooLong: 'Eremuaren balioa luzeegia da, gehienez :max letra izan ditzake.',
   invalidFormat: 'Eremuaren balioak formatu baliogabea edo debekatuta dauden karaktereak dauzka.',
-  confirmChangeLibrary: 'Hori egiten baduzu oraingo eduki motaz egindako lan guztia galduko duzu. Ziur zaude ariketa mota aldatu nahi duzula?',
+  confirmChangeLibrary:
+    'Hori egiten baduzu oraingo eduki motaz egindako lan guztia galduko duzu. Ziur zaude ariketa mota aldatu nahi duzula?',
   commonFields: 'Testu gainidazteak eta itzulpenak',
   commonFieldsDescription: 'Bertan ezarpenak editatzen ahal dutuzu edo ariketa honetako testua itzul dezakezu.',
   uploading: 'Kargatzen, itxaron...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Itsatsi YouTubeko esteka edo bestelako bideo iturriaren URLa',
   uploadAudioTitle: 'Kargatu audio fitxategia',
   uploadVideoTitle: 'Kargatu bideo fitxategia',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Txertatu',
   cancel: 'Utzi',
   height: 'altuera',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Kargatu',
   uploadPlaceholder: 'Ez dago hautaturiko fitxategirik',
   uploadInstructionsTitle: 'Kargatu H5P fitxategi bat.',
-  uploadInstructionsContent: 'Has zaitezke <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a> web-orriko adibideekin.',
+  uploadInstructionsContent:
+    'Has zaitezke <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a> web-orriko adibideekin.',
   uploadFileButtonLabel: 'Kargatu fitxategi bat',
   uploadFileButtonChangeLabel: 'Aldatu fitxategia',
   uploadingThrobber: 'Oraintxe ari da kargatzen...',
@@ -84,11 +87,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Hautatu duzun fitxategia ezin da kargatu',
   h5pFileWrongExtensionContent: 'Soilik .h5p luzapena duten fitxategiak onartzen dira.',
   h5pFileValidationFailedTitle: 'Ezin da ontzat eman H5P fitxategia.',
-  h5pFileValidationFailedContent: 'Egiaztatu igo duzun H5P fitxategia benetan H5P ariketa duela. H5p' +
- ' fitxategiak soilik liburutegiak badituzte kargatu daitezke H5P     liburutegien orriaren bidezf.',
+  h5pFileValidationFailedContent:
+    'Egiaztatu igo duzun H5P fitxategia benetan H5P ariketa duela. H5p fitxategiak soilik liburutegiak badituzte kargatu daitezke H5P     liburutegien orriaren bidezf.',
   h5pFileUploadServerErrorTitle: 'H5P fitxategia ezin izan da kargatu',
-  h5pFileUploadServerErrorContent: 'Ustegabeko errorea gertatu da. Aztertu zure zerbitzariaren akatsen erregistroa' +
- ' zehaztasun gehiago izateko.',
+  h5pFileUploadServerErrorContent:
+    'Ustegabeko errorea gertatu da. Aztertu zure zerbitzariaren akatsen erregistroa zehaztasun gehiago izateko.',
   contentTypeSectionAll: 'ariketa mota guztiak',
   searchResults: 'Bilatu emaitzak',
   contentTypeSearchFieldPlaceholder: 'Bilatu ariketa motaren arabera',
@@ -118,7 +121,8 @@ H5PEditor.language.core = {
   readLess: 'Irakurri gutxiago',
   contentTypeOwner: ':owner arabera',
   contentTypeUnsupportedApiVersionTitle: 'ariketa mota honek bertsio nagusi berriago bat behar du',
-  contentTypeUnsupportedApiVersionContent: 'Jarri harremantan zure sistemaren administratzailearekin eskuratzeko behar dituzun eguneraketak',
+  contentTypeUnsupportedApiVersionContent:
+    'Jarri harremantan zure sistemaren administratzailearekin eskuratzeko behar dituzun eguneraketak',
   contentTypeUpdateAvailable: 'Eguneraketa eskuragarria',
   contentTypeRestricted: 'Ariketa mota mugatua',
   contentTypeRestrictedDesc: 'Administratzaileak ariketa mota honen erabilera mugatu du.',
@@ -128,11 +132,12 @@ H5PEditor.language.core = {
   currentMenuSelected: 'uneko hautapena',
   errorCommunicatingHubTitle: 'Ezin da komunikatu konektorearekin.',
   errorCommunicatingHubContent: 'Errore bat gertatu da. Saia zaitez beranduago.',
-  warningNoContentTypesInstalled: "Ez daukazu inolako ariketa motarik instalatuta.",
+  warningNoContentTypesInstalled: 'Ez daukazu inolako ariketa motarik instalatuta.',
   warningChangeBrowsingToSeeResults: 'Klikatu <em>Dena</em> ikusteko instala dezakezun ariketa mota guztien zerrenda.',
   warningUpdateAvailableTitle: ':contentType motaren bertsio berria eskuragarri dago.',
   warningUpdateAvailableBody: 'Eguneratu azken bertsiora esperientzia hobea izateko.',
-  licenseDescription: 'Lizentzia honen hainbat berritasun behean zerrendatuta daude. Klikatu goiko info ikonoan irakurtzeko jatorrizko lizentziaren testua.',
+  licenseDescription:
+    'Lizentzia honen hainbat berritasun behean zerrendatuta daude. Klikatu goiko info ikonoan irakurtzeko jatorrizko lizentziaren testua.',
   licenseModalTitle: 'Lizentziaren xehetasunak',
   licenseModalSubtitle: 'Hautatu lizentzia bat ikusteko erabilera egokiaren gaineko informazioa',
   licenseUnspecified: 'Zehazkabea',
@@ -152,40 +157,47 @@ H5PEditor.language.core = {
   screenshots: 'Pantaila kapturak',
   reloadButtonLabel: 'Birkargatu',
   videoQuality: 'Kalitatezko bideoaren etiketa',
-  videoQualityDescription: 'Etiketa honek laguntzen dio erabiltzaileari bideoaren uneko kalitatea identifikatzen. Esate baterako 1080p, 720p, HD edo Mugikorra',
+  videoQualityDescription:
+    'Etiketa honek laguntzen dio erabiltzaileari bideoaren uneko kalitatea identifikatzen. Esate baterako 1080p, 720p, HD edo Mugikorra',
   videoQualityDefaultLabel: 'Kalitatea :index',
   noContentTypesAvailable: 'Ez dago ariketa mota erabilgarririk',
-  noContentTypesAvailableDesc: 'Zure guneak zailtasunak ditu H5P.org-ekin konektatzeko ariketa mota erabilgarriak zerrendatzeko.',
+  noContentTypesAvailableDesc:
+    'Zure guneak zailtasunak ditu H5P.org-ekin konektatzeko ariketa mota erabilgarriak zerrendatzeko.',
   contentTypeCacheOutdated: 'Ariketa moten zerrenda zaharkitua dago',
-  contentTypeCacheOutdatedDesc: 'Zure guneak zailtasunak ditu H5P.org-ekin konektatzeko ariketa mota erabilgarriak zerrendatzeko. Ezingo duzu ariketa mota berriak instalatu edo eguneratu.',
+  contentTypeCacheOutdatedDesc:
+    'Zure guneak zailtasunak ditu H5P.org-ekin konektatzeko ariketa mota erabilgarriak zerrendatzeko. Ezingo duzu ariketa mota berriak instalatu edo eguneratu.',
   tryAgain: 'Saiatu berriro',
   getHelp: 'Lortu laguntza',
   untitled: ':libraryTitle izengabea',
   title: 'Titulua',
   metadata: 'Metadatuak',
   addTitle: 'Gehitu titulua',
-  usedForSearchingReportsAndCopyrightInformation: 'Bilaketetarako, txostenetarako eta copyright informaziorako erabilgarria',
+  usedForSearchingReportsAndCopyrightInformation:
+    'Bilaketetarako, txostenetarako eta copyright informaziorako erabilgarria',
   metadataSharingAndLicensingInfo: 'Metadatuak (partekatze eta lizentziaren gaineko informazioa)',
-  fillInTheFieldsBelow : 'Bete azpiko eremuetan',
+  fillInTheFieldsBelow: 'Bete azpiko eremuetan',
   saveMetadata: 'Gorde metadatuak',
   addAuthor: 'Gorde egilea',
   confirmRemoveAuthor: 'Ziur zaude egile hori kendu nahi duzula?',
   addNewChange: 'Gehitu aldaketa berri bat',
   confirmDeleteChangeLog: 'Ziur zaude aldaketaren erregistro sarrera hau kendu nahi duzula?',
-  changelogDescription: 'Lizentzi batzuk jatorrizko lanari egindako aldaketak edo lan eratorriak adierazi eta erakustea exijitzen dute. Zure aldaketak hemen adierazi ditzakezu lizentziaren betekin gisa edo besterik gabe besteei edo zure buruari edukiari egindako aldaketen berri emateko.',
+  changelogDescription:
+    'Lizentzi batzuk jatorrizko lanari egindako aldaketak edo lan eratorriak adierazi eta erakustea exijitzen dute. Zure aldaketak hemen adierazi ditzakezu lizentziaren betekin gisa edo besterik gabe besteei edo zure buruari edukiari egindako aldaketen berri emateko.',
   logThisChange: 'Erregistratu aldaketa',
   newChangeHasBeenLogged: 'Aldaketa berri bat erregistratu daN',
   loggedChanges: 'Erregistratutako aldaketak',
   noChangesHaveBeenLogged: 'Ez da aldaketarik erregistratu',
   errorHeader: 'Errore bat gertatu da',
-  errorCalculatingMaxScore: 'Ezin da kalkulatu ariketa honen puntuazio maximoa.  Suposatzen da puntuazio maximoa 0 dela. Jar zaitez harrematenetan administratzailearekin itsaten saiatzen ari dena ez bada onartzen testuinguru honetan.',
+  errorCalculatingMaxScore:
+    'Ezin da kalkulatu ariketa honen puntuazio maximoa.  Suposatzen da puntuazio maximoa 0 dela. Jar zaitez harrematenetan administratzailearekin itsaten saiatzen ari dena ez bada onartzen testuinguru honetan.',
   maxScoreSemanticsMissing: 'Edukian ezin izan da aurkitu espero zen semantika.',
   copyButton: 'Kopiatu',
   copiedButton: 'Kopiatu da',
   pasteButton: 'Itsatsi',
   pasteAndReplaceButton: 'Itsatsi eta Ordezkatu',
   pasteContent: 'Ordezkatu Edukia',
-  confirmPasteContent: 'Hau eginda oraingo edukia arbelean daukazunarekin ordezkatuko da. Oraingo edukia galduko da. Ziur jarraitu nahi duzula?',
+  confirmPasteContent:
+    'Hau eginda oraingo edukia arbelean daukazunarekin ordezkatuko da. Oraingo edukia galduko da. Ziur jarraitu nahi duzula?',
   confirmPasteButtonText: 'Ordezkatu edukia',
   copyToClipboard: 'Kopiatu H5P edukia to the clipboard',
   copiedToClipboard: 'Edukia arbelera kopiatzen da',
@@ -195,8 +207,10 @@ H5PEditor.language.core = {
   pasteError: 'Ezin izan da arbeltik itsatsi',
   pasteContentNotSupported: 'Arbeleko H5P edukia ez da testuinguru honetan onartzen',
   pasteContentRestricted: 'Arbeleko edukia mugatuta dago gune honetan',
-  pasteTooOld: 'Arbeleko H5P edukia testuinguru honetan onartzen den bertsioa (:local) baino bertsio baxuago batekoa da (:clip), posiblea bada saiatu zaitez itsatsi nahi duzun edukia eguneratzen, berriz kopiatu ezazu eta ondoren hemen itsaten saiatu zaitez.',
-  pasteTooNew: 'Arbeleko H5P edukia testuinguru honetan onartzen den bertsioa (:local) baino bertsio altuago batekoa da (:clip), posiblea bada saiatu zaitez itsatsi nahi duzun eduki-motaren liburutegia eguneratzen, eta ondoren berriz itsaten saiatu zaitez.',
+  pasteTooOld:
+    'Arbeleko H5P edukia testuinguru honetan onartzen den bertsioa (:local) baino bertsio baxuago batekoa da (:clip), posiblea bada saiatu zaitez itsatsi nahi duzun edukia eguneratzen, berriz kopiatu ezazu eta ondoren hemen itsaten saiatu zaitez.',
+  pasteTooNew:
+    'Arbeleko H5P edukia testuinguru honetan onartzen den bertsioa (:local) baino bertsio altuago batekoa da (:clip), posiblea bada saiatu zaitez itsatsi nahi duzun eduki-motaren liburutegia eguneratzen, eta ondoren berriz itsaten saiatu zaitez.',
   ok: 'ados',
   avTablistLabel: 'Txertatu hau erabiltzen:',
   tabTitleBasicFileUpload: 'Igo fitxategia',
@@ -209,14 +223,86 @@ H5PEditor.language.core = {
   language: 'Hizkuntza',
   noLanguagesSupported: 'Ez dago onartutako hizkuntzarik',
   changeLanguage: 'Aldatu hizkuntza honetara :language?',
-  thisWillPotentially: "Honek testu eta itzulpen guztiak berrabiarazi ditzake. Ekintza hau ezin da desegin. Edukia bera ez da aldatuko. Jarraitu nahi duzu??",
+  thisWillPotentially:
+    'Honek testu eta itzulpen guztiak berrabiarazi ditzake. Ekintza hau ezin da desegin. Edukia bera ez da aldatuko. Jarraitu nahi duzu??',
   notAllTextsChanged: 'Ez dira testu guztiak aldatu, :language hizkuntzaren sostengua ez da osoa.',
-  contributeTranslations: ':language hizkuntzako itzulpena osatu nahi baduzu <a href=":url" target="_new">contributing translations to H5P</a> estekan nola egin ikasi dezakezu',
-  unknownLibrary: 'Tamalez, aukeratutako \'%lib\' eduki-mota ez dago sistema honetan instalatuta.',
+  contributeTranslations:
+    ':language hizkuntzako itzulpena osatu nahi baduzu <a href=":url" target="_new">contributing translations to H5P</a> estekan nola egin ikasi dezakezu',
+  unknownLibrary: "Tamalez, aukeratutako '%lib' eduki-mota ez dago sistema honetan instalatuta.",
   proceedButtonLabel: 'Jarraitu eta gorde',
   enterFullscreenButtonLabel: 'Erabili pantaila osoko modura',
   exitFullscreenButtonLabel: 'Irten pantaila osoko modutik',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
