@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Lisää :entity',
   tooLong: 'Kentän arvo on liiaan pitkä. Saa sisältää :max kirjainta tai vähemmän.',
   invalidFormat: 'Kentän arvo on väärässä muodossa tai sisältää kiellettyjä merkkejä.',
-  confirmChangeLibrary: 'Tämä hävittää tekemäsi työn nykyiselle sisältötyypille. Oletko varma, että haluat vaihtaa sisältötyyppiä?',
+  confirmChangeLibrary:
+    'Tämä hävittää tekemäsi työn nykyiselle sisältötyypille. Oletko varma, että haluat vaihtaa sisältötyyppiä?',
   commonFields: 'Omat tekstit ja käännökset',
   commonFieldsDescription: 'Täällä voit muuttaa asetuksia tai kääntää tekstejä tälle sisällölle.',
   uploading: 'Lataa, odota...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Liitä YouTube linkki tai muu videon osoite',
   uploadAudioTitle: 'Lataa äänitiedosto',
   uploadVideoTitle: 'Lataa videotiedosto',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Lisää',
   cancel: 'Peruuta',
   height: 'korkeus',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Lataa',
   uploadPlaceholder: 'Ei tiedostoa valittuna',
   uploadInstructionsTitle: 'Lataa H5P tiedosto.',
-  uploadInstructionsContent: 'Sisältöesimerkkejä löytyy osoitteesta <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Sisältöesimerkkejä löytyy osoitteesta <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Lataa tiedosto',
   uploadFileButtonChangeLabel: 'Vaihda tiedosto',
   uploadingThrobber: 'Lataa...',
@@ -84,7 +87,8 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Valittua tiedostoa ei pystytty lataamaan',
   h5pFileWrongExtensionContent: 'Vain .h5p päätteiset tiedostot ovat sallittuja.',
   h5pFileValidationFailedTitle: 'H5P tiedostoa ei pystytty varmistamaan.',
-  h5pFileValidationFailedContent: 'Varmista että ladattu H5P tiedosto sisältää H5P sisältöä. H5P tiedostot joissa on ainoastaan kirjastoja täytyy ladata H5P Libraries sivulla.',
+  h5pFileValidationFailedContent:
+    'Varmista että ladattu H5P tiedosto sisältää H5P sisältöä. H5P tiedostot joissa on ainoastaan kirjastoja täytyy ladata H5P Libraries sivulla.',
   h5pFileUploadServerErrorTitle: 'H5P tiedostoa ei pystytty lataamaan',
   h5pFileUploadServerErrorContent: 'Odottamaton virhe. Tarkista palvelinlogisi lisätietoja varten.',
   contentTypeSectionAll: 'Kaikki sisältötyypit',
@@ -127,10 +131,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Yhteyttä hubiin ei saatu muodostettua.',
   errorCommunicatingHubContent: 'Virhe tapahtui. Yritä uudelleen.',
   warningNoContentTypesInstalled: 'Sinulla ei ole yhtään sisältötyyppiä asennettuna.',
-  warningChangeBrowsingToSeeResults: 'Klikkaa <em>Kaikki</em> saadaksesi listan kaikista asennettavista sisältötyypeistä.',
+  warningChangeBrowsingToSeeResults:
+    'Klikkaa <em>Kaikki</em> saadaksesi listan kaikista asennettavista sisältötyypeistä.',
   warningUpdateAvailableTitle: 'Sisällöstä :contentType on uusi versio saatavilla.',
   warningUpdateAvailableBody: 'Päivitä viimeisimpään versioon parannetuilla ominaisuuksilla.',
-  licenseDescription: 'Tämän lisenssin jotkin ominaisuudet on annettu alempana. Klikkaa infoikonia lukeaksi alkuperäisen lisenssitekstin.',
+  licenseDescription:
+    'Tämän lisenssin jotkin ominaisuudet on annettu alempana. Klikkaa infoikonia lukeaksi alkuperäisen lisenssitekstin.',
   licenseModalTitle: 'Lisenssin yksityiskohdat',
   licenseModalSubtitle: 'Valitse lisenssi tarkastellaksesi tietoa oikeasta käytöstä',
   licenseUnspecified: 'Määrittelemätön',
@@ -150,12 +156,15 @@ H5PEditor.language.core = {
   screenshots: 'Kuvakaappaukset',
   reloadButtonLabel: 'Lataa uudelleen',
   videoQuality: 'Videon laadun nimike',
-  videoQualityDescription: 'Tämä nimike auttaa käyttäjää tunnistamaan videon nykyisen laadun. Esim. 1080p, 720p, HD tai Mobiili',
+  videoQualityDescription:
+    'Tämä nimike auttaa käyttäjää tunnistamaan videon nykyisen laadun. Esim. 1080p, 720p, HD tai Mobiili',
   videoQualityDefaultLabel: 'Laatu :index',
   noContentTypesAvailable: 'Sisältötyyppejä ei saatavilla.',
-  noContentTypesAvailableDesc: 'Sivustollasi on ongelmia yhdistettäessä h5p.org-sivustoon ja sisältötyyppien listaus ei tästä syystä toimi.',
+  noContentTypesAvailableDesc:
+    'Sivustollasi on ongelmia yhdistettäessä h5p.org-sivustoon ja sisältötyyppien listaus ei tästä syystä toimi.',
   contentTypeCacheOutdated: 'Sisältötyyppilista on vanhentunut',
-  contentTypeCacheOutdatedDesc: 'Sivustollasi on ongelmia yhdistettäessä h5p.org-sivustoon tarkistaakseen onko uusia sisältötyyppipäivityksiä saatavilla. Et välttämättä pysty päivittämään vanhoja etkä lataamaan uusia sisältötyyppejä.',
+  contentTypeCacheOutdatedDesc:
+    'Sivustollasi on ongelmia yhdistettäessä h5p.org-sivustoon tarkistaakseen onko uusia sisältötyyppipäivityksiä saatavilla. Et välttämättä pysty päivittämään vanhoja etkä lataamaan uusia sisältötyyppejä.',
   tryAgain: 'Yritä uudelleen',
   getHelp: 'Apu',
   untitled: 'Untitled :libraryTitle',
@@ -170,20 +179,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Oletko varma että halua että haluat poistaa tämän tekijän?',
   addNewChange: 'Lisää uusi muutos',
   confirmDeleteChangeLog: 'Oletko varma että haluat poistaa tämän muutoslogitietueen?',
-  changelogDescription: 'Jotkut lisenssit vaativat että alkuperäiseen työhön tehtävä muutokset merkitään logiin ja esitetään. Voit merkitä nämä muutokset logiin tänne tai vain itseäsi varten jotta pysyt kärryillä mitä tietoja olet muuttanut.',
+  changelogDescription:
+    'Jotkut lisenssit vaativat että alkuperäiseen työhön tehtävä muutokset merkitään logiin ja esitetään. Voit merkitä nämä muutokset logiin tänne tai vain itseäsi varten jotta pysyt kärryillä mitä tietoja olet muuttanut.',
   logThisChange: 'Kirjoita tämä muutos logiin',
   newChangeHasBeenLogged: 'Uusi muutos kirjattiin logiin',
   loggedChanges: 'Muutokset logitettiin.',
   noChangesHaveBeenLogged: 'Yhtään muutosta ei kirjoitettu logiin',
   errorHeader: 'Tapahtui virhe',
-  errorCalculatingMaxScore: 'Maksimiarvosanaa ei pystytty laskemaan tälle sisällölle, täten maksimiarvosanaksi asetetaan 0. Olethan yhteydessä ylläpitäjiin mikäli tämä on virhetilanne.',
+  errorCalculatingMaxScore:
+    'Maksimiarvosanaa ei pystytty laskemaan tälle sisällölle, täten maksimiarvosanaksi asetetaan 0. Olethan yhteydessä ylläpitäjiin mikäli tämä on virhetilanne.',
   maxScoreSemanticsMissing: 'Odotettuja semantiikkoja ei löydetty aineistosta.',
   copyButton: 'Kopioi',
   copiedButton: 'Kopioitu',
   pasteButton: 'Liitä',
   pasteAndReplaceButton: 'Liitä & Korvaa',
   pasteContent: 'Korvaa sisältö',
-  confirmPasteContent: 'Jos teet tämän niin korvaat nykyisen sisällön sisällöllä jonka olet liittämässä leikepöydältä. Nykyinen sisältö menetetään pysyvästi. Oletko varma että haluat jatkaa?',
+  confirmPasteContent:
+    'Jos teet tämän niin korvaat nykyisen sisällön sisällöllä jonka olet liittämässä leikepöydältä. Nykyinen sisältö menetetään pysyvästi. Oletko varma että haluat jatkaa?',
   confirmPasteButtonText: 'Korvaa sisältö',
   copyToClipboard: 'Kopioi H5P sisältö leikepöydälle',
   copiedToClipboard: 'H5P sisältö on nyt kopioitu leikepöydälle',
@@ -193,8 +205,10 @@ H5PEditor.language.core = {
   pasteError: 'Ei pystytä liittämään leikepöydältä',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Lisää käyttäen',
   tabTitleBasicFileUpload: 'Tiedoston lähetys',
@@ -207,14 +221,87 @@ H5PEditor.language.core = {
   language: 'Kieli',
   noLanguagesSupported: 'Ei tuettuja kieliä',
   changeLanguage: 'Vaihda kieleksi :language?',
-  thisWillPotentially: "Tämä mahdollisesti nollaa kaikki tähän asti tehdy kielikäännösmuutokset. Tätä ei voi perua. Itse aineisto ei muutu. Haluatko varmasti jatkaa?",
+  thisWillPotentially:
+    'Tämä mahdollisesti nollaa kaikki tähän asti tehdy kielikäännösmuutokset. Tätä ei voi perua. Itse aineisto ei muutu. Haluatko varmasti jatkaa?',
   notAllTextsChanged: 'Kaikkia tekstejä ei voitu vaihtaa :language kielelle koska niiden käännökset puuttuvat',
-  contributeTranslations: 'Jos haluat auttaa käännöstyössä löydät lisätietoja täältä:<a href=":url" target="_new">H5P käännöksissä auttaminen</a>',
-  unknownLibrary: 'Valitettavasti valittua sisältötyyppiä \'%lib\' ei ole asennettuna tähän järjestelmään. Olethan yhteydessä ylläpitoon.',
+  contributeTranslations:
+    'Jos haluat auttaa käännöstyössä löydät lisätietoja täältä:<a href=":url" target="_new">H5P käännöksissä auttaminen</a>',
+  unknownLibrary:
+    "Valitettavasti valittua sisältötyyppiä '%lib' ei ole asennettuna tähän järjestelmään. Olethan yhteydessä ylläpitoon.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
