@@ -11,13 +11,13 @@ H5PEditor.language.core = {
   exceedsMax: 'La valeur du champ :property est supérieure au maximum de :max.',
   listExceedsMax: 'La liste dépasse le maximum de :max éléments.',
   belowMin: 'La valeur du champ :property est inférieure au minimum de :min.',
-  listBelowMin: 'La liste requiert un minimum de :min éléments pour que l\'activité fonctionne correctement.',
+  listBelowMin: "La liste requiert un minimum de :min éléments pour que l'activité fonctionne correctement.",
   outOfStep: 'La :property ne peut être effectuée que par étape :step.',
   add: 'Ajouter',
   addFile: 'Ajouter un fichier',
   removeFile: 'Supprimer un fichier',
   confirmRemoval: 'Voulez-vous vraiment supprimer :type&nbsp;?',
-  removeImage: 'Supprimer l\'image',
+  removeImage: "Supprimer l'image",
   confirmImageRemoval: 'Votre image va être supprimée. Êtes-vous sûr de vouloir continuer ?',
   changeFile: 'Changer le fichier',
   changeLibrary: 'Changer le type de contenu',
@@ -37,34 +37,35 @@ H5PEditor.language.core = {
   tutorial: 'Tutoriel',
   editMode: 'Mode de rédaction',
   listLabel: 'Liste',
-  uploadError: 'Erreur pendant l\'envoi du fichier',
+  uploadError: "Erreur pendant l'envoi du fichier",
   fileToLarge: 'Le fichier que vous envoyez est trop lourd.',
   unknownFileUploadError: 'Erreur inconnue lors du téléversement du fichier',
   noSemantics: 'Erreur, impossible de charger ce type de contenu.',
-  editImage: 'Editer l\'image',
+  editImage: "Editer l'image",
   saveLabel: 'Enregistrer',
   cancelLabel: 'Annuler',
   resetToOriginalLabel: 'Réinitialiser',
-  loadingImageEditor: 'Editeur d\'image en cours de chargement, patientez...',
+  loadingImageEditor: "Editeur d'image en cours de chargement, patientez...",
   selectFiletoUpload: 'Choisissez le fichier à charger',
   or: 'ou',
-  enterAudioUrl: 'Entrez l\'URL de la source audio',
+  enterAudioUrl: "Entrez l'URL de la source audio",
   enterVideoUrl: 'Enter video URL',
-  enterAudioTitle: 'Coller le lien ou l\'URL d\'une autre source audio',
-  enterVideoTitle: 'Coller le lien YouTube link ou l\'URL d\'une autre source vidéo',
+  enterAudioTitle: "Coller le lien ou l'URL d'une autre source audio",
+  enterVideoTitle: "Coller le lien YouTube link ou l'URL d'une autre source vidéo",
   uploadAudioTitle: 'Téléverser le fichier audio',
   uploadVideoTitle: 'Téléverser le fichier vidéo',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Insérer',
   cancel: 'Annuler',
   height: 'hauteur',
   width: 'largeur',
   textField: 'texte',
   numberField: 'numérique',
-  orderItemUp: 'Monter l\'élément',
-  orderItemDown: 'Descendre l\'élément',
-  removeItem: 'Supprimer l\'élément',
-  hubPanelLabel: 'Sélectionner le type d\'activité',
+  orderItemUp: "Monter l'élément",
+  orderItemDown: "Descendre l'élément",
+  removeItem: "Supprimer l'élément",
+  hubPanelLabel: "Sélectionner le type d'activité",
   importantInstructions: 'Instructions importantes',
   showImportantInstructions: 'Afficher les instructions',
   hideImportantInstructions: 'Masquer les instructions importantes',
@@ -74,22 +75,25 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Téléverser',
   uploadPlaceholder: 'Aucun fichier sélectionné',
   uploadInstructionsTitle: 'Téléverser un fichier H5P.',
-  uploadInstructionsContent: 'Inspirez-vous des exemples disponibles sur <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Inspirez-vous des exemples disponibles sur <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Téléverser un fichier',
   uploadFileButtonChangeLabel: 'Modifier le fichier',
   uploadingThrobber: 'Téléversement en cours...',
   uploadSuccess: ':title a été téléversé avec succès !',
-  unableToInterpretError: 'Impossible d\'interpréter la réponse.',
+  unableToInterpretError: "Impossible d'interpréter la réponse.",
   unableToInterpretSolution: 'Merci de vérifier le journal des erreurs.',
-  h5pFileWrongExtensionTitle: 'Le fichier sélectionné n\'a pas pu être téléversé.',
-  h5pFileWrongExtensionContent: 'Seuls les fichiers avec l\'extentions .h5p sont autorisés.',
+  h5pFileWrongExtensionTitle: "Le fichier sélectionné n'a pas pu être téléversé.",
+  h5pFileWrongExtensionContent: "Seuls les fichiers avec l'extentions .h5p sont autorisés.",
   h5pFileValidationFailedTitle: 'Impossible de valider le fichier H5P.',
-  h5pFileValidationFailedContent: 'Merci de vérifier que le fichier H5P téléversé contient un contenu H5P valide. Les fichiers contenant uniquement des bibliothèques doivent être téléversés en utilisant la page des Librairies H5P.',
-  h5pFileUploadServerErrorTitle: 'Le fichier H5P n\'a pas pu être téléversé.',
-  h5pFileUploadServerErrorContent: 'Une erreur inattendue s\'est produite. Vérifier le journal des erreurs de votre serveur pour plus d\'informations.',
-  contentTypeSectionAll: 'Tous les types d\'activités',
+  h5pFileValidationFailedContent:
+    'Merci de vérifier que le fichier H5P téléversé contient un contenu H5P valide. Les fichiers contenant uniquement des bibliothèques doivent être téléversés en utilisant la page des Librairies H5P.',
+  h5pFileUploadServerErrorTitle: "Le fichier H5P n'a pas pu être téléversé.",
+  h5pFileUploadServerErrorContent:
+    "Une erreur inattendue s'est produite. Vérifier le journal des erreurs de votre serveur pour plus d'informations.",
+  contentTypeSectionAll: "Tous les types d'activités",
   searchResults: 'Résultats de la recherche',
-  contentTypeSearchFieldPlaceholder: 'Rechercher des types d\'activités',
+  contentTypeSearchFieldPlaceholder: "Rechercher des types d'activités",
   contentTypeInstallButtonLabel: 'Installer',
   contentTypeInstallingButtonLabel: 'Installation en cours',
   contentTypeUseButtonLabel: 'Utiliser',
@@ -101,9 +105,9 @@ H5PEditor.language.core = {
   contentTypeIconAltText: 'Icône',
   contentTypeInstallSuccess: ':contentType installé avec succès !',
   contentTypeUpdateSuccess: ':contentType mis à jour avec succès !',
-  contentTypeInstallError: ':contentType n\'a pas pu être installé. Contactez votre administrateur.',
+  contentTypeInstallError: ":contentType n'a pas pu être installé. Contactez votre administrateur.",
   contentTypeLicensePanelTitle: 'Licence',
-  contentTypeDemoButtonLabel: 'Démonstration de l\'activité',
+  contentTypeDemoButtonLabel: "Démonstration de l'activité",
   numResults: ':num résultats',
   show: 'Afficher',
   recentlyUsedFirst: 'Utilisés récemment en premier',
@@ -111,28 +115,31 @@ H5PEditor.language.core = {
   newestFirst: 'Plus récents en premier',
   aToZ: 'A à Z',
   noResultsFound: 'Aucun résultat trouvé',
-  noResultsFoundDesc: 'Il n\'y a aucune activité qui corresponde à vos critères de recherche.',
+  noResultsFoundDesc: "Il n'y a aucune activité qui corresponde à vos critères de recherche.",
   readMore: 'Lire plus',
   readLess: 'Lire moins',
   contentTypeOwner: 'Par :owner',
-  contentTypeUnsupportedApiVersionTitle: 'Ce type d\'activité requiert une version du noyau H5P plus récente.',
-  contentTypeUnsupportedApiVersionContent: 'Contactez votre administrateur système pour qu\'il vous fournisse les mises à jour nécessaires.',
+  contentTypeUnsupportedApiVersionTitle: "Ce type d'activité requiert une version du noyau H5P plus récente.",
+  contentTypeUnsupportedApiVersionContent:
+    "Contactez votre administrateur système pour qu'il vous fournisse les mises à jour nécessaires.",
   contentTypeUpdateAvailable: 'Mise à jour disponible',
-  contentTypeRestricted: 'Type d\'activité à usage restreint',
-  contentTypeRestrictedDesc: 'L\'utilisation de ce type d\'activité a été restreint par un administrateur.',
-  contentTypeNotInstalled: 'Type d\'activité non installé',
-  contentTypeNotInstalledDesc: 'Vous n\'avez pas les droits requis pour installer des types d\'activités.',
-  theContentType: 'Le type d\'activité',
+  contentTypeRestricted: "Type d'activité à usage restreint",
+  contentTypeRestrictedDesc: "L'utilisation de ce type d'activité a été restreint par un administrateur.",
+  contentTypeNotInstalled: "Type d'activité non installé",
+  contentTypeNotInstalledDesc: "Vous n'avez pas les droits requis pour installer des types d'activités.",
+  theContentType: "Le type d'activité",
   currentMenuSelected: 'sélection en cours',
   errorCommunicatingHubTitle: 'Impossible de communiquer avec le hub.',
-  errorCommunicatingHubContent: 'Une erreur s\'est produite. Merci d\'essayer à nouveau.',
-  warningNoContentTypesInstalled: 'Vous n\'avez aucun type d\'activité installé.',
-  warningChangeBrowsingToSeeResults: 'Cliquez <em>Tout</em> pour afficher la liste de tous les types d\'activités que vous pouvez installer.',
+  errorCommunicatingHubContent: "Une erreur s'est produite. Merci d'essayer à nouveau.",
+  warningNoContentTypesInstalled: "Vous n'avez aucun type d'activité installé.",
+  warningChangeBrowsingToSeeResults:
+    "Cliquez <em>Tout</em> pour afficher la liste de tous les types d'activités que vous pouvez installer.",
   warningUpdateAvailableTitle: 'Une nouvelle version de :contentType est disponible.',
-  warningUpdateAvailableBody: 'Mettez à jour vers la dernière version pour profiter d\'une expérience améliorée.',
-  licenseDescription: 'Quelques caractéristiques de cette licence sont indiquées ci-dessous. Cliquez l\'icône information au dessus pour afficher le texte de la licence originelle.',
+  warningUpdateAvailableBody: "Mettez à jour vers la dernière version pour profiter d'une expérience améliorée.",
+  licenseDescription:
+    "Quelques caractéristiques de cette licence sont indiquées ci-dessous. Cliquez l'icône information au dessus pour afficher le texte de la licence originelle.",
   licenseModalTitle: 'Détails de la licence',
-  licenseModalSubtitle: 'Sélectionner une licence pour afficher les informations à propos de ses modalités d\'usage.',
+  licenseModalSubtitle: "Sélectionner une licence pour afficher les informations à propos de ses modalités d'usage.",
   licenseUnspecified: 'Non spécifié',
   licenseCanUseCommercially: 'Utilisation commerciale autorisée',
   licenseCanModify: 'Modification autorisée',
@@ -147,43 +154,50 @@ H5PEditor.language.core = {
   imageLightBoxProgress: ':num sur :total',
   nextImage: 'Image suivante',
   previousImage: 'Image précédente',
-  screenshots: 'Captures d\'écran',
+  screenshots: "Captures d'écran",
   reloadButtonLabel: 'Recharger',
   videoQuality: 'Label pour la qualité vidéo',
-  videoQualityDescription: 'Ce label permet à l\'utilisateur d\'identifier la qualité de la vidéo, par exemple 1080p, 720p, HD ou Mobile',
+  videoQualityDescription:
+    "Ce label permet à l'utilisateur d'identifier la qualité de la vidéo, par exemple 1080p, 720p, HD ou Mobile",
   videoQualityDefaultLabel: 'Qualité :index',
   noContentTypesAvailable: 'Aucun type de contenu disponible',
-  noContentTypesAvailableDesc: 'Votre site rencontre des difficultés pour se connecter à H5P.org et afficher la liste des types d\'activités disponibles.',
-  contentTypeCacheOutdated: 'Liste des types d\'activités obsolète',
-  contentTypeCacheOutdatedDesc: 'Votre site rencontre des difficultés pour se connecter à H5P.org pour vérifier les mises à jour des types d\'activités. Vous pourriez être incapable de mettre à jour ou d\'installer de nouveaux types d\'activités.',
+  noContentTypesAvailableDesc:
+    "Votre site rencontre des difficultés pour se connecter à H5P.org et afficher la liste des types d'activités disponibles.",
+  contentTypeCacheOutdated: "Liste des types d'activités obsolète",
+  contentTypeCacheOutdatedDesc:
+    "Votre site rencontre des difficultés pour se connecter à H5P.org pour vérifier les mises à jour des types d'activités. Vous pourriez être incapable de mettre à jour ou d'installer de nouveaux types d'activités.",
   tryAgain: 'Recommencer',
-  getHelp: 'Obtenir de l\'aide',
+  getHelp: "Obtenir de l'aide",
   untitled: 'Sans titre :libraryTitle',
   title: 'Title',
   metadata: 'Métadonnées',
   addTitle: 'Ajouter un titre',
-  usedForSearchingReportsAndCopyrightInformation: 'Utilisé pour la recherche, les rapports et les informations de copyright',
+  usedForSearchingReportsAndCopyrightInformation:
+    'Utilisé pour la recherche, les rapports et les informations de copyright',
   metadataSharingAndLicensingInfo: 'Métadonnées (informations de partage et de licence)',
   fillInTheFieldsBelow: 'Remplissez les champs ci-dessous',
   saveMetadata: 'Enregistrer les métadonnées',
-  addAuthor: 'Enregistrer l\'auteur',
+  addAuthor: "Enregistrer l'auteur",
   confirmRemoveAuthor: 'Êtes-vous sûr de vouloir supprimer cet auteur?',
   addNewChange: 'Ajouter une nouvelle modification',
   confirmDeleteChangeLog: 'Êtes-vous sûr de vouloir supprimer cette entrée du journal des modifications ?',
-  changelogDescription: 'Certaines licences nécessitent que les modifications apportées au travail original ou aux dérivés soient consignées et affichées. Vous pouvez enregistrer vos modifications ici pour des raisons de licence ou simplement pour vous permettre, à vous-même et à d\'autres, de suivre les modifications apportées à ce contenu.',
+  changelogDescription:
+    "Certaines licences nécessitent que les modifications apportées au travail original ou aux dérivés soient consignées et affichées. Vous pouvez enregistrer vos modifications ici pour des raisons de licence ou simplement pour vous permettre, à vous-même et à d'autres, de suivre les modifications apportées à ce contenu.",
   logThisChange: 'Enregistrer cette modification',
   newChangeHasBeenLogged: 'La nouvelle modification a été enregistrée',
   loggedChanges: 'Modifications enregistrées',
   noChangesHaveBeenLogged: 'Aucune modification enregistrée',
-  errorHeader: 'Une erreur s\'est produite',
-  errorCalculatingMaxScore: 'Impossible de calculer le score maximal pour ce contenu. La note maximale est supposée être 0. Contactez votre administrateur si cela n’est pas correct.',
+  errorHeader: "Une erreur s'est produite",
+  errorCalculatingMaxScore:
+    'Impossible de calculer le score maximal pour ce contenu. La note maximale est supposée être 0. Contactez votre administrateur si cela n’est pas correct.',
   maxScoreSemanticsMissing: 'Impossible de trouver la sémantique attendue dans le contenu.',
   copyButton: 'Copier',
   copiedButton: 'Copié',
   pasteButton: 'Coller',
   pasteAndReplaceButton: 'Coller et Remplacer',
   pasteContent: 'Remplacer le contenu',
-  confirmPasteContent: 'En faisant cela, vous remplacerez le contenu actuel par le contenu de votre presse-papiers. Le contenu actuel sera perdu. Êtes-vous sûr de vouloir continuer?',
+  confirmPasteContent:
+    'En faisant cela, vous remplacerez le contenu actuel par le contenu de votre presse-papiers. Le contenu actuel sera perdu. Êtes-vous sûr de vouloir continuer?',
   confirmPasteButtonText: 'Remplacer le contenu',
   copyToClipboard: 'Copier le contenu H5P dans le Presse-papiers',
   copiedToClipboard: 'Le contenu est copié dans le presse-papier',
@@ -193,13 +207,16 @@ H5PEditor.language.core = {
   pasteError: 'Impossible de coller le contenu du presse-papiers',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insérer via',
   tabTitleBasicFileUpload: 'Téléverser fichier',
   tabTitleInputLinkURL: 'Lien/URL',
-  errorTooHighVersion: 'Les paramètres contiennent le paramètre %used alors que seulement le paramètre %supported ou antérieur est pris en charge.',
+  errorTooHighVersion:
+    'Les paramètres contiennent le paramètre %used alors que seulement le paramètre %supported ou antérieur est pris en charge.',
   errorNotSupported: 'Les paramètres contiennent le paramètre %used qui n’est pas pris en charge.',
   errorParamsBroken: 'Paramètres invalides.',
   libraryMissing: 'Bibliothèque requise manquante %lib.',
@@ -207,14 +224,87 @@ H5PEditor.language.core = {
   language: 'Langue',
   noLanguagesSupported: 'Aucune langue supportée',
   changeLanguage: 'Changer la langue pour :language ?',
-  thisWillPotentially: "Cela réinitialisera potentiellement tout le texte et les traductions. Vous ne pouvez pas annuler cela. Le contenu lui-même ne sera pas changé. Voulez-vous poursuivre ?",
-  notAllTextsChanged: 'Les textes n\'ont pas tous été traduits, la traduction pour la langue :language est incomplète.',
-  contributeTranslations: 'Si vous souhaitez compléter la traduction de : language, cliquez pour en savoir plus sur la façon de <a href=":url" target="_new"> contribuer aux traductions de H5P </a>.',
-  unknownLibrary: 'Malheureusement, la bibliothèque pour le type de contenu sélectionné \'%lib\' n\'est pas installée sur votre système.',
+  thisWillPotentially:
+    'Cela réinitialisera potentiellement tout le texte et les traductions. Vous ne pouvez pas annuler cela. Le contenu lui-même ne sera pas changé. Voulez-vous poursuivre ?',
+  notAllTextsChanged: "Les textes n'ont pas tous été traduits, la traduction pour la langue :language est incomplète.",
+  contributeTranslations:
+    'Si vous souhaitez compléter la traduction de : language, cliquez pour en savoir plus sur la façon de <a href=":url" target="_new"> contribuer aux traductions de H5P </a>.',
+  unknownLibrary:
+    "Malheureusement, la bibliothèque pour le type de contenu sélectionné '%lib' n'est pas installée sur votre système.",
   proceedButtonLabel: 'Enregistrer',
   enterFullscreenButtonLabel: 'Afficher en plein écran',
   exitFullscreenButtonLabel: 'Quitter le mode plein écran',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
