@@ -1,201 +1,214 @@
 H5PEditor.language.core = {
-  missingTranslation: "[Traduzione mancante :key]",
-  loading: "Caricamento, si prega di attendere...",
-  selectLibrary: "Seleziona la libreria che userai per i tuoi contenuti.",
-  unknownFieldPath: "Impossibibile trovare :path.",
+  missingTranslation: '[Traduzione mancante :key]',
+  loading: 'Caricamento, si prega di attendere...',
+  selectLibrary: 'Seleziona la libreria che userai per i tuoi contenuti.',
+  unknownFieldPath: 'Impossibibile trovare :path.',
   notImageField: ":path non è un'immagine.",
   notImageOrDimensionsField: ":path non è un'immagine o dimensioni campo.",
-  requiredProperty: ":property è richiesta e deve avere un valore.",
-  onlyNumbers: "Il valore di :property può contenere solo numeri.",
+  requiredProperty: ':property è richiesta e deve avere un valore.',
+  onlyNumbers: 'Il valore di :property può contenere solo numeri.',
   illegalDecimalNumber: ':property can only contain numbers with max :decimals decimals.',
-  exceedsMax: "Il valore di :property eccede il massimo di :max.",
-  listExceedsMax: "The list exceeds the maximum of :max items.",
-  belowMin: "Il valore di :property eccede il minimo di :min.",
-  listBelowMin: "The list needs at least :min items for the content to function properly.",
-  outOfStep: "Il valore di :property può essere modificato solo negli step di :step.",
-  add: "Add",
-  addFile: "Aggiungi file",
-  removeFile: "Rimuovi file",
-  confirmRemoval: "Sei sicuro di voler rimuovere questo :type?",
-  removeImage: "Rimuovi immagine",
+  exceedsMax: 'Il valore di :property eccede il massimo di :max.',
+  listExceedsMax: 'The list exceeds the maximum of :max items.',
+  belowMin: 'Il valore di :property eccede il minimo di :min.',
+  listBelowMin: 'The list needs at least :min items for the content to function properly.',
+  outOfStep: 'Il valore di :property può essere modificato solo negli step di :step.',
+  add: 'Add',
+  addFile: 'Aggiungi file',
+  removeFile: 'Rimuovi file',
+  confirmRemoval: 'Sei sicuro di voler rimuovere questo :type?',
+  removeImage: 'Rimuovi immagine',
   confirmImageRemoval: "Questo rimuoverà l'immagine . Sei sicuro di voler procedere ?",
-  changeFile: "Cambia file",
-  changeLibrary: "Cambiare tipo di contenuto",
-  semanticsError: "Errore semantico: :error",
-  missingProperty: "Nel campo :index manca la sua proprietà :property.",
-  expandCollapse: "Espandi/Collassa",
-  addEntity: "Aggiungi :entity",
-  tooLong: "Il valore del campo è troppo lungo, dovrebbe contenere al massimo :max lettere.",
-  invalidFormat: "Il valore del campo contiene un formato non valido o caratteri non consentiti.",
-  confirmChangeLibrary: "Sei sicuro di voler cambiare libreria?",
-  commonFields: "Impostazioni e testi",
-  commonFieldsDescription: "Da qui puoi modificare le impostazioni o tradurre testi utilizzati in questo contenuto.",
-  uploading: "Caricamento, si prega di attendere...",
-  noFollow: "Impossibile seguire il campo :path.",
-  editCopyright: "Modifica copyright",
-  close: "Chiudi",
-  tutorial: "Tutorial",
-  editMode: "Modalità modifica",
-  listLabel: "Lista",
-  uploadError: "File Upload Error",
-  fileToLarge: "The file you are trying to upload might be too large.",
-  unknownFileUploadError: "Unknown file upload error",
-  noSemantics: "Error, could not load the content type form.",
-  editImage: "Edit image",
-  saveLabel: "Save",
-  cancelLabel: "Cancel",
-  resetToOriginalLabel: "Reset to original",
-  loadingImageEditor: "Loading image editor, please wait...",
-  selectFiletoUpload: "Select file to upload",
-  or: "or",
-  enterAudioUrl: "Enter audio source URL",
+  changeFile: 'Cambia file',
+  changeLibrary: 'Cambiare tipo di contenuto',
+  semanticsError: 'Errore semantico: :error',
+  missingProperty: 'Nel campo :index manca la sua proprietà :property.',
+  expandCollapse: 'Espandi/Collassa',
+  addEntity: 'Aggiungi :entity',
+  tooLong: 'Il valore del campo è troppo lungo, dovrebbe contenere al massimo :max lettere.',
+  invalidFormat: 'Il valore del campo contiene un formato non valido o caratteri non consentiti.',
+  confirmChangeLibrary: 'Sei sicuro di voler cambiare libreria?',
+  commonFields: 'Impostazioni e testi',
+  commonFieldsDescription: 'Da qui puoi modificare le impostazioni o tradurre testi utilizzati in questo contenuto.',
+  uploading: 'Caricamento, si prega di attendere...',
+  noFollow: 'Impossibile seguire il campo :path.',
+  editCopyright: 'Modifica copyright',
+  close: 'Chiudi',
+  tutorial: 'Tutorial',
+  editMode: 'Modalità modifica',
+  listLabel: 'Lista',
+  uploadError: 'File Upload Error',
+  fileToLarge: 'The file you are trying to upload might be too large.',
+  unknownFileUploadError: 'Unknown file upload error',
+  noSemantics: 'Error, could not load the content type form.',
+  editImage: 'Edit image',
+  saveLabel: 'Save',
+  cancelLabel: 'Cancel',
+  resetToOriginalLabel: 'Reset to original',
+  loadingImageEditor: 'Loading image editor, please wait...',
+  selectFiletoUpload: 'Select file to upload',
+  or: 'or',
+  enterAudioUrl: 'Enter audio source URL',
   enterVideoUrl: 'Enter video URL',
-  enterAudioTitle: "Paste link or other audio source URL",
-  enterVideoTitle: "Paste YouTube link or other video source URL",
-  uploadAudioTitle: "Upload audio file",
-  uploadVideoTitle: "Upload video file",
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
-  insert: "Insert",
-  cancel: "Cancel",
-  height: "height",
-  width: "width",
-  textField: "text field",
-  numberField: "number field",
-  orderItemUp: "Order item up",
-  orderItemDown: "Order item down",
-  removeItem: "Remove item",
-  hubPanelLabel: "Select content type",
-  importantInstructions: "Important instructions",
-  showImportantInstructions: "Show instructions",
-  hideImportantInstructions: "Hide important instructions",
-  hide: "Hide",
-  example: "Example",
-  createContentTabLabel: "Create Content",
-  uploadTabLabel: "Upload",
-  uploadPlaceholder: "No file chosen",
-  uploadInstructionsTitle: "Upload an H5P file.",
-  uploadInstructionsContent: 'You may start with examples from <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
-  uploadFileButtonLabel: "Upload a file",
-  uploadFileButtonChangeLabel: "Change file",
-  uploadingThrobber: "Now uploading...",
-  uploadSuccess: ":title was successfully uploaded!",
-  unableToInterpretError: "Unable to interpret response.",
-  unableToInterpretSolution: "Please check your error log.",
-  h5pFileWrongExtensionTitle: "The selected file could not be uploaded",
-  h5pFileWrongExtensionContent: "Only files with the .h5p extension are allowed.",
-  h5pFileValidationFailedTitle: "Could not validate H5P file.",
-  h5pFileValidationFailedContent: "Make sure the uploaded H5P contains valid H5P content. H5P files containing only libraries should be uploaded through the H5P Libraries page.",
-  h5pFileUploadServerErrorTitle: "The H5P file could not be uploaded",
-  h5pFileUploadServerErrorContent: "An unexpected error occured. Check your server error log for more details.",
-  contentTypeSectionAll: "All Content Types",
-  searchResults: "Search Results",
-  contentTypeSearchFieldPlaceholder: "Search for Content Types",
-  contentTypeInstallButtonLabel: "Install",
-  contentTypeInstallingButtonLabel: "Installing",
-  contentTypeUseButtonLabel: "Use",
-  contentTypeDetailButtonLabel: "Details",
-  contentTypeUpdateButtonLabel: "Update",
-  contentTypeUpdatingButtonLabel: "Updating",
-  contentTypeGetButtonLabel: "Get",
-  contentTypeBackButtonLabel: "Back",
-  contentTypeIconAltText: "Icon",
-  contentTypeInstallSuccess: ":contentType successfully installed!",
-  contentTypeUpdateSuccess: ":contentType successfully updated!",
-  contentTypeInstallError: ":contentType could not be installed. Contact your administrator.",
-  contentTypeLicensePanelTitle: "License",
-  contentTypeDemoButtonLabel: "Content Demo",
-  numResults: ":num results",
-  show: "Show",
-  recentlyUsedFirst: "Recently Used First",
-  popularFirst: "Popular First",
-  newestFirst: "Newest First",
-  aToZ: "A to Z",
-  noResultsFound: "No results found",
-  noResultsFoundDesc: "There is no content type that matches your search criteria.",
-  readMore: "Read more",
-  readLess: "Read less",
-  contentTypeOwner: "By :owner",
-  contentTypeUnsupportedApiVersionTitle: "This content type requires a newer core version",
-  contentTypeUnsupportedApiVersionContent: "Contact your system administrator to provide you with the necessary updates",
-  contentTypeUpdateAvailable: "Update available",
-  contentTypeRestricted: "Restricted content type",
-  contentTypeRestrictedDesc: "The use of this content type has been restricted by an administrator.",
-  contentTypeNotInstalled: "Content type not installed",
-  contentTypeNotInstalledDesc: "You do not have permission to install content types.",
-  theContentType: "the content type",
-  currentMenuSelected: "current selection",
-  errorCommunicatingHubTitle: "Not able to communicate with hub.",
-  errorCommunicatingHubContent: "An error occured. Please try again.",
+  enterAudioTitle: 'Paste link or other audio source URL',
+  enterVideoTitle: 'Paste YouTube link or other video source URL',
+  uploadAudioTitle: 'Upload audio file',
+  uploadVideoTitle: 'Upload video file',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  insert: 'Insert',
+  cancel: 'Cancel',
+  height: 'height',
+  width: 'width',
+  textField: 'text field',
+  numberField: 'number field',
+  orderItemUp: 'Order item up',
+  orderItemDown: 'Order item down',
+  removeItem: 'Remove item',
+  hubPanelLabel: 'Select content type',
+  importantInstructions: 'Important instructions',
+  showImportantInstructions: 'Show instructions',
+  hideImportantInstructions: 'Hide important instructions',
+  hide: 'Hide',
+  example: 'Example',
+  createContentTabLabel: 'Create Content',
+  uploadTabLabel: 'Upload',
+  uploadPlaceholder: 'No file chosen',
+  uploadInstructionsTitle: 'Upload an H5P file.',
+  uploadInstructionsContent:
+    'You may start with examples from <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadFileButtonLabel: 'Upload a file',
+  uploadFileButtonChangeLabel: 'Change file',
+  uploadingThrobber: 'Now uploading...',
+  uploadSuccess: ':title was successfully uploaded!',
+  unableToInterpretError: 'Unable to interpret response.',
+  unableToInterpretSolution: 'Please check your error log.',
+  h5pFileWrongExtensionTitle: 'The selected file could not be uploaded',
+  h5pFileWrongExtensionContent: 'Only files with the .h5p extension are allowed.',
+  h5pFileValidationFailedTitle: 'Could not validate H5P file.',
+  h5pFileValidationFailedContent:
+    'Make sure the uploaded H5P contains valid H5P content. H5P files containing only libraries should be uploaded through the H5P Libraries page.',
+  h5pFileUploadServerErrorTitle: 'The H5P file could not be uploaded',
+  h5pFileUploadServerErrorContent: 'An unexpected error occured. Check your server error log for more details.',
+  contentTypeSectionAll: 'All Content Types',
+  searchResults: 'Search Results',
+  contentTypeSearchFieldPlaceholder: 'Search for Content Types',
+  contentTypeInstallButtonLabel: 'Install',
+  contentTypeInstallingButtonLabel: 'Installing',
+  contentTypeUseButtonLabel: 'Use',
+  contentTypeDetailButtonLabel: 'Details',
+  contentTypeUpdateButtonLabel: 'Update',
+  contentTypeUpdatingButtonLabel: 'Updating',
+  contentTypeGetButtonLabel: 'Get',
+  contentTypeBackButtonLabel: 'Back',
+  contentTypeIconAltText: 'Icon',
+  contentTypeInstallSuccess: ':contentType successfully installed!',
+  contentTypeUpdateSuccess: ':contentType successfully updated!',
+  contentTypeInstallError: ':contentType could not be installed. Contact your administrator.',
+  contentTypeLicensePanelTitle: 'License',
+  contentTypeDemoButtonLabel: 'Content Demo',
+  numResults: ':num results',
+  show: 'Show',
+  recentlyUsedFirst: 'Recently Used First',
+  popularFirst: 'Popular First',
+  newestFirst: 'Newest First',
+  aToZ: 'A to Z',
+  noResultsFound: 'No results found',
+  noResultsFoundDesc: 'There is no content type that matches your search criteria.',
+  readMore: 'Read more',
+  readLess: 'Read less',
+  contentTypeOwner: 'By :owner',
+  contentTypeUnsupportedApiVersionTitle: 'This content type requires a newer core version',
+  contentTypeUnsupportedApiVersionContent:
+    'Contact your system administrator to provide you with the necessary updates',
+  contentTypeUpdateAvailable: 'Update available',
+  contentTypeRestricted: 'Restricted content type',
+  contentTypeRestrictedDesc: 'The use of this content type has been restricted by an administrator.',
+  contentTypeNotInstalled: 'Content type not installed',
+  contentTypeNotInstalledDesc: 'You do not have permission to install content types.',
+  theContentType: 'the content type',
+  currentMenuSelected: 'current selection',
+  errorCommunicatingHubTitle: 'Not able to communicate with hub.',
+  errorCommunicatingHubContent: 'An error occured. Please try again.',
   warningNoContentTypesInstalled: "You don't have any content types installed.",
-  warningChangeBrowsingToSeeResults: "Click <em>All</em> to get the list of all the content types you can install.",
-  warningUpdateAvailableTitle: "A new version of :contentType is available.",
-  warningUpdateAvailableBody: "Update to the latest version for an improved experience.",
-  licenseDescription: "Some of the features of this license are indicated below. Click the info icon above to read the original license text.",
-  licenseModalTitle: "License Details",
-  licenseModalSubtitle: "Select a license to view information about proper usage",
-  licenseUnspecified: "Unspecified",
-  licenseCanUseCommercially: "Can use commercially",
-  licenseCanModify: "Can modify",
-  licenseCanDistribute: "Can distribute",
-  licenseCanSublicense: "Can sublicense",
-  licenseCanHoldLiable: "Can hold liable",
-  licenseCannotHoldLiable: "Cannot hold liable",
-  licenseMustIncludeCopyright: "Must include copyright",
-  licenseMustIncludeLicense: "Must include license",
-  licenseFetchDetailsFailed: "Failed fetching license details",
-  imageLightboxTitle: "Images",
-  imageLightBoxProgress: ":num of :total",
-  nextImage: "Next image",
-  previousImage: "Previous image",
-  screenshots: "Screenshots",
-  reloadButtonLabel: "Reload",
-  videoQuality: "Video quality label",
-  videoQualityDescription: "This label helps the user identify the current quality of the video. E.g. 1080p, 720p, HD or Mobile",
-  videoQualityDefaultLabel: "Quality :index",
-  noContentTypesAvailable: "No content types are available",
-  noContentTypesAvailableDesc: "Your site is having difficulties connecting to H5P.org and listing the available content types.",
-  contentTypeCacheOutdated: "Content type list outdated",
-  contentTypeCacheOutdatedDesc: "Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.",
-  tryAgain: "Try again",
-  getHelp: "Get help",
-  untitled: "Untitled :libraryTitle",
-  title: "Title",
-  metadata: "Metadata",
-  addTitle: "Add title",
-  usedForSearchingReportsAndCopyrightInformation: "Used for searching, reports and copyright information",
-  metadataSharingAndLicensingInfo: "Metadata (sharing and licensing info)",
-  fillInTheFieldsBelow: "Fill in the fields below",
-  saveMetadata: "Save metadata",
-  addAuthor: "Save author",
-  confirmRemoveAuthor: "Are you sure you would like to remove this author?",
-  addNewChange: "Add new change",
-  confirmDeleteChangeLog: "Are you sure you would like to delete this change log entry?",
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
-  logThisChange: "Log this change",
-  newChangeHasBeenLogged: "New change has been logged",
-  loggedChanges: "Logged changes",
-  noChangesHaveBeenLogged: "No changes have been logged",
-  errorHeader: "An error occured",
-  errorCalculatingMaxScore: "Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.",
-  maxScoreSemanticsMissing: "Could not find the expected semantics in the content.",
-  copyButton: "Copy",
-  copiedButton: "Copied",
-  pasteButton: "Paste",
+  warningChangeBrowsingToSeeResults: 'Click <em>All</em> to get the list of all the content types you can install.',
+  warningUpdateAvailableTitle: 'A new version of :contentType is available.',
+  warningUpdateAvailableBody: 'Update to the latest version for an improved experience.',
+  licenseDescription:
+    'Some of the features of this license are indicated below. Click the info icon above to read the original license text.',
+  licenseModalTitle: 'License Details',
+  licenseModalSubtitle: 'Select a license to view information about proper usage',
+  licenseUnspecified: 'Unspecified',
+  licenseCanUseCommercially: 'Can use commercially',
+  licenseCanModify: 'Can modify',
+  licenseCanDistribute: 'Can distribute',
+  licenseCanSublicense: 'Can sublicense',
+  licenseCanHoldLiable: 'Can hold liable',
+  licenseCannotHoldLiable: 'Cannot hold liable',
+  licenseMustIncludeCopyright: 'Must include copyright',
+  licenseMustIncludeLicense: 'Must include license',
+  licenseFetchDetailsFailed: 'Failed fetching license details',
+  imageLightboxTitle: 'Images',
+  imageLightBoxProgress: ':num of :total',
+  nextImage: 'Next image',
+  previousImage: 'Previous image',
+  screenshots: 'Screenshots',
+  reloadButtonLabel: 'Reload',
+  videoQuality: 'Video quality label',
+  videoQualityDescription:
+    'This label helps the user identify the current quality of the video. E.g. 1080p, 720p, HD or Mobile',
+  videoQualityDefaultLabel: 'Quality :index',
+  noContentTypesAvailable: 'No content types are available',
+  noContentTypesAvailableDesc:
+    'Your site is having difficulties connecting to H5P.org and listing the available content types.',
+  contentTypeCacheOutdated: 'Content type list outdated',
+  contentTypeCacheOutdatedDesc:
+    'Your site is having difficulties connecting to H5P.org to check for content type updates. You may not be able to update or install new content types.',
+  tryAgain: 'Try again',
+  getHelp: 'Get help',
+  untitled: 'Untitled :libraryTitle',
+  title: 'Title',
+  metadata: 'Metadata',
+  addTitle: 'Add title',
+  usedForSearchingReportsAndCopyrightInformation: 'Used for searching, reports and copyright information',
+  metadataSharingAndLicensingInfo: 'Metadata (sharing and licensing info)',
+  fillInTheFieldsBelow: 'Fill in the fields below',
+  saveMetadata: 'Save metadata',
+  addAuthor: 'Save author',
+  confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
+  addNewChange: 'Add new change',
+  confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  logThisChange: 'Log this change',
+  newChangeHasBeenLogged: 'New change has been logged',
+  loggedChanges: 'Logged changes',
+  noChangesHaveBeenLogged: 'No changes have been logged',
+  errorHeader: 'An error occured',
+  errorCalculatingMaxScore:
+    'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
+  maxScoreSemanticsMissing: 'Could not find the expected semantics in the content.',
+  copyButton: 'Copy',
+  copiedButton: 'Copied',
+  pasteButton: 'Paste',
   pasteAndReplaceButton: 'Paste & Replace',
   pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
+  confirmPasteContent:
+    'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
   confirmPasteButtonText: 'Replace content',
-  copyToClipboard: "Copy H5P content to the clipboard",
-  copiedToClipboard: "Content is copied to the clipboard",
-  pasteFromClipboard: "Paste H5P content from the clipboard",
+  copyToClipboard: 'Copy H5P content to the clipboard',
+  copiedToClipboard: 'Content is copied to the clipboard',
+  pasteFromClipboard: 'Paste H5P content from the clipboard',
   pasteAndReplaceFromClipboard: 'Replace existing content with H5P Content from the clipboard',
-  pasteNoContent: "No H5P content on the clipboard",
-  pasteError: "Cannot paste from clipboard",
+  pasteNoContent: 'No H5P content on the clipboard',
+  pasteError: 'Cannot paste from clipboard',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
-  ok: "OK",
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  ok: 'OK',
   avTablistLabel: 'Insert using',
   tabTitleBasicFileUpload: 'File Upload',
   tabTitleInputLinkURL: 'Link/URL',
@@ -207,14 +220,86 @@ H5PEditor.language.core = {
   language: 'Language',
   noLanguagesSupported: 'No languages supported',
   changeLanguage: 'Change language to :language?',
-  thisWillPotentially: "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
+  thisWillPotentially:
+    "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
   notAllTextsChanged: 'Not all texts were changed, there is only partial coverage for :language.',
-  contributeTranslations: 'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Unfortunately, the selected content type \'%lib\' isn\'t installed on this system.',
+  contributeTranslations:
+    'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Unfortunately, the selected content type '%lib' isn't installed on this system.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
