@@ -7,7 +7,7 @@ H5PEditor.language.core = {
   notImageOrDimensionsField: '":path" er verken et bilde- eller dimensjonsfelt.',
   requiredProperty: '":property" er påkrevd og må ha en verdi.',
   onlyNumbers: '":property" kan bare innholde tall.',
-  illegalDecimalNumber: ':property can only contain numbers with max :decimals decimals.',
+  illegalDecimalNumber: ':property kan bare inneholde tall med maksimalt :decimals desimaler.',
   exceedsMax: '":property" overstiger maksverdien på :max.',
   listExceedsMax: 'Antallet elementer i listen overstiger maksantallet på :max elementer.',
   belowMin: '":property" er mindre enn minimumsverdien på :min.',
@@ -26,8 +26,9 @@ H5PEditor.language.core = {
   expandCollapse: 'Utvid/Slå sammen',
   addEntity: 'Legg til :entity',
   tooLong: 'Feltets verdi er for lang, den må være på :max tegn eller mindre.',
-  invalidFormat: 'Feltets verdi er på et ugyldig format eller bruker ulovlige tegn.',
-  confirmChangeLibrary: 'Ved å gjøre dette mister du alt arbeid gjort med nåværende innholdstype. Er du sikker på at du ønsker å bytte innholdstype?',
+  invalidFormat: 'Feltets verdi har ugyldig format eller bruker ulovlige tegn.',
+  confirmChangeLibrary:
+    'Ved å gjøre dette mister du alt arbeid gjort med nåværende innholdstype. Er du sikker på at du ønsker å bytte innholdstype?',
   commonFields: 'Innstillinger og tekster',
   commonFieldsDescription: 'Her kan du redigere innstillinger eller oversette tekster som brukes i dette innholdet.',
   uploading: 'Laster opp fil, vennligst vent...',
@@ -40,7 +41,7 @@ H5PEditor.language.core = {
   uploadError: 'Filopplasting feilet',
   fileToLarge: 'Filen du prøver å laste opp kan være for stor.',
   unknownFileUploadError: 'Ukjent filopplastingsfeil',
-  noSemantics: 'Feil, kunne ikke laste skjemaet for innholdstypen.',
+  noSemantics: 'Feil: kunne ikke laste skjemaet for innholdstypen.',
   editImage: 'Rediger bilde',
   saveLabel: 'Lagre',
   cancelLabel: 'Avbryt',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Lim inn YouTube lenke eller annen videokilde URL',
   uploadAudioTitle: 'Last opp lydfil',
   uploadVideoTitle: 'Last opp videofil',
-  addVideoDescription: 'H5P støtter alle eksterne videokilder på formatene mp4, webm eller ogv, slik som Vimeo Pro, og har støtte for YouTube-lenker og Panopto-lenker.',
+  addVideoDescription:
+    'H5P støtter alle eksterne videokilder på formatene mp4, webm eller ogv, slik som Vimeo Pro, og har støtte for YouTube-lenker og Panopto-lenker.',
   insert: 'Sett inn',
   cancel: 'Avbryt',
   height: 'Høyde',
@@ -74,26 +76,28 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Last opp',
   uploadPlaceholder: 'Ingen fil er valgt',
   uploadInstructionsTitle: 'Last opp en .h5p fil.',
-  uploadInstructionsContent: 'Du kan starte med eksempler fra <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Du kan starte med eksempler fra <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Last opp en fil',
   uploadFileButtonChangeLabel: 'Bytt fil',
   uploadingThrobber: 'Laster opp...',
   uploadSuccess: ':title ble lastet opp!',
   unableToInterpretError: 'Kan ikke tolke respons.',
   unableToInterpretSolution: 'Vennligst sjekk feilloggen din.',
-  h5pFileWrongExtensionTitle: '.h5p fil ble ikke funnet.',
+  h5pFileWrongExtensionTitle: '.h5p-fil ble ikke funnet.',
   h5pFileWrongExtensionContent: 'Du må laste opp en fil med fil-endelsen .h5p',
-  h5pFileValidationFailedTitle: 'Kunne ikke validere H5P filen.',
-  h5pFileValidationFailedContent: 'Pass på at .h5p filen du laster opp inneholder valid H5P innhold. H5P filer som bare inneholder biblioteker skal lastes opp på bibliotekssiden og kan ikke lastes opp som innhold.',
-  h5pFileUploadServerErrorTitle: 'H5P filen kunne ikke bli lastet opp.',
-  h5pFileUploadServerErrorContent: 'En uventet feil har oppstått. Sjekk din server log for detaljer.',
+  h5pFileValidationFailedTitle: 'Kunne ikke validere H5P-filen.',
+  h5pFileValidationFailedContent:
+    'Pass på at .h5p filen du laster opp inneholder valid H5P innhold. H5P filer som bare inneholder biblioteker skal lastes opp på bibliotekssiden og kan ikke lastes opp som innhold.',
+  h5pFileUploadServerErrorTitle: 'H5P-filen kunne ikke bli lastet opp.',
+  h5pFileUploadServerErrorContent: 'En uventet feil har oppstått. Sjekk din serverlogg for detaljer.',
   contentTypeSectionAll: 'Alle innholdstyper',
   searchResults: 'Søkeresultater',
   contentTypeSearchFieldPlaceholder: 'Søk etter innholdstyper',
   contentTypeInstallButtonLabel: 'Installer',
   contentTypeInstallingButtonLabel: 'Installerer',
   contentTypeUseButtonLabel: 'Bruk',
-  contentTypeDetailButtonLabel: 'Details',
+  contentTypeDetailButtonLabel: 'Detaljer',
   contentTypeUpdateButtonLabel: 'Oppdater',
   contentTypeUpdatingButtonLabel: 'Oppdaterer',
   contentTypeGetButtonLabel: 'Hent',
@@ -101,7 +105,7 @@ H5PEditor.language.core = {
   contentTypeIconAltText: 'Ikon',
   contentTypeInstallSuccess: ':contentType ble installert!',
   contentTypeUpdateSuccess: ':contentType ble oppdatert!',
-  contentTypeInstallError: ':contentType kunne ikke bli installert. Kontakt din administrator.',
+  contentTypeInstallError: ':contentType kunne ikke bli installert. Kontakt administrator for siden.',
   contentTypeLicensePanelTitle: 'Lisens',
   contentTypeDemoButtonLabel: 'Innholdsdemonstrasjon',
   numResults: ':num treff',
@@ -109,13 +113,13 @@ H5PEditor.language.core = {
   recentlyUsedFirst: 'Nylig brukt først',
   popularFirst: 'Populære først',
   newestFirst: 'Nyeste først',
-  aToZ: 'A til Z',
+  aToZ: 'A til Å',
   noResultsFound: 'Ingen resultater',
   noResultsFoundDesc: 'Det finnes ingen innholdstype som samsvarer med søkekriteriene dine.',
   readMore: 'Vis mer',
   readLess: 'Vis mindre',
   contentTypeOwner: 'Av :owner',
-  contentTypeUnsupportedApiVersionTitle: 'Denne innholdstypen krever en ny versjon av H5P Kjernen',
+  contentTypeUnsupportedApiVersionTitle: 'Denne innholdstypen krever en ny versjon av H5P Core',
   contentTypeUnsupportedApiVersionContent: 'Kontakt din systemadministrator for nødvendige oppdateringer.',
   contentTypeUpdateAvailable: 'Oppdatering tilgjengelig',
   contentTypeRestricted: 'Begrenset tilgang til innholdstype',
@@ -130,7 +134,8 @@ H5PEditor.language.core = {
   warningChangeBrowsingToSeeResults: 'Klikk <em>Alle</em> for å få en liste over alle innholdstyper du kan installere.',
   warningUpdateAvailableTitle: 'En ny versjon av :contentType er tilgjengelig.',
   warningUpdateAvailableBody: 'Oppdater til siste versjon for en best mulig opplevelse.',
-  licenseDescription: 'Noen av egenskapene til denne lisensen indikeres under. Klikk på informasjonsikonet for å lese lisensteksten.',
+  licenseDescription:
+    'Noen av egenskapene til denne lisensen indikeres under. Klikk på informasjonsikonet for å lese lisensteksten.',
   licenseModalTitle: 'License Details',
   licenseModalSubtitle: 'Velg en lisens for å se en oversikt over riktig bruk',
   licenseUnspecified: 'Uspesifisert',
@@ -142,7 +147,7 @@ H5PEditor.language.core = {
   licenseCannotHoldLiable: 'Kan ikke holde ansvarlig',
   licenseMustIncludeCopyright: 'Må inkludere opphavsrettsinformasjon',
   licenseMustIncludeLicense: 'Må inkludere lisens',
-  licenseFetchDetailsFailed: 'Feil ved henting av lisens detaljer',
+  licenseFetchDetailsFailed: 'Feil ved henting av lisens-detaljer',
   imageLightboxTitle: 'Bilder',
   imageLightBoxProgress: ':num av :total',
   nextImage: 'Neste bilde',
@@ -150,56 +155,64 @@ H5PEditor.language.core = {
   screenshots: 'Skjermbilder',
   reloadButtonLabel: 'Last på nytt',
   videoQuality: 'Navn på videokvalitet',
-  videoQualityDescription: 'Dette navnet hjelper brukeren med å identifisere den aktuelle videokvaliteten. F.eks. 1080p, 720p, HD or Mobile',
+  videoQualityDescription:
+    'Dette navnet hjelper brukeren med å identifisere den aktuelle videokvaliteten. F.eks. 1080p, 720p, HD eller for mobil',
   videoQualityDefaultLabel: 'Kvalitet :index',
   noContentTypesAvailable: 'Ingen innholdstyper er tilgjengelig',
-  noContentTypesAvailableDesc: 'Nettstedet ditt har problemer med å koble til H5P.org og liste tilgjengelige innholdstyper.',
+  noContentTypesAvailableDesc:
+    'Nettstedet ditt har problemer med å koble til H5P.org og kan ikke liste tilgjengelige innholdstyper.',
   contentTypeCacheOutdated: 'Innholdstype-listen er utdatert',
-  contentTypeCacheOutdatedDesc: 'Nettstedet ditt har problemer med å koble til H5P.org for å se etter innholdstype-oppdateringer. Det kan hende at du ikke får oppdatert eller installert nye innholdstyper.',
+  contentTypeCacheOutdatedDesc:
+    'Nettstedet ditt har problemer med å koble til H5P.org for å se etter innholdstype-oppdateringer. Det kan hende at du ikke får oppdatert eller installert nye innholdstyper.',
   tryAgain: 'Prøv på nytt',
-  getHelp: 'Get help',
-  untitled: 'Untitled :libraryTitle',
-  title: 'Title',
+  getHelp: 'Få hjelp',
+  untitled: 'Uten tittel :libraryTitle',
+  title: 'Tittel',
   metadata: 'Metadata',
-  addTitle: 'Add title',
-  usedForSearchingReportsAndCopyrightInformation: 'Used for searching, reports and copyright information',
-  metadataSharingAndLicensingInfo: 'Metadata (sharing and licensing info)',
-  fillInTheFieldsBelow: 'Fill in the fields below',
-  saveMetadata: 'Save metadata',
-  addAuthor: 'Save author',
-  confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
-  addNewChange: 'Add new change',
-  confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
-  logThisChange: 'Log this change',
-  newChangeHasBeenLogged: 'New change has been logged',
-  loggedChanges: 'Logged changes',
-  noChangesHaveBeenLogged: 'No changes have been logged',
+  addTitle: 'Legg til tittel',
+  usedForSearchingReportsAndCopyrightInformation: 'Brukes til søk, rapporter og copyright-informasjon',
+  metadataSharingAndLicensingInfo: 'Metadata (deling- og lisensinformasjon)',
+  fillInTheFieldsBelow: 'Fyll inn feltene nedenfor',
+  saveMetadata: 'Lagre metadata',
+  addAuthor: 'Lagre forfatter',
+  confirmRemoveAuthor: 'Er du sikker på at du ønsker å fjerne denne forfatteren?',
+  addNewChange: 'Legg til endring',
+  confirmDeleteChangeLog: 'Er du sikker på at du ønsker å slette denne loggoppføringen?',
+  changelogDescription:
+    'Enkelte lisenser krever at endringer gjort på det originale verket, eller verk avledet av dette, loggføres og vises. Du kan loggføre endringene dine her for lisenshensyn, slik at du og andre skal kunne holde styr på endringene som er gjort på dette innholdet.',
+  logThisChange: 'Loggfør denne endringen',
+  newChangeHasBeenLogged: 'Ny endring er blitt loggført',
+  loggedChanges: 'Loggførte endringer',
+  noChangesHaveBeenLogged: 'Ingen endring er blitt loggført',
   errorHeader: 'En feil har oppstått',
-  errorCalculatingMaxScore: 'Kunne ikke kalkulerer maks poengsum for denne innholdstypen. Maks poengsum er antatt til å være 0. Kontakt administrator dersom dette ikke er korrekt.',
+  errorCalculatingMaxScore:
+    'Kunne ikke kalkulerer maks poengsum for denne innholdstypen. Maks poengsum er antatt til å være 0. Kontakt administrator dersom dette ikke er korrekt.',
   maxScoreSemanticsMissing: 'Kunne ikke finne den forventede semantikken i den valgte innholdstypen.',
   copyButton: 'Kopier',
   copiedButton: 'Kopiert',
   pasteButton: 'Lim inn',
-  pasteAndReplaceButton: 'Paste & Replace',
-  pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
-  confirmPasteButtonText: 'Replace content',
-  copyToClipboard: 'Copy H5P content to the clipboard',
-  copiedToClipboard: 'Content is copied to the clipboard',
-  pasteFromClipboard: 'Paste H5P content from the clipboard',
-  pasteAndReplaceFromClipboard: 'Replace existing content with H5P Content from the clipboard',
-  pasteNoContent: 'No H5P content on the clipboard',
-  pasteError: 'Cannot paste from clipboard',
-  pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
-  pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteAndReplaceButton: 'Lim inn og erstatt',
+  pasteContent: 'Erstatt innhold',
+  confirmPasteContent:
+    'Dette vil erstatte det eksisterende innholdet med innholdet på din utklippstavle. Det eksisterende innholdet vil gå tapt. Er du sikker på at du ønsker å fortsette? ',
+  confirmPasteButtonText: 'Erstatt innhold',
+  copyToClipboard: 'Kopier H5P-innhold til utklippstavlen',
+  copiedToClipboard: 'Innholdet er kopiert til utklippstavlen',
+  pasteFromClipboard: 'Lim inn H5P-innhold fra utklippstavlen',
+  pasteAndReplaceFromClipboard: 'Erstatt eksisterende innhold med H5P-innhold fra utklippstavlen',
+  pasteNoContent: 'Utklippstavlen har ikke H5P-innhold',
+  pasteError: 'Kan ikke lime inn fra utklippstavlen',
+  pasteContentNotSupported: 'Innholdet på utklippstavlen støttes ikke i denne sammenhengen',
+  pasteContentRestricted: 'Innholdet på utklippstavlen er begrenset på denne siden.',
+  pasteTooOld:
+    'Innholdet på utklippstavlen har lavere versjon (:clip) enn det som er støttet i denne sammenhengen (:local). Om det er mulig, kan du oppdatere innholdet du ønsker å kopiere, for så å forsøke og kopiere og lime det inn på nytt.',
+  pasteTooNew:
+    'Innholdet på utklippstavlen har høyere versjon (:clip) enn det som er støttet i denne sammenhengen (:local). Om det er mulig, kan du oppdatere innholdet du ønsker å kopiere til, for så å forsøke og kopiere og lime det inn på nytt.',
   ok: 'OK',
   avTablistLabel: 'Sett inn via',
   tabTitleBasicFileUpload: 'Filopplasting',
   tabTitleInputLinkURL: 'Lenke/URL',
-  errorTooHighVersion: 'Parameterene innholder %used mens kunn %supported eller tidligere er støttet.',
+  errorTooHighVersion: 'Parameterene innholder %used mens kun %supported eller tidligere er støttet.',
   errorNotSupported: 'Parameterene inneholder %used som ikke er støttet.',
   errorParamsBroken: 'Parameterene er ødelagt.',
   libraryMissing: 'Mangler påkrevd bibliotek %lib.',
@@ -207,87 +220,86 @@ H5PEditor.language.core = {
   language: 'Språk',
   noLanguagesSupported: 'Ingen språk støttet',
   changeLanguage: 'Bytt språk til :language?',
-  thisWillPotentially: "Dette kan potensielt tilbakestille all teksten og oversettelsene. Du kan ikke omgjøre dette. Innholdet vil ikke bli endret. Vil du fortsette?",
+  thisWillPotentially:
+    'Dette kan potensielt tilbakestille all teksten og oversettelsene. Du kan ikke omgjøre dette. Innholdet vil ikke bli endret. Vil du fortsette?',
   notAllTextsChanged: 'Ikke alle tekstene ble endret, det er kun delvis støtte for :language.',
-  contributeTranslations: 'Hvis du vil fullføre oversettelsen for :language kan du lære mer om <a href=":url" target="_new">bidra med oversettelser til H5P</a>',
-  unknownLibrary: 'Dessverre var ikke den valgte \'%lib\'-innholdstypen installert på dette systemet.',
-  proceedButtonLabel: 'Proceed to save',
-  enterFullscreenButtonLabel: 'Enter fullscreen',
-  exitFullscreenButtonLabel: 'Exit fullscreen',
-  noContentHeader: 'No Fitting Content?',
-  noContentSuggestion: 'Create one yourself!',
-  tutorials: 'Tutorials',
-  contentSectionAll: 'All shared content',
-  popularContent: 'Popular Content',
-  allPopular: 'All Popular',
-  newOnTheHub: 'New on the Hub',
-  allNew: 'All New',
-  filterBy: 'Filter by',
-  filter: 'Filter',
+  contributeTranslations:
+    'Hvis du vil fullføre oversettelsen for :language kan du lære mer om <a href=":url" target="_new">bidra med oversettelser til H5P</a>',
+  unknownLibrary: "Dessverre var ikke den valgte '%lib'-innholdstypen installert på dette systemet.",
+  proceedButtonLabel: 'Fortsett for å lagre',
+  enterFullscreenButtonLabel: 'Åpne i fullskjerm',
+  exitFullscreenButtonLabel: 'Avslutt fullskjerm',
+  noContentHeader: 'Ikke noe innhold som passer?',
+  noContentSuggestion: 'Lag innhold selv!',
+  tutorials: 'Veiledninger',
+  contentSectionAll: 'Alt delt innhold',
+  popularContent: 'Populært innhold',
+  allPopular: 'Alt populært innhold',
+  newOnTheHub: 'Nytt på H5P Hub',
+  allNew: 'Alt nytt',
+  filterBy: 'Filtrer etter',
+  filter: 'Filtrer',
   filters: {
     level: {
-      dropdownLabel: 'Level',
-      dialogHeader: 'Select level of education',
-      dialogButtonLabel: 'Filter level of education'
+      dropdownLabel: 'Nivå',
+      dialogHeader: 'Velg utdanningsnivå',
+      dialogButtonLabel: 'Filtrer etter utdanningsnivå',
     },
     language: {
-      dropdownLabel: 'Language',
-      dialogHeader: 'Select language(s)',
-      dialogButtonLabel: 'Filter languages',
-      searchPlaceholder: 'Select one or more languages'
+      dropdownLabel: 'Språk',
+      dialogHeader: 'Velg språk',
+      dialogButtonLabel: 'Filtrer etter språk',
+      searchPlaceholder: 'Velg ett eller flere språk',
     },
     reviewed: {
-      dropdownLabel: 'Reviewed',
-      dialogHeader: 'Reviewed Content',
-      dialogButtonLabel: 'Filter',
-      optionLabel: 'Show only reviewed content'
+      dropdownLabel: 'Gjennomgått',
+      dialogHeader: 'Gjennomgått innhold',
+      dialogButtonLabel: 'Filtrer',
+      optionLabel: 'Vis bare gjennomgått innhold',
     },
     contentTypes: {
-      dropdownLabel: 'Content types',
-      dialogHeader: 'Select Content type(s)',
-      dialogButtonLabel: 'Filter Content Types',
-      searchPlaceholder: 'Select one or more content types'
+      dropdownLabel: 'Innholdstyper',
+      dialogHeader: 'Velg innholdstype(r)',
+      dialogButtonLabel: 'Filtrer på innholdstyper',
+      searchPlaceholder: 'Velg en eller flere innholdstyper',
     },
     disciplines: {
-      dropdownLabel: 'Discipline',
-      dialogHeader: 'Select Your Discipline',
-      dialogButtonLabel: 'Filter Disciplines',
-      searchPlaceholder: 'Select one or more disciplines'
+      dropdownLabel: 'Disiplin',
+      dialogHeader: 'Velg din disiplin',
+      dialogButtonLabel: 'Filtrer disipliner',
+      searchPlaceholder: 'Velg en eller flere disipliner',
     },
     licenses: {
-      dropdownLabel: 'License',
-      dialogHeader: 'Select preferred rights of use',
-      dialogButtonLabel: 'Filter licenses',
-      options: {
-        modified: 'Can be modified',
-        commercial: 'Allows commercial use'
-      }
-    }
+      dropdownLabel: 'Lisens',
+      dialogHeader: 'Velg rettigheter for bruk',
+      dialogButtonLabel: 'Filtrer på lisenser',
+      options: { modified: 'Kan endres', commercial: 'Tillater kommersiell bruk' },
+    },
   },
-  clearFilters: 'Clear all filters',
-  contentSearchFieldPlaceholder: 'Search for Content',
-  loadingContentTitle: 'We are loading content for you...',
-  loadingContentSubtitle: 'Please wait',
-  by: 'By',
-  dropdownButton: 'Open dropdown',
-  paginationNavigation: 'Pagination navigation',
-  page: 'Page',
-  currentPage: 'Current page',
-  nextPage: 'Go to next page',
-  previousPage: 'Go to previous page',
-  contentPreviewButtonLabel: 'Preview',
-  contentDownloadButtonLabel: 'Get Content',
-  reuseContentTabLabel: 'Get Shared Content',
-  contentPublisherPanelHeader: 'Publisher Info',
-  noContentFoundDesc: 'There is no content that matches your search criteria.',
-  h5pType: 'H5P Type',
-  level: 'Level',
-  size: 'Size',
-  failedFetchingData: 'Failed fetching data',
-  filterErrorMessage: 'Something went wrong. Please reload the page.',
-  in: 'in',
-  navigateToParent: 'Navigate to parent',
-  a11yTitleShowLabel: 'Show label for AT',
-  a11yTitleHideLabel: 'Hide label for AT',
-  reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  clearFilters: 'Nullstill alle filtre',
+  contentSearchFieldPlaceholder: 'Søk etter innhold',
+  loadingContentTitle: 'Vi laster innhold for deg...',
+  loadingContentSubtitle: 'Vennligst vent',
+  by: 'Av',
+  dropdownButton: 'Åpne nedtrekksmeny',
+  paginationNavigation: 'Sidenavigasjon',
+  page: 'Side',
+  currentPage: 'Gjeldende side',
+  nextPage: 'Gå til neste side',
+  previousPage: 'Gå til forrige side',
+  contentPreviewButtonLabel: 'Forhåndsvis',
+  contentDownloadButtonLabel: 'Hent innhold',
+  reuseContentTabLabel: 'Finn delt innhold',
+  contentPublisherPanelHeader: 'Informasjon om publiserer',
+  noContentFoundDesc: 'Det finnes ikke noe innhold som samsvarer med ditt søk.',
+  h5pType: 'H5P-type',
+  level: 'Nivå',
+  size: 'Størrelse',
+  failedFetchingData: 'Henting av data mislyktes',
+  filterErrorMessage: 'Noe gikk galt. Vennligst last siden på nytt.',
+  in: 'i',
+  navigateToParent: 'Naviger til overordnet nivå',
+  a11yTitleShowLabel: 'Vis etikett for tilgjengelighet',
+  a11yTitleHideLabel: 'Skjul etikett for tilgjengelighet',
+  reuseSuccess: ':title ble vellykket importert fra H5P Hub.',
 };
