@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Voeg :entity toe',
   tooLong: 'Veldwaarde is te lang, mag :max letters of minder bevatten.',
   invalidFormat: 'Veldwaarde bevat een ongeldig format of karakters die verboden zijn.',
-  confirmChangeLibrary: 'Hierdoor verlies je al het werk dat je met het huidige content type hebt gedaan. Weet je zeker dat je het content type wilt wijzigen?',
+  confirmChangeLibrary:
+    'Hierdoor verlies je al het werk dat je met het huidige content type hebt gedaan. Weet je zeker dat je het content type wilt wijzigen?',
   commonFields: 'Tekst overschrijven en vertalingen',
   commonFieldsDescription: 'Hier kun je instellingen bewerken of in dit content type gebruikte teksten vertalen.',
   uploading: 'Uploaden, wachten alstublieft ...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Plak YouTube-link of andere URL van videobron',
   uploadAudioTitle: 'Upload audiobestand',
   uploadVideoTitle: 'Upload videobestand',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Invoegen',
   cancel: 'Annuleren',
   height: 'hoogte',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Uploaden',
   uploadPlaceholder: 'Geen bestand gekozen',
   uploadInstructionsTitle: 'Een H5P bestand uploaden.',
-  uploadInstructionsContent: 'Je kunt beginnen met voorbeelden van <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Je kunt beginnen met voorbeelden van <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Upload een bestand',
   uploadFileButtonChangeLabel: 'Wijzig bestand',
   uploadingThrobber: 'Uploaden...',
@@ -84,9 +87,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Het geselecteerde bestand kon niet worden geüpload',
   h5pFileWrongExtensionContent: 'Alleen bestanden met de .h5p extensie toegestaan.',
   h5pFileValidationFailedTitle: 'Kon het H5P bestand niet valideren.',
-  h5pFileValidationFailedContent: 'Zorg ervoor dat de geüploade H5P geldige H5P content bevat. H5P bestanden die alleen bibliotheken bevatten, dienen te worden geüpload via de H5P Bibliotheken pagina.',
+  h5pFileValidationFailedContent:
+    'Zorg ervoor dat de geüploade H5P geldige H5P content bevat. H5P bestanden die alleen bibliotheken bevatten, dienen te worden geüpload via de H5P Bibliotheken pagina.',
   h5pFileUploadServerErrorTitle: 'Het H5P bestand kon niet worden geüpload',
-  h5pFileUploadServerErrorContent: 'Er is een onverwachte fout opgetreden. Controleer het server log voor meer informatie.',
+  h5pFileUploadServerErrorContent:
+    'Er is een onverwachte fout opgetreden. Controleer het server log voor meer informatie.',
   contentTypeSectionAll: 'Alle Content Types',
   searchResults: 'Zoekresultaten',
   contentTypeSearchFieldPlaceholder: 'Zoeken naar Content Types',
@@ -127,10 +132,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Kan niet communiceren met hub.',
   errorCommunicatingHubContent: 'Er is een fout opgetreden. Probeer het opnieuw.',
   warningNoContentTypesInstalled: 'Je hebt geen geïnstalleerde content types.',
-  warningChangeBrowsingToSeeResults: 'Klik <em>Alle</em> om de lijst op te halen van alle content types die je kunt installeren.',
+  warningChangeBrowsingToSeeResults:
+    'Klik <em>Alle</em> om de lijst op te halen van alle content types die je kunt installeren.',
   warningUpdateAvailableTitle: 'Er is een nieuwe versie van :contentType beschikbaar.',
   warningUpdateAvailableBody: 'Update naar de laatste versie voor een verbeterde ervaring.',
-  licenseDescription: 'Sommige van de kenmerken van deze licentie zijn hieronder weergegeven. Klik op het info icoon hierboven om de oorspronkelijke tekst van de licentie te lezen.',
+  licenseDescription:
+    'Sommige van de kenmerken van deze licentie zijn hieronder weergegeven. Klik op het info icoon hierboven om de oorspronkelijke tekst van de licentie te lezen.',
   licenseModalTitle: 'Details van licentie',
   licenseModalSubtitle: 'Selecteer een licentie om informatie te krijgen over het correcte gebruik ervan',
   licenseUnspecified: 'Niet gespecificeerd',
@@ -150,12 +157,15 @@ H5PEditor.language.core = {
   screenshots: 'Screenshots',
   reloadButtonLabel: 'Opnieuw laden',
   videoQuality: 'Videokwaliteitslabel',
-  videoQualityDescription: 'Dit label helpt de gebruiker bij het vaststelllen van de huidige kwaliteit van de video. Bijv. 1080p, 720p, HD of Mobile',
+  videoQualityDescription:
+    'Dit label helpt de gebruiker bij het vaststelllen van de huidige kwaliteit van de video. Bijv. 1080p, 720p, HD of Mobile',
   videoQualityDefaultLabel: 'Kwaliteit :index',
   noContentTypesAvailable: 'Geen content types beschikbaar',
-  noContentTypesAvailableDesc: 'Je site heeft problemen bij het verbinden met H5P.org en het tonen van de beschikbare content types.',
+  noContentTypesAvailableDesc:
+    'Je site heeft problemen bij het verbinden met H5P.org en het tonen van de beschikbare content types.',
   contentTypeCacheOutdated: 'Content types lijst is verlopen',
-  contentTypeCacheOutdatedDesc: 'Je site heeft problemen bij het verbinden met H5P.org om te controleren op updates voor content types. Het kan zijn dat je geen content types kunt updaten of installeren.',
+  contentTypeCacheOutdatedDesc:
+    'Je site heeft problemen bij het verbinden met H5P.org om te controleren op updates voor content types. Het kan zijn dat je geen content types kunt updaten of installeren.',
   tryAgain: 'Probeer nogmaals',
   getHelp: 'Vraag hulp',
   untitled: 'Zonder titel :libraryTitle',
@@ -170,20 +180,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Weet je zeker dat je deze auteur wilt verwijderen?',
   addNewChange: 'Voeg nieuwe wijziging toe',
   confirmDeleteChangeLog: 'Weet je zeker dat je deze invoer van het wijzigingslog wilt verwijderen?',
-  changelogDescription: 'Sommige licenties vereisen dat wijzigingen aan het originele werk, of afgeleiden daarvan, worden gelogd en getoond. U kunt hier uw wijzigingen vanwege licentieredenen loggen of alleen om uzelf en anderen te informeren over wijzigingen aan deze inhoud.',
+  changelogDescription:
+    'Sommige licenties vereisen dat wijzigingen aan het originele werk, of afgeleiden daarvan, worden gelogd en getoond. U kunt hier uw wijzigingen vanwege licentieredenen loggen of alleen om uzelf en anderen te informeren over wijzigingen aan deze inhoud.',
   logThisChange: 'Log deze wijziging',
   newChangeHasBeenLogged: 'Nieuwe wijziging is gelogd',
   loggedChanges: 'Gelogde wijzigingen',
   noChangesHaveBeenLogged: 'Er zijn geen wijzigingen gelogd',
   errorHeader: 'Er is een fout opgetreden',
-  errorCalculatingMaxScore: 'Kon de maximumscore voor deze inhoud niet berekenen. De maximumscore lijkt 0 te zijn. Neem contact op met uw beheerder als dit niet juist is.',
+  errorCalculatingMaxScore:
+    'Kon de maximumscore voor deze inhoud niet berekenen. De maximumscore lijkt 0 te zijn. Neem contact op met uw beheerder als dit niet juist is.',
   maxScoreSemanticsMissing: 'Kon de verwachte semantics in de content niet vinden.',
   copyButton: 'Kopiëren',
   copiedButton: 'Gekopieerd',
   pasteButton: 'Plakken',
   pasteAndReplaceButton: 'Plak en vervang',
   pasteContent: 'Plak inhoud',
-  confirmPasteContent: 'Hiermee vervang je de huidige inhoud door de inhoud van het klembord. De huidige inhoud gaat verloren. Weet je zeker dat je wilt doorgaan?',
+  confirmPasteContent:
+    'Hiermee vervang je de huidige inhoud door de inhoud van het klembord. De huidige inhoud gaat verloren. Weet je zeker dat je wilt doorgaan?',
   confirmPasteButtonText: 'Vervang inhoud',
   copyToClipboard: 'Kopieer H5P inhoud naar het klembord',
   copiedToClipboard: 'Inhoud is gekopieerd naar het klembord',
@@ -193,8 +206,10 @@ H5PEditor.language.core = {
   pasteError: 'Kan niet plakkken van klembord',
   pasteContentNotSupported: 'De inhoud in het H5P klembord wordt niet ondersteund in deze context',
   pasteContentRestricted: 'De inhoud in het H5P klembord bevat afgeschermde content',
-  pasteTooOld: 'De inhoud in het H5P klembord is van een lagere versie (:clip) dan wat wordt ondersteund in deze context (:local), probeer indien mogelijk de inhoud die u wilt plakken te upgraden, kopieer het opnieuw en plak het hier opnieuw.',
-  pasteTooNew: 'De inhoud in het H5P klembord is van een hogere versie (:clip) dan wat wordt ondersteund in deze context (:local), probeer indien mogelijk deze content eerst te upgraden, en probeer dan opnieuw de inhoud hier te plakken.',
+  pasteTooOld:
+    'De inhoud in het H5P klembord is van een lagere versie (:clip) dan wat wordt ondersteund in deze context (:local), probeer indien mogelijk de inhoud die u wilt plakken te upgraden, kopieer het opnieuw en plak het hier opnieuw.',
+  pasteTooNew:
+    'De inhoud in het H5P klembord is van een hogere versie (:clip) dan wat wordt ondersteund in deze context (:local), probeer indien mogelijk deze content eerst te upgraden, en probeer dan opnieuw de inhoud hier te plakken.',
   ok: 'OK',
   avTablistLabel: 'Invoegen m.b.v.',
   tabTitleBasicFileUpload: 'Bestandsupload',
@@ -207,14 +222,86 @@ H5PEditor.language.core = {
   language: 'Taal',
   noLanguagesSupported: 'Geen taal ondersteuning',
   changeLanguage: 'Wijzig taal naar :language?',
-  thisWillPotentially: "Mogelijk stelt dit alle tekst en vertalingen opnieuw in. Je kunt dit niet ongedaan maken. De content zelf wordt niet gewijzigd. Wil je doorgaan?",
+  thisWillPotentially:
+    'Mogelijk stelt dit alle tekst en vertalingen opnieuw in. Je kunt dit niet ongedaan maken. De content zelf wordt niet gewijzigd. Wil je doorgaan?',
   notAllTextsChanged: 'Niet alle teksten zijn gewijzigd, er is slechts gedeeltelijke dekking voor :language.',
-  contributeTranslations: 'Als je de vertaling voor :language wilt completeren, kun je hier meer leren over <a href=":url" target="_new">het bijdragen van vertalingen aan H5P</a>',
-  unknownLibrary: 'Jammer genoeg is het geselecteerde content type \'%lib\' niet geïnstalleerd op dit systeem.',
+  contributeTranslations:
+    'Als je de vertaling voor :language wilt completeren, kun je hier meer leren over <a href=":url" target="_new">het bijdragen van vertalingen aan H5P</a>',
+  unknownLibrary: "Jammer genoeg is het geselecteerde content type '%lib' niet geïnstalleerd op dit systeem.",
   proceedButtonLabel: 'Doorgaan naar opslaan',
   enterFullscreenButtonLabel: 'In volledig scherm werken',
   exitFullscreenButtonLabel: 'Volledig scherm afsluiten',
   a11yTitleShowLabel: 'Toon label voor AT',
   a11yTitleHideLabel: 'Verberg label voor AT',
   reuseSuccess: ':title is met succes geïmporteerd van de H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
