@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Legg til :entity',
   tooLong: 'Feltets verdi er for lang, den må vere på :max tegn eller mindre.',
   invalidFormat: 'Feltets verdi er på eit ugyldig format eller bruker ulovlege tegn.',
-  confirmChangeLibrary: 'Gjør du dette mister du alt arbeid gjort med nåværande innhaldstype. Er du sikker på at du ønskjer å byte innhaldstype?',
+  confirmChangeLibrary:
+    'Gjør du dette mister du alt arbeid gjort med nåværande innhaldstype. Er du sikker på at du ønskjer å byte innhaldstype?',
   commonFields: 'Innstillinger og tekstar',
   commonFieldsDescription: 'Her kan du redigere innstillinger eller oversette tekstar som brukast i dette innhaldet.',
   uploading: 'Lastar opp fil, ver venleg og vent...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Lim inn YouTube lenke eller annen videokjelde URL',
   uploadAudioTitle: 'Last opp lydfil',
   uploadVideoTitle: 'Last opp videofil',
-  addVideoDescription: 'H5P støtter alle eksterne videokilder på formatene mp4, webm eller ogv, slik som Vimeo Pro, og har støtte for YouTube-lenker og Panopto-lenker.',
+  addVideoDescription:
+    'H5P støtter alle eksterne videokilder på formatene mp4, webm eller ogv, slik som Vimeo Pro, og har støtte for YouTube-lenker og Panopto-lenker.',
   insert: 'Sett inn',
   cancel: 'Avbryt',
   height: 'Høyde',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Upload',
   uploadPlaceholder: 'No file chosen',
   uploadInstructionsTitle: 'Upload an H5P file.',
-  uploadInstructionsContent: 'You may start with examples from <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'You may start with examples from <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Upload a file',
   uploadFileButtonChangeLabel: 'Change file',
   uploadingThrobber: 'Now uploading...',
@@ -84,7 +87,8 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'The selected file could not be uploaded',
   h5pFileWrongExtensionContent: 'Only files with the .h5p extension are allowed.',
   h5pFileValidationFailedTitle: 'Could not validate H5P file.',
-  h5pFileValidationFailedContent: 'Make sure the uploaded H5P contains valid H5P content. H5P files containing only libraries should be uploaded through the H5P Libraries page.',
+  h5pFileValidationFailedContent:
+    'Make sure the uploaded H5P contains valid H5P content. H5P files containing only libraries should be uploaded through the H5P Libraries page.',
   h5pFileUploadServerErrorTitle: 'The H5P file could not be uploaded',
   h5pFileUploadServerErrorContent: 'An unexpected error occured. Check your server error log for more details.',
   contentTypeSectionAll: 'Alle innhaldstypar',
@@ -116,7 +120,8 @@ H5PEditor.language.core = {
   readLess: 'Read less',
   contentTypeOwner: 'By :owner',
   contentTypeUnsupportedApiVersionTitle: 'This content type requires a newer core version',
-  contentTypeUnsupportedApiVersionContent: 'Contact your system administrator to provide you with the necessary updates',
+  contentTypeUnsupportedApiVersionContent:
+    'Contact your system administrator to provide you with the necessary updates',
   contentTypeUpdateAvailable: 'Update available',
   contentTypeRestricted: 'Begrenset tilgong til innhaldstype',
   contentTypeRestrictedDesc: 'Bruken av denne innhaldstypen er begrenset av ein administrator.',
@@ -126,11 +131,12 @@ H5PEditor.language.core = {
   currentMenuSelected: 'current selection',
   errorCommunicatingHubTitle: 'Not able to communicate with hub.',
   errorCommunicatingHubContent: 'An error occured. Please try again.',
-  warningNoContentTypesInstalled: 'You don\'t have any content types installed.',
+  warningNoContentTypesInstalled: "You don't have any content types installed.",
   warningChangeBrowsingToSeeResults: 'Click <em>All</em> to get the list of all the content types you can install.',
   warningUpdateAvailableTitle: 'A new version of :contentType is available.',
   warningUpdateAvailableBody: 'Update to the latest version for an improved experience.',
-  licenseDescription: 'Some of the features of this license are indicated below. Click the info icon above to read the original license text.',
+  licenseDescription:
+    'Some of the features of this license are indicated below. Click the info icon above to read the original license text.',
   licenseModalTitle: 'License Details',
   licenseModalSubtitle: 'Select a license to view information about proper usage',
   licenseUnspecified: 'Unspecified',
@@ -150,12 +156,15 @@ H5PEditor.language.core = {
   screenshots: 'Screenshots',
   reloadButtonLabel: 'Reload',
   videoQuality: 'Navn på videokvalitet',
-  videoQualityDescription: 'Dette navnet hjelper brukeren med å identifisere den aktuelle videokvaliteten. F.eks. 1080p, 720p, HD or Mobile',
+  videoQualityDescription:
+    'Dette navnet hjelper brukeren med å identifisere den aktuelle videokvaliteten. F.eks. 1080p, 720p, HD or Mobile',
   videoQualityDefaultLabel: 'Kvalitet :index',
   noContentTypesAvailable: 'Ingen innhaldstyper er tilgjengeleg',
-  noContentTypesAvailableDesc: 'Nettstedet ditt har problemer med å koble til H5P.org og liste tilgjengelege innhaldstyper.',
+  noContentTypesAvailableDesc:
+    'Nettstedet ditt har problemer med å koble til H5P.org og liste tilgjengelege innhaldstyper.',
   contentTypeCacheOutdated: 'Innhaldstype-lista er utdatert',
-  contentTypeCacheOutdatedDesc: 'Nettstedet ditt har problemer med å koble til H5P.org for å se etter innhaldstype-oppdateringer. Det kan hende at du ikke får oppdatert eller installert nye innhaldstyper.',
+  contentTypeCacheOutdatedDesc:
+    'Nettstedet ditt har problemer med å koble til H5P.org for å se etter innhaldstype-oppdateringer. Det kan hende at du ikke får oppdatert eller installert nye innhaldstyper.',
   tryAgain: 'Prøv på nytt',
   getHelp: 'Get help',
   untitled: 'Untitled :libraryTitle',
@@ -170,20 +179,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
   addNewChange: 'Add new change',
   confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
   logThisChange: 'Log this change',
   newChangeHasBeenLogged: 'New change has been logged',
   loggedChanges: 'Logged changes',
   noChangesHaveBeenLogged: 'No changes have been logged',
   errorHeader: 'An error occured',
-  errorCalculatingMaxScore: 'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
+  errorCalculatingMaxScore:
+    'Could not calculate the max score for this content. The max score is assumed to be 0. Contact your administrator if this isn’t correct.',
   maxScoreSemanticsMissing: 'Could not find the expected semantics in the content.',
   copyButton: 'Kopier',
   copiedButton: 'Kopiert',
   pasteButton: 'Lim inn',
   pasteAndReplaceButton: 'Paste & Replace',
   pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
+  confirmPasteContent:
+    'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
   confirmPasteButtonText: 'Replace content',
   copyToClipboard: 'Copy H5P content to the clipboard',
   copiedToClipboard: 'Content is copied to the clipboard',
@@ -193,8 +205,10 @@ H5PEditor.language.core = {
   pasteError: 'Cannot paste from clipboard',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insert using',
   tabTitleBasicFileUpload: 'File Upload',
@@ -207,14 +221,86 @@ H5PEditor.language.core = {
   language: 'Language',
   noLanguagesSupported: 'No languages supported',
   changeLanguage: 'Change language to :language?',
-  thisWillPotentially: "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
+  thisWillPotentially:
+    "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
   notAllTextsChanged: 'Not all texts were changed, there is only partial coverage for :language.',
-  contributeTranslations: 'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Dessverre var ikke den valgte \'%lib\'-innholdstypen installert på dette systemet.',
+  contributeTranslations:
+    'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Dessverre var ikke den valgte '%lib'-innholdstypen installert på dette systemet.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
