@@ -54,7 +54,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Wklej link YouTube lub dodaj plik wideo',
   uploadAudioTitle: 'Dodaj plik audio',
   uploadVideoTitle: 'Dodaj plik wideo',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Wstaw',
   cancel: 'Anuluj',
   height: 'długość',
@@ -74,7 +75,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Dodaj',
   uploadPlaceholder: 'Nie wybrano pliku',
   uploadInstructionsTitle: 'Dodaj plik h5.',
-  uploadInstructionsContent: 'Możesz rozpocząć korzystając z przykładów, które znajdziesz tutaj <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Możesz rozpocząć korzystając z przykładów, które znajdziesz tutaj <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Dodaj plik',
   uploadFileButtonChangeLabel: 'Zmień plik',
   uploadingThrobber: 'Trwa dodawanie...',
@@ -84,7 +86,8 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Nie można przesłać wybranego pliku',
   h5pFileWrongExtensionContent: 'Dopuszczalne są tylko pliki z rozszerzeniem .h5p.',
   h5pFileValidationFailedTitle: 'Wybrany plik H5P może być nieprawidłowy.',
-  h5pFileValidationFailedContent: 'Upewnij się, że przesyłane pliki zawierają prawidłową treść H5P. Pliki zawierające tylko biblioteki powinny być dodawane z formularza bibliotek H5P.',
+  h5pFileValidationFailedContent:
+    'Upewnij się, że przesyłane pliki zawierają prawidłową treść H5P. Pliki zawierające tylko biblioteki powinny być dodawane z formularza bibliotek H5P.',
   h5pFileUploadServerErrorTitle: 'Nie można przesłać pliku H5P',
   h5pFileUploadServerErrorContent: 'Wystąpił nieoczekiwany błąd. Sprawdź dziennik błędów serwera więcej szczegółów.',
   contentTypeSectionAll: 'Wszystkie typy zasobu',
@@ -116,7 +119,8 @@ H5PEditor.language.core = {
   readLess: 'Czytaj mniej',
   contentTypeOwner: 'Przez właściciela',
   contentTypeUnsupportedApiVersionTitle: 'Ten typ zasobu wymaga nowszej wersji podstawowej',
-  contentTypeUnsupportedApiVersionContent: 'Aby uzyskać niezbędne aktualizacje, skontaktuj się z administratorem systemu',
+  contentTypeUnsupportedApiVersionContent:
+    'Aby uzyskać niezbędne aktualizacje, skontaktuj się z administratorem systemu',
   contentTypeUpdateAvailable: 'Dostępna aktualizacja',
   contentTypeRestricted: 'Zablokowany typ zasobu',
   contentTypeRestrictedDesc: 'Użycie tego typu zasobu zostało zablokowane przez administratora.',
@@ -127,10 +131,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Nie można skomunikować się z hubem.',
   errorCommunicatingHubContent: 'Wystąpił błąd. Proszę spróbuj ponownie.',
   warningNoContentTypesInstalled: 'Nie masz zainstalowanych żadnych typów zasobu.',
-  warningChangeBrowsingToSeeResults: 'Kliknij przycisk <em>Wszystko</ em>, aby wyświetlić listę wszystkich typów zasobu, które można zainstalować.',
+  warningChangeBrowsingToSeeResults:
+    'Kliknij przycisk <em>Wszystko</ em>, aby wyświetlić listę wszystkich typów zasobu, które można zainstalować.',
   warningUpdateAvailableTitle: 'Dostępna jest nowa wersja :contentType.',
   warningUpdateAvailableBody: 'Zaktualizuj do najnowszej wersji, aby uzyskać pełną funkcjonalność.',
-  licenseDescription: 'Niektóre szczegóły licencji są wymienione poniżej. Kliknij ikonę informacji powyżej, aby przeczytać oryginalny tekst licencji.',
+  licenseDescription:
+    'Niektóre szczegóły licencji są wymienione poniżej. Kliknij ikonę informacji powyżej, aby przeczytać oryginalny tekst licencji.',
   licenseModalTitle: 'Szczegóły licencji',
   licenseModalSubtitle: 'Wybierz licencję, aby wyświetlić informacje o prawidłowym użyciu',
   licenseUnspecified: 'Nieokreślony',
@@ -150,12 +156,14 @@ H5PEditor.language.core = {
   screenshots: 'Miniatury',
   reloadButtonLabel: 'Przeładować',
   videoQuality: 'Etykieta jakości wideo',
-  videoQualityDescription: 'Ta etykieta pomaga użytkownikowi zorientować się w jakości wideo, np. 1080p, 720p, HD lub mobilne',
+  videoQualityDescription:
+    'Ta etykieta pomaga użytkownikowi zorientować się w jakości wideo, np. 1080p, 720p, HD lub mobilne',
   videoQualityDefaultLabel: 'Jakość :index',
   noContentTypesAvailable: 'Żaden typ zasobu nie jest dostępny',
   noContentTypesAvailableDesc: 'Twoja strona nie może połączyć się z H5P.org, aby wyświetlić dostępne typy zasobu.',
   contentTypeCacheOutdated: 'Lista typów zasobu jest nieaktualna',
-  contentTypeCacheOutdatedDesc: 'Twoja strona nie może połączyć się z H5P.org, aby wyświetlić dostępne aktualizacje typów zasobu. Aktualizowanie ani instalowanie typów zasobu nie jest teraz możliwe.',
+  contentTypeCacheOutdatedDesc:
+    'Twoja strona nie może połączyć się z H5P.org, aby wyświetlić dostępne aktualizacje typów zasobu. Aktualizowanie ani instalowanie typów zasobu nie jest teraz możliwe.',
   tryAgain: 'Spróbuj ponownie',
   getHelp: 'Pomoc',
   untitled: 'Bez tytułu :libraryTitle',
@@ -170,20 +178,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
   addNewChange: 'Dodaj nową zmianę',
   confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
   logThisChange: 'Zarejestruj tę zmianę',
   newChangeHasBeenLogged: 'Nowa zmiana została zarejestrowana',
   loggedChanges: 'Zarejestrowane zmiany',
   noChangesHaveBeenLogged: 'Nie ma zarejestrowanych zmian',
   errorHeader: 'Wystąpił błąd',
-  errorCalculatingMaxScore: 'Nie udało się obliczyć maksymalnego wyniku dla tego zasobu, zatem został on ustalony na 0. Jeśli to błąd, skontaktuj się z administratorem.',
+  errorCalculatingMaxScore:
+    'Nie udało się obliczyć maksymalnego wyniku dla tego zasobu, zatem został on ustalony na 0. Jeśli to błąd, skontaktuj się z administratorem.',
   maxScoreSemanticsMissing: 'W zasobie nie znaleziono spodziewanej semantyki.',
   copyButton: 'Kopiuj',
   copiedButton: 'Skopiowane',
   pasteButton: 'Wklej',
   pasteAndReplaceButton: 'Paste & Replace',
   pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
+  confirmPasteContent:
+    'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
   confirmPasteButtonText: 'Replace content',
   copyToClipboard: 'Skopiuj zasób H5P do schowka',
   copiedToClipboard: 'Zasób skopiowany do schowka',
@@ -193,8 +204,10 @@ H5PEditor.language.core = {
   pasteError: 'Nie można wkleić ze schowka',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insert using',
   tabTitleBasicFileUpload: 'File Upload',
@@ -207,14 +220,86 @@ H5PEditor.language.core = {
   language: 'Language',
   noLanguagesSupported: 'No languages supported',
   changeLanguage: 'Change language to :language?',
-  thisWillPotentially: "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
+  thisWillPotentially:
+    "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
   notAllTextsChanged: 'Not all texts were changed, there is only partial coverage for :language.',
-  contributeTranslations: 'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Unfortunately, the selected content type \'%lib\' isn\'t installed on this system.',
+  contributeTranslations:
+    'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Unfortunately, the selected content type '%lib' isn't installed on this system.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
