@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Adicionar :entity',
   tooLong: 'Valor do campo é muito longo, deve conter :max caracteres ou menos.',
   invalidFormat: 'Valor do campo contém um formato inválido ou caracteres proibidos.',
-  confirmChangeLibrary: 'Fazendo isso você perderá todo o trabalho concluído com o tipo atual de conteúdo. Tem certeza de que deseja alterar o tipo de conteúdo?',
+  confirmChangeLibrary:
+    'Fazendo isso você perderá todo o trabalho concluído com o tipo atual de conteúdo. Tem certeza de que deseja alterar o tipo de conteúdo?',
   commonFields: 'Sobreposição de texto e traduções',
   commonFieldsDescription: 'Aqui você pode editar configurações ou traduzir textos usados neste conteúdo.',
   uploading: 'Enviando, aguarde...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Cole aqui a URL do vídeo ou o link do YouTube',
   uploadAudioTitle: 'Enviar arquivo de áudio',
   uploadVideoTitle: 'Enviar arquivo de vídeo',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Inserir',
   cancel: 'Cancelar',
   height: 'altura',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Carregar',
   uploadPlaceholder: 'Nenhum arquivo selecionado',
   uploadInstructionsTitle: 'Carregar um arquivo H5P.',
-  uploadInstructionsContent: 'Você pode começar com exemplos em <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Você pode começar com exemplos em <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Carregar um arquivo',
   uploadFileButtonChangeLabel: 'Alterar arquivo',
   uploadingThrobber: 'Carregando...',
@@ -84,9 +87,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'O arquivo selecionado não pôde ser carregado',
   h5pFileWrongExtensionContent: 'Somente arquivos com a extensão .h5p são permitidos.',
   h5pFileValidationFailedTitle: 'Não foi possível validar o arquivo H5P.',
-  h5pFileValidationFailedContent: 'Tenha certeza que o H5P carregado possui conteúdo válido H5P. Arquivos H5P que contém apenas bibliotecas devem ser enviados através da página de Bibliotecas H5P.',
+  h5pFileValidationFailedContent:
+    'Tenha certeza que o H5P carregado possui conteúdo válido H5P. Arquivos H5P que contém apenas bibliotecas devem ser enviados através da página de Bibliotecas H5P.',
   h5pFileUploadServerErrorTitle: 'O arquivo H5P não pôde ser enviado',
-  h5pFileUploadServerErrorContent: 'Um erro inesperado ocorreu. Verifique seu o log de erros do servidor para maiores informações.',
+  h5pFileUploadServerErrorContent:
+    'Um erro inesperado ocorreu. Verifique seu o log de erros do servidor para maiores informações.',
   contentTypeSectionAll: 'Todos os tipos de conteúdo',
   searchResults: 'Procurar resultados',
   contentTypeSearchFieldPlaceholder: 'Procurar por tipos de conteúdo',
@@ -116,7 +121,8 @@ H5PEditor.language.core = {
   readLess: 'Veja menos',
   contentTypeOwner: 'Por :owner',
   contentTypeUnsupportedApiVersionTitle: 'Este tipo de conteúdo requer uma nova versão principal',
-  contentTypeUnsupportedApiVersionContent: 'Contate o administrador de seu sistema para lhe prover as atualizações necessárias',
+  contentTypeUnsupportedApiVersionContent:
+    'Contate o administrador de seu sistema para lhe prover as atualizações necessárias',
   contentTypeUpdateAvailable: 'Atualização disponível',
   contentTypeRestricted: 'Tipo de conteúdo restrito',
   contentTypeRestrictedDesc: 'O uso deste tipo de conteúdo foi restringido pelo administrador.',
@@ -127,10 +133,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Incapaz de se comunicar com o hub.',
   errorCommunicatingHubContent: 'Um erro ocorreu. Por favor, tente novamente.',
   warningNoContentTypesInstalled: 'Você não possui nenhum tipo de conteúdo instalado.',
-  warningChangeBrowsingToSeeResults: 'Clique em <em>todos</em> para obter uma lista de todos os conteúdos que você pode instalar.',
+  warningChangeBrowsingToSeeResults:
+    'Clique em <em>todos</em> para obter uma lista de todos os conteúdos que você pode instalar.',
   warningUpdateAvailableTitle: 'Uma nova versão de :contentType está disponível.',
   warningUpdateAvailableBody: 'Atualize para a última versão para uma experiência melhorada.',
-  licenseDescription: 'Algumas das características desta licença são indicadas abaixo. Clique o ícone de informações acima para ler o texto original da licença. ',
+  licenseDescription:
+    'Algumas das características desta licença são indicadas abaixo. Clique o ícone de informações acima para ler o texto original da licença. ',
   licenseModalTitle: 'Detalhes da licença',
   licenseModalSubtitle: 'Selecione uma licença para ver informações sobre seu uso apropriado',
   licenseUnspecified: 'Não informado',
@@ -150,12 +158,15 @@ H5PEditor.language.core = {
   screenshots: 'Capturas de tela',
   reloadButtonLabel: 'Recarregar',
   videoQuality: 'Rótulo de qualidade de vídeo',
-  videoQualityDescription: 'Este rótulo ajuda o usuário a identificar a qualidade atual do vídeo. Ex.: 1080p, 720p, HD ou Mobile',
+  videoQualityDescription:
+    'Este rótulo ajuda o usuário a identificar a qualidade atual do vídeo. Ex.: 1080p, 720p, HD ou Mobile',
   videoQualityDefaultLabel: 'Qualidade :index',
   noContentTypesAvailable: 'Nenhum tipo de conteúdo está disponível',
-  noContentTypesAvailableDesc: 'Seu site está tendo dificuldades ao conectar ao H5P.org e listar os tipos de conteúdo disponíveis.',
+  noContentTypesAvailableDesc:
+    'Seu site está tendo dificuldades ao conectar ao H5P.org e listar os tipos de conteúdo disponíveis.',
   contentTypeCacheOutdated: 'Lista de tipos de conteúdo desatualizada',
-  contentTypeCacheOutdatedDesc: 'Seu site está tendo dificuldades ao conectar ao H5P.org e verificar os tipos de conteúdo disponíveis. Você pode não ser capaz de atualizar ou instalar novos tipos de conteúdo.',
+  contentTypeCacheOutdatedDesc:
+    'Seu site está tendo dificuldades ao conectar ao H5P.org e verificar os tipos de conteúdo disponíveis. Você pode não ser capaz de atualizar ou instalar novos tipos de conteúdo.',
   tryAgain: 'Tentar novamente',
   getHelp: 'Obter ajuda',
   untitled: 'Sem título',
@@ -170,20 +181,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Are you sure you would like to remove this author?',
   addNewChange: 'Adicionar nova alteração',
   confirmDeleteChangeLog: 'Are you sure you would like to delete this change log entry?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
   logThisChange: 'Registrar alteração',
   newChangeHasBeenLogged: 'A nova alteração foi registrada no log',
   loggedChanges: 'Alterações registradas',
   noChangesHaveBeenLogged: 'Nenhuma alteração foi registrada',
   errorHeader: 'Um erro ocorreu',
-  errorCalculatingMaxScore: 'Não foi possível calcular a pontuação máxima para este conteúdo. A pontuação máxima é assumida como sendo 0. Contate seu administrador se isto não estiver correto.',
+  errorCalculatingMaxScore:
+    'Não foi possível calcular a pontuação máxima para este conteúdo. A pontuação máxima é assumida como sendo 0. Contate seu administrador se isto não estiver correto.',
   maxScoreSemanticsMissing: 'Não foi possível encontrar a semântica esperada no conteúdo.',
   copyButton: 'Copiar',
   copiedButton: 'Copiado',
   pasteButton: 'Colar',
   pasteAndReplaceButton: 'Paste & Replace',
   pasteContent: 'Replace Content',
-  confirmPasteContent: 'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
+  confirmPasteContent:
+    'By doing this you will replace the current content with the content from your clipboard. The current content will be lost. Are you sure you wish to continue?',
   confirmPasteButtonText: 'Replace content',
   copyToClipboard: 'Copiar conteúdo H5P para o painel',
   copiedToClipboard: 'O conteúdo foi copiado para o painel',
@@ -193,8 +207,10 @@ H5PEditor.language.core = {
   pasteError: 'Não foi possível colar',
   pasteContentNotSupported: 'The content in the H5P clipboard is not supported in this context',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
-  pasteTooNew: 'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
+  pasteTooOld:
+    'The content in the H5P clipboard is of a lower version (:clip) than what is supported in this context (:local), if possible try to have the content you want to paste upgraded, copy it again and try pasting it here.',
+  pasteTooNew:
+    'The content in the H5P clipboard is of a higher version (:clip) than what is supported in this context (:local), if possible try to have this content upgraded first, and then try pasting the content here again.',
   ok: 'OK',
   avTablistLabel: 'Insert using',
   tabTitleBasicFileUpload: 'File Upload',
@@ -207,14 +223,86 @@ H5PEditor.language.core = {
   language: 'Language',
   noLanguagesSupported: 'No languages supported',
   changeLanguage: 'Change language to :language?',
-  thisWillPotentially: "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
+  thisWillPotentially:
+    "This will potentially reset all the text and translations. You can't undo this. The content itself will not be changed. Do you want to proceed?",
   notAllTextsChanged: 'Not all texts were changed, there is only partial coverage for :language.',
-  contributeTranslations: 'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Unfortunately, the selected content type \'%lib\' isn\'t installed on this system.',
+  contributeTranslations:
+    'If you want to complete the translation for :language you can learn about <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Unfortunately, the selected content type '%lib' isn't installed on this system.",
   proceedButtonLabel: 'Proceed to save',
   enterFullscreenButtonLabel: 'Enter fullscreen',
   exitFullscreenButtonLabel: 'Exit fullscreen',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
