@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Adicionar :entity',
   tooLong: 'O valor do campo é muito longo. Deve conter :max caracteres no máximo.',
   invalidFormat: 'O valor do campo contém um formato inválido ou caracteres não permitidos.',
-  confirmChangeLibrary: 'Se continuar, perderá todo o trabalho concluído com o tipo atual de conteúdo. Tem a certeza de que pretende alterar o tipo de conteúdo?',
+  confirmChangeLibrary:
+    'Se continuar, perderá todo o trabalho concluído com o tipo atual de conteúdo. Tem a certeza de que pretende alterar o tipo de conteúdo?',
   commonFields: 'Substituição de texto e traduções',
   commonFieldsDescription: 'Aqui pode editar as configurações ou traduzir textos usados neste conteúdo.',
   uploading: 'A enviar. Por favor, aguarde...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Cole aqui o URL do vídeo ou a hiperligação do YouTube',
   uploadAudioTitle: 'Enviar ficheiro de áudio',
   uploadVideoTitle: 'Enviar ficheiro de vídeo',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Inserir',
   cancel: 'Cancelar',
   height: 'altura',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Carregar',
   uploadPlaceholder: 'Nenhum ficheiro selecionado',
   uploadInstructionsTitle: 'Carregar um ficheiro H5P.',
-  uploadInstructionsContent: 'Pode começar com exemplos existentes em <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Pode começar com exemplos existentes em <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Carregar um ficheiro',
   uploadFileButtonChangeLabel: 'Alterar ficheiro',
   uploadingThrobber: 'A carregar...',
@@ -84,9 +87,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Não foi possível carregar o ficheiro',
   h5pFileWrongExtensionContent: 'Apenas são permitidos ficheiros com a extensão .h5p.',
   h5pFileValidationFailedTitle: 'Não foi possível validar o ficheiro H5P.',
-  h5pFileValidationFailedContent: 'Certifique-se que o ficheiros H5P carregado possui conteúdo H5P válido. Ficheiros H5P com bibliotecas devem ser enviados através da página de Bibliotecas H5P.',
+  h5pFileValidationFailedContent:
+    'Certifique-se que o ficheiros H5P carregado possui conteúdo H5P válido. Ficheiros H5P com bibliotecas devem ser enviados através da página de Bibliotecas H5P.',
   h5pFileUploadServerErrorTitle: 'Não foi possível enviar o ficheiro H5P',
-  h5pFileUploadServerErrorContent: 'Ocorreu um erro inesperado. Para mais informações, consulte o registo de erros do servidor.',
+  h5pFileUploadServerErrorContent:
+    'Ocorreu um erro inesperado. Para mais informações, consulte o registo de erros do servidor.',
   contentTypeSectionAll: 'Todos os tipos de conteúdo',
   searchResults: 'Procurar resultados',
   contentTypeSearchFieldPlaceholder: 'Procurar por tipos de conteúdo',
@@ -116,7 +121,8 @@ H5PEditor.language.core = {
   readLess: 'Ver menos',
   contentTypeOwner: 'Por :owner',
   contentTypeUnsupportedApiVersionTitle: 'Este tipo de conteúdo requer uma nova versão do core',
-  contentTypeUnsupportedApiVersionContent: 'Contacte o administrador de seu sistema para lhe disponibilize as atualizações necessárias',
+  contentTypeUnsupportedApiVersionContent:
+    'Contacte o administrador de seu sistema para lhe disponibilize as atualizações necessárias',
   contentTypeUpdateAvailable: 'Atualização disponível',
   contentTypeRestricted: 'Tipo de conteúdo restrito',
   contentTypeRestrictedDesc: 'O uso deste tipo de conteúdo foi restringido pelo administrador.',
@@ -127,10 +133,12 @@ H5PEditor.language.core = {
   errorCommunicatingHubTitle: 'Não foi possível comunicar com o Hub.',
   errorCommunicatingHubContent: 'Ocorreu um erro. Por favor, tente novamente.',
   warningNoContentTypesInstalled: 'Nenhum tipo de conteúdo está instalado.',
-  warningChangeBrowsingToSeeResults: 'Clique em <em>todos</em> para obter a lista de todos os conteúdos que pode instalar.',
+  warningChangeBrowsingToSeeResults:
+    'Clique em <em>todos</em> para obter a lista de todos os conteúdos que pode instalar.',
   warningUpdateAvailableTitle: 'Está disponível uma nova versão de :contentType.',
   warningUpdateAvailableBody: 'Atualize para a última versão para uma melhor experiência.',
-  licenseDescription: 'Algumas das características desta licença são indicadas abaixo. Clique no ícone de informações acima para ler o texto original da licença.',
+  licenseDescription:
+    'Algumas das características desta licença são indicadas abaixo. Clique no ícone de informações acima para ler o texto original da licença.',
   licenseModalTitle: 'Detalhes da licença',
   licenseModalSubtitle: 'Selecione uma licença para ver informações sobre o uso apropriado',
   licenseUnspecified: 'Não especificado',
@@ -150,12 +158,15 @@ H5PEditor.language.core = {
   screenshots: 'Capturas de ecrã',
   reloadButtonLabel: 'Recarregar',
   videoQuality: 'Etiqueta da qualidade de vídeo',
-  videoQualityDescription: 'Esta etiqueta ajuda o utilizador a identificar a qualidade atual do vídeo. Ex.: 1080p, 720p, HD ou Mobile',
+  videoQualityDescription:
+    'Esta etiqueta ajuda o utilizador a identificar a qualidade atual do vídeo. Ex.: 1080p, 720p, HD ou Mobile',
   videoQualityDefaultLabel: 'Qualidade :index',
   noContentTypesAvailable: 'Nenhum tipo de conteúdo está disponível',
-  noContentTypesAvailableDesc: 'O seu site está a ter dificuldades na ligação a H5P.org para obter a listagem os tipos de conteúdo disponíveis.',
+  noContentTypesAvailableDesc:
+    'O seu site está a ter dificuldades na ligação a H5P.org para obter a listagem os tipos de conteúdo disponíveis.',
   contentTypeCacheOutdated: 'Lista de tipos de conteúdo desatualizada',
-  contentTypeCacheOutdatedDesc: 'O seu site está a ter dificuldades na ligação a H5P.org para verificar a existência de atualizações disponíveis. Pode não ser capaz de atualizar ou instalar novos tipos de conteúdo.',
+  contentTypeCacheOutdatedDesc:
+    'O seu site está a ter dificuldades na ligação a H5P.org para verificar a existência de atualizações disponíveis. Pode não ser capaz de atualizar ou instalar novos tipos de conteúdo.',
   tryAgain: 'Tentar novamente',
   getHelp: 'Ajuda',
   untitled: 'Sem título :libraryTitle',
@@ -170,20 +181,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Tem a certeza de que pretende remover este autor?',
   addNewChange: 'Adicionar nova alteração',
   confirmDeleteChangeLog: 'Tem a certeza de que pretende excluir esta entrada do registo de alterações?',
-  changelogDescription: 'Algumas licenças exigem que sejam registadas e exibidas as alterações feitas no trabalho original ou derivados. Pode registar aqui as suas alterações por motivos de licenciamento ou apenas para permitir que possa, tal como outras pessoas, acompanhar as alterações feitas neste conteúdo.',
+  changelogDescription:
+    'Algumas licenças exigem que sejam registadas e exibidas as alterações feitas no trabalho original ou derivados. Pode registar aqui as suas alterações por motivos de licenciamento ou apenas para permitir que possa, tal como outras pessoas, acompanhar as alterações feitas neste conteúdo.',
   logThisChange: 'Registar alteração',
   newChangeHasBeenLogged: 'A nova alteração foi guardada no registo',
   loggedChanges: 'Alterações registadas',
   noChangesHaveBeenLogged: 'Nenhuma alteração foi registada',
   errorHeader: 'Ocorreu um erro',
-  errorCalculatingMaxScore: 'Não foi possível calcular a pontuação máxima para este conteúdo. A pontuação máxima é assumida como sendo 0. Se achar que não está correto, contacte o seu administrador.',
+  errorCalculatingMaxScore:
+    'Não foi possível calcular a pontuação máxima para este conteúdo. A pontuação máxima é assumida como sendo 0. Se achar que não está correto, contacte o seu administrador.',
   maxScoreSemanticsMissing: 'Não foi possível encontrar a semântica esperada no conteúdo.',
   copyButton: 'Copiar',
   copiedButton: 'Copiado',
   pasteButton: 'Colar',
   pasteAndReplaceButton: 'Colar & Substituir',
   pasteContent: 'Substituir conteúdo',
-  confirmPasteContent: 'Se continuar, irá substituir o conteúdo atual pelo conteúdo da área de transferência. O conteúdo atual será perdido. Tem a certeza de que pretende continuar?',
+  confirmPasteContent:
+    'Se continuar, irá substituir o conteúdo atual pelo conteúdo da área de transferência. O conteúdo atual será perdido. Tem a certeza de que pretende continuar?',
   confirmPasteButtonText: 'Substituir conteúdo',
   copyToClipboard: 'Copiar conteúdo H5P para a área de transferência',
   copiedToClipboard: 'O conteúdo foi copiado para a área de transferência',
@@ -193,8 +207,10 @@ H5PEditor.language.core = {
   pasteError: 'Não foi possível colar',
   pasteContentNotSupported: 'O conteúdo que está a tentar colar não é suportado neste contexto',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'O conteúdo que está a tentar colar é de uma versão inferior (:clip) à versão suportada neste contexto (:local). Se possível, tente atualizar o conteúdo que deseja colar, copie-o novamente e tente colá-lo aqui.',
-  pasteTooNew: 'O conteúdo que está a tentar colar é de uma versão superior (:clip) à versão suportada neste contexto (:local). Se possível, tente atualizar este conteúdo, e depois tente colar novamente.',
+  pasteTooOld:
+    'O conteúdo que está a tentar colar é de uma versão inferior (:clip) à versão suportada neste contexto (:local). Se possível, tente atualizar o conteúdo que deseja colar, copie-o novamente e tente colá-lo aqui.',
+  pasteTooNew:
+    'O conteúdo que está a tentar colar é de uma versão superior (:clip) à versão suportada neste contexto (:local). Se possível, tente atualizar este conteúdo, e depois tente colar novamente.',
   ok: 'OK',
   avTablistLabel: 'Inserir usando',
   tabTitleBasicFileUpload: 'Carregar ficheiro',
@@ -207,14 +223,86 @@ H5PEditor.language.core = {
   language: 'Idioma',
   noLanguagesSupported: 'Nenhum idioma suportado',
   changeLanguage: 'Alterar idioma para :language?',
-  thisWillPotentially: 'Isto redefinirá potencialmente todos os textos e traduções. Não pode desfazer esta ação. O conteúdo em si não será alterado. Pretende prosseguir?',
+  thisWillPotentially:
+    'Isto redefinirá potencialmente todos os textos e traduções. Não pode desfazer esta ação. O conteúdo em si não será alterado. Pretende prosseguir?',
   notAllTextsChanged: 'Nem todos os textos foram alterados. Existe apenas uma tradução parcial para :language.',
-  contributeTranslations: 'Se pretende completar a tradução para :language, clique para obter informações sobre <a href=":url" target="_new">contribuir com traduções para H5P</a>',
-  unknownLibrary: 'Infelizmente, o tipo de conteúdo selecionado \'%lib\' não está instalado neste sistema.',
+  contributeTranslations:
+    'Se pretende completar a tradução para :language, clique para obter informações sobre <a href=":url" target="_new">contribuir com traduções para H5P</a>',
+  unknownLibrary: "Infelizmente, o tipo de conteúdo selecionado '%lib' não está instalado neste sistema.",
   proceedButtonLabel: 'Prosseguir para guardar',
   enterFullscreenButtonLabel: 'Mostrar em ecrã inteiro',
   exitFullscreenButtonLabel: 'Sair de ecrã inteiro',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
-  reuseSuccess: ':title was successfully imported from the H5P Hub.'
+  reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
