@@ -7,7 +7,8 @@ H5PEditor.language.core = {
   notImageOrDimensionsField: '":path" не является изображением или полем для размеров.',
   requiredProperty: 'Это :property обязательно и должно иметь значение.',
   onlyNumbers: 'Это :property значение может содержать только цифры.',
-  illegalDecimalNumber: 'Это :property может содержать только числа с максимальным количеством :decimals десятичных знаков.',
+  illegalDecimalNumber:
+    'Это :property может содержать только числа с максимальным количеством :decimals десятичных знаков.',
   exceedsMax: 'Это :property значение превышает максимум :max.',
   listExceedsMax: 'Список превышает максимальное количество :max элементов.',
   belowMin: 'Это :property значение ниже минимума :min.',
@@ -27,9 +28,11 @@ H5PEditor.language.core = {
   addEntity: 'Добавить :entity',
   tooLong: 'Значение поля слишком длинное, должно содержать :max знаков или меньше.',
   invalidFormat: 'Значение поля содержит неверный формат или символы, которые запрещены.',
-  confirmChangeLibrary: 'Делая это, вы потеряете всю работу, выполненную с текущим типом контента. Вы уверены, что хотите изменить тип контента?',
+  confirmChangeLibrary:
+    'Делая это, вы потеряете всю работу, выполненную с текущим типом контента. Вы уверены, что хотите изменить тип контента?',
   commonFields: 'Переопределение текста и переводы',
-  commonFieldsDescription: 'Здесь вы можете редактировать настройки или переводить тексты, используемые в этом контенте.',
+  commonFieldsDescription:
+    'Здесь вы можете редактировать настройки или переводить тексты, используемые в этом контенте.',
   uploading: 'Загрузка, пожалуйста, подождите ...',
   noFollow: 'Недопустимо использовать поле ":path".',
   editCopyright: 'Изменить авторское право',
@@ -54,7 +57,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Вставить ссылку на YouTube или другой источник видео',
   uploadAudioTitle: 'Загрузить аудио файл',
   uploadVideoTitle: 'Загрузить видео файл',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Вставить',
   cancel: 'Отмена',
   height: 'высота',
@@ -74,7 +78,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Загрузить',
   uploadPlaceholder: 'Файл не выбран',
   uploadInstructionsTitle: 'Загрузить H5P файл.',
-  uploadInstructionsContent: 'Вы можете начать с примеров из <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Вы можете начать с примеров из <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Загрузить файл',
   uploadFileButtonChangeLabel: 'Изменить файл',
   uploadingThrobber: 'Загрузка...',
@@ -84,11 +89,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Выбранный файл не может быть загружен',
   h5pFileWrongExtensionContent: 'Разрешены только файлы с расширением .h5p.',
   h5pFileValidationFailedTitle: 'Не удалось проверить H5P файл.',
-  h5pFileValidationFailedContent: 'Убедитесь, что загруженный H5P содержит действительный контент H5P. H5P' +
-  ' файлы, содержащие только библиотеки, должны быть загружены через страницу библиотек H5P.',
+  h5pFileValidationFailedContent:
+    'Убедитесь, что загруженный H5P содержит действительный контент H5P. H5P файлы, содержащие только библиотеки, должны быть загружены через страницу библиотек H5P.',
   h5pFileUploadServerErrorTitle: 'Файл H5P не может быть загружен',
-  h5pFileUploadServerErrorContent: 'Произошла непредвиденная ошибка. Проверьте журнал ошибок вашего сервера для' +
-  ' больших деталей.',
+  h5pFileUploadServerErrorContent:
+    'Произошла непредвиденная ошибка. Проверьте журнал ошибок вашего сервера для больших деталей.',
   contentTypeSectionAll: 'Все типы контента',
   searchResults: 'Результаты поиска',
   contentTypeSearchFieldPlaceholder: 'Поиск типов контента',
@@ -118,7 +123,8 @@ H5PEditor.language.core = {
   readLess: 'Читать меньше',
   contentTypeOwner: 'От :owner',
   contentTypeUnsupportedApiVersionTitle: 'Этот тип контента требует более новой версии ядра',
-  contentTypeUnsupportedApiVersionContent: 'Обратитесь к системному администратору, чтобы предоставить вам необходимые обновления',
+  contentTypeUnsupportedApiVersionContent:
+    'Обратитесь к системному администратору, чтобы предоставить вам необходимые обновления',
   contentTypeUpdateAvailable: 'Доступно обновление',
   contentTypeRestricted: 'Ограниченный тип контента',
   contentTypeRestrictedDesc: 'Использование этого типа контента было ограничено администратором.',
@@ -128,11 +134,13 @@ H5PEditor.language.core = {
   currentMenuSelected: 'текущий выбор',
   errorCommunicatingHubTitle: 'Не может соединиться с хабом.',
   errorCommunicatingHubContent: 'Произошла ошибка. Пожалуйста, попробуйте еще раз.',
-  warningNoContentTypesInstalled: "У вас не установлены типы контента",
-  warningChangeBrowsingToSeeResults: 'Нажмите на <em>Все</em> чтобы получить список всех типов контента, которые вы можете установить.',
+  warningNoContentTypesInstalled: 'У вас не установлены типы контента',
+  warningChangeBrowsingToSeeResults:
+    'Нажмите на <em>Все</em> чтобы получить список всех типов контента, которые вы можете установить.',
   warningUpdateAvailableTitle: 'Новая версия :contentType доступна.',
   warningUpdateAvailableBody: 'Обновление до последней версии для улучшения опыта.',
-  licenseDescription: 'Некоторые функции этой лицензии указаны ниже. Нажмите на иконку информации выше, чтобы прочитать исходный текст лицензии.',
+  licenseDescription:
+    'Некоторые функции этой лицензии указаны ниже. Нажмите на иконку информации выше, чтобы прочитать исходный текст лицензии.',
   licenseModalTitle: 'Детали лицензии',
   licenseModalSubtitle: 'Выберите лицензию для просмотра информации о правильном использовании',
   licenseUnspecified: 'Unspecified',
@@ -152,12 +160,14 @@ H5PEditor.language.core = {
   screenshots: 'Скриншоты',
   reloadButtonLabel: 'Обновить',
   videoQuality: 'Надпись разрешения видео',
-  videoQualityDescription: 'Эта надпись помогает пользователю определить текущее качество видео. Например. 1080p, 720p, HD или мобильный',
+  videoQualityDescription:
+    'Эта надпись помогает пользователю определить текущее качество видео. Например. 1080p, 720p, HD или мобильный',
   videoQualityDefaultLabel: 'Разрешение видео :index',
   noContentTypesAvailable: 'Нет доступных типов контента',
   noContentTypesAvailableDesc: 'У вашего сайта проблемы с подключением к H5P.org и списку доступных типов контента.',
   contentTypeCacheOutdated: 'Список типов контента устарел',
-  contentTypeCacheOutdatedDesc: 'Ваш сайт испытывает трудности при подключении к H5P.org для проверки наличия обновлений типа контента. Возможно, вы не сможете обновить или установить новые типы контента.',
+  contentTypeCacheOutdatedDesc:
+    'Ваш сайт испытывает трудности при подключении к H5P.org для проверки наличия обновлений типа контента. Возможно, вы не сможете обновить или установить новые типы контента.',
   tryAgain: 'Попробуйте снова',
   getHelp: 'Получить помощь',
   untitled: 'Без названия :libraryTitle',
@@ -166,26 +176,29 @@ H5PEditor.language.core = {
   addTitle: 'Добавить название',
   usedForSearchingReportsAndCopyrightInformation: 'Используется для поиска, отчетов и информации об авторских правах',
   metadataSharingAndLicensingInfo: 'Метаданные (информация о совместном использовании и лицензировании)',
-  fillInTheFieldsBelow : 'Заполните поля ниже',
+  fillInTheFieldsBelow: 'Заполните поля ниже',
   saveMetadata: 'Сохранить метаданные',
   addAuthor: 'Сохранить автора',
   confirmRemoveAuthor: 'Вы уверены, что хотите удалить этого автора?',
   addNewChange: 'Добавить новое изменение',
   confirmDeleteChangeLog: 'Вы уверены, что хотите удалить эту запись в журнале изменений?',
-  changelogDescription: 'Некоторые лицензии требуют, чтобы изменения, внесенные в оригинальную работу или производные, регистрировались и отображались. Вы можете регистрировать свои изменения здесь по причинам лицензирования или просто для того, чтобы позволить себе и другим отслеживать изменения, внесенные в этот контент.',
+  changelogDescription:
+    'Некоторые лицензии требуют, чтобы изменения, внесенные в оригинальную работу или производные, регистрировались и отображались. Вы можете регистрировать свои изменения здесь по причинам лицензирования или просто для того, чтобы позволить себе и другим отслеживать изменения, внесенные в этот контент.',
   logThisChange: 'Зарегистрировать это изменение',
   newChangeHasBeenLogged: 'Новое изменение было зарегистрировано',
   loggedChanges: 'Зарегистрированные изменения',
   noChangesHaveBeenLogged: 'Изменение не было зарегистрировано',
   errorHeader: 'Произошла ошибка',
-  errorCalculatingMaxScore: 'Не удалось рассчитать максимальный балл для этого контента. Максимальный балл принимается равным 0. Свяжитесь с администратором, если это не так.',
+  errorCalculatingMaxScore:
+    'Не удалось рассчитать максимальный балл для этого контента. Максимальный балл принимается равным 0. Свяжитесь с администратором, если это не так.',
   maxScoreSemanticsMissing: 'Не удалось найти ожидаемую семантику в содержании.',
   copyButton: 'Копировать',
   copiedButton: 'Скопировано',
   pasteButton: 'Вставить',
   pasteAndReplaceButton: 'Вставить и Заменить',
   pasteContent: 'Заменить содержимое',
-  confirmPasteContent: 'Сделав это, вы замените текущий контент содержимым из буфера обмена. Текущий контент будет потерян. Вы уверены, что хотите продолжить?',
+  confirmPasteContent:
+    'Сделав это, вы замените текущий контент содержимым из буфера обмена. Текущий контент будет потерян. Вы уверены, что хотите продолжить?',
   confirmPasteButtonText: 'Заменить контент',
   copyToClipboard: 'Скопировать содержимое H5P в буфер обмена',
   copiedToClipboard: 'Контент скопирован в буфер обмена',
@@ -195,8 +208,10 @@ H5PEditor.language.core = {
   pasteError: 'Не удалось вставить из буфера обмена',
   pasteContentNotSupported: 'Содержимое в буфере обмена H5P не поддерживается в этом контексте.',
   pasteContentRestricted: 'Содержимое буфера обмена было ограничено на этом сайте.',
-  pasteTooOld: 'Содержимое буфера обмена H5P имеет более низкую версию (:clip), чем поддерживаемое в данном контексте (:local), если возможно, попытайтесь вставить обновленное содержимое, скопировать его снова и вставить его сюда.',
-  pasteTooNew: 'Содержимое буфера обмена H5P имеет более высокую версию (:clip), чем поддерживаемое в данном контексте (:local), по возможности сначала попытайтесь обновить это содержимое, а затем попытайтесь вставить его сюда снова.',
+  pasteTooOld:
+    'Содержимое буфера обмена H5P имеет более низкую версию (:clip), чем поддерживаемое в данном контексте (:local), если возможно, попытайтесь вставить обновленное содержимое, скопировать его снова и вставить его сюда.',
+  pasteTooNew:
+    'Содержимое буфера обмена H5P имеет более высокую версию (:clip), чем поддерживаемое в данном контексте (:local), по возможности сначала попытайтесь обновить это содержимое, а затем попытайтесь вставить его сюда снова.',
   ok: 'ОК',
   avTablistLabel: 'Вставить с помощью',
   tabTitleBasicFileUpload: 'Загрузка файла',
@@ -209,14 +224,86 @@ H5PEditor.language.core = {
   language: 'Язык',
   noLanguagesSupported: 'Языки не поддерживаются',
   changeLanguage: 'Изменить язык на :language?',
-  thisWillPotentially: "Это потенциально сбросит весь текст и переводы. Вы не можете отменить это. Сам контент не будет изменен. Вы хотите продолжить?",
+  thisWillPotentially:
+    'Это потенциально сбросит весь текст и переводы. Вы не можете отменить это. Сам контент не будет изменен. Вы хотите продолжить?',
   notAllTextsChanged: 'Не все тексты были изменены, есть только частичные переводы для :language.',
-  contributeTranslations: 'Если вы хотите завершить перевод для :language вы можете узнать о том как <a href=":url" target="_new">внести переводы в H5P</a>',
-  unknownLibrary: 'К сожалению, выбранный тип контента \'%lib\' isn\ не установлен в этой системе.',
+  contributeTranslations:
+    'Если вы хотите завершить перевод для :language вы можете узнать о том как <a href=":url" target="_new">внести переводы в H5P</a>',
+  unknownLibrary: "К сожалению, выбранный тип контента '%lib' isn не установлен в этой системе.",
   proceedButtonLabel: 'Перейти к сохранению',
   enterFullscreenButtonLabel: 'Войти в полноэкранный режим',
   exitFullscreenButtonLabel: 'Покинуть полноэкранный режим',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
