@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Dodaj :entity',
   tooLong: 'Polje je predolgo. Vsebovati mora največ naslednje število črk: :max.',
   invalidFormat: 'Vrednost polja vsebuje neveljaven format ali nedovoljene znake.',
-  confirmChangeLibrary: 'S spremembo boste izgubili vso opravljeno delo v trenutni vsebini. Ali ste prepričani, da želite nadaljevati?',
+  confirmChangeLibrary:
+    'S spremembo boste izgubili vso opravljeno delo v trenutni vsebini. Ali ste prepričani, da želite nadaljevati?',
   commonFields: 'Preglasitve besed in prevodi',
   commonFieldsDescription: 'Spremembe besed v tem razdelku bodo veljale le za to vsebino.',
   uploading: 'Nalagam, počakajte prosim ... ',
@@ -49,12 +50,13 @@ H5PEditor.language.core = {
   selectFiletoUpload: 'Izberite datoteko za prenos',
   or: 'ali',
   enterAudioUrl: 'Vpis spletnega naslova (URL)',
-  enterVideoUrl: 'Enter video URL',
+  enterVideoUrl: 'Dodaj spletni naslov (URL) do videoposnetka',
   enterAudioTitle: 'Dodaj spletni naslov (URL) do zvočnega posnetka',
   enterVideoTitle: 'Dodaj spletni naslov (URL) do YouTube ali drugega videoposnetka',
   uploadAudioTitle: 'Naloži zvočni posnetek',
   uploadVideoTitle: 'Naloži videoposnetek',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P podpira vse zunanje videoposnetke v formatu mp4, webm ali ogv. Tudi povezave do posnetkov na Vimeo Pro, YouTube in Panopto.',
   insert: 'Vstavi',
   cancel: 'Prekliči',
   height: 'višina',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Naloži iz datoteke',
   uploadPlaceholder: 'Izbrana ni nobena datoteka',
   uploadInstructionsTitle: 'Naloži datoteko H5P.',
-  uploadInstructionsContent: 'Začnete lahko s primeri iz <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Začnete lahko s primeri iz <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Naloži',
   uploadFileButtonChangeLabel: 'Zamenjaj datoteko',
   uploadingThrobber: 'Nalagam datoteko ...',
@@ -84,11 +87,10 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Izbrane datoteke ni mogoče naložiti',
   h5pFileWrongExtensionContent: 'Dovoljene so le datoteke s končnico .h5p.',
   h5pFileValidationFailedTitle: 'Datoteke H5P ni bilo mogoče preveriti.',
-  h5pFileValidationFailedContent: 'Prepričajte se, da naložena datoteka vsebuje veljavno vsebino H5P. Datoteke' +
-  ' knjižnic je treba naložiti preko strani H5P knjižnic.',
+  h5pFileValidationFailedContent:
+    'Prepričajte se, da naložena datoteka vsebuje veljavno vsebino H5P. Datoteke knjižnic je treba naložiti preko strani H5P knjižnic.',
   h5pFileUploadServerErrorTitle: 'Datoteke H5P ni bilo mogoče naložiti',
-  h5pFileUploadServerErrorContent: 'Prišlo je do neznane napake. Preverite dnevnik napak za' +
-  ' več informacij.',
+  h5pFileUploadServerErrorContent: 'Prišlo je do neznane napake. Preverite dnevnik napak za več informacij.',
   contentTypeSectionAll: 'Vsi tipi vsebin',
   searchResults: 'Rezultati iskanja',
   contentTypeSearchFieldPlaceholder: 'Iskanje tipov vsebin',
@@ -128,11 +130,13 @@ H5PEditor.language.core = {
   currentMenuSelected: 'trenutni izbor',
   errorCommunicatingHubTitle: 'Ni mogoče komunicirati s središčem H5P.',
   errorCommunicatingHubContent: 'Prišlo je do neznane napake. Poskusite znova.',
-  warningNoContentTypesInstalled: "Nimate nameščenih nobenih tipov vsebine.",
-  warningChangeBrowsingToSeeResults: 'Izberite <em>All</em> za prikaz seznama vseh tipov vsebine, ki jih lahko namestite.',
+  warningNoContentTypesInstalled: 'Nimate nameščenih nobenih tipov vsebine.',
+  warningChangeBrowsingToSeeResults:
+    'Izberite <em>All</em> za prikaz seznama vseh tipov vsebine, ki jih lahko namestite.',
   warningUpdateAvailableTitle: 'Na voljo je nova različica tipa vsebine :contentType.',
   warningUpdateAvailableBody: 'Posodobitev na najnovejšo različico izboljša uporabniško izkušnjo.',
-  licenseDescription: 'Nekatere značilnosti te licence so navedene spodaj. Ogled izvornih informacij o licenci je dosegljiv s klikom na ikono (i) zgoraj.',
+  licenseDescription:
+    'Nekatere značilnosti te licence so navedene spodaj. Ogled izvornih informacij o licenci je dosegljiv s klikom na ikono(i) zgoraj.',
   licenseModalTitle: 'Podrobnosti o licenci',
   licenseModalSubtitle: 'Izberite licenco za ogled informacij o pravilni uporabi',
   licenseUnspecified: 'Neopredeljeno',
@@ -143,7 +147,7 @@ H5PEditor.language.core = {
   licenseCanHoldLiable: 'Lahko prenaša odgovornost',
   licenseCannotHoldLiable: 'Ne prenaša odgovornosti',
   licenseMustIncludeCopyright: 'Vsebovati mora avtorske pravice',
-  licenseMustIncludeLicense: 'Vsebovati mora license',
+  licenseMustIncludeLicense: 'Vsebovati mora licenco',
   licenseFetchDetailsFailed: 'Podatkov o licenci ni bilo mogoče pridobiti',
   imageLightboxTitle: 'Slike',
   imageLightBoxProgress: ':num od :total',
@@ -152,40 +156,47 @@ H5PEditor.language.core = {
   screenshots: 'Zaslonski posnetek',
   reloadButtonLabel: 'Ponovno naloži',
   videoQuality: 'Oznaka kakovosti videoposnetka',
-  videoQualityDescription: 'Ta oznaka pomaga uporabniku prepoznati trenutno kakovost videoposnetka. Npr. 1080p, 720p, HD ali Mobilne naprave',
+  videoQualityDescription:
+    'Ta oznaka pomaga uporabniku prepoznati trenutno kakovost videoposnetka. Npr. 1080p, 720p, HD ali Mobilne naprave',
   videoQualityDefaultLabel: 'Kakovost :index',
   noContentTypesAvailable: 'Na voljo ni nobenega tipa vsebine',
-  noContentTypesAvailableDesc: 'Vaše spletno mesto ima težave pri povezovanju s H5P.org in pripravo seznama razpoložljivih tipov vsebine.',
+  noContentTypesAvailableDesc:
+    'Vaše spletno mesto ima težave pri povezovanju s H5P.org in pripravo seznama razpoložljivih tipov vsebine.',
   contentTypeCacheOutdated: 'Seznam tipov vsebine je zastarel',
-  contentTypeCacheOutdatedDesc: 'Vaše spletno mesto ima težave pri povezovanju s H5P.org in preverjanju posodobitev za tipe vsebine. Morda ne boste mogli posodobiti ali namestiti novih tipov vsebine.',
+  contentTypeCacheOutdatedDesc:
+    'Vaše spletno mesto ima težave pri povezovanju s H5P.org in preverjanju posodobitev za tipe vsebine. Morda ne boste mogli posodobiti ali namestiti novih tipov vsebine.',
   tryAgain: 'Poskusite znova',
   getHelp: 'Pridobi pomoč',
   untitled: 'Brez naslova :libraryTitle',
   title: 'Naslov',
   metadata: 'Metapodatki',
   addTitle: 'Dodaj naslov',
-  usedForSearchingReportsAndCopyrightInformation: 'Uporablja se za iskanje, poročila in informacije o avtorskih pravicah',
+  usedForSearchingReportsAndCopyrightInformation:
+    'Uporablja se za iskanje, poročila in informacije o avtorskih pravicah',
   metadataSharingAndLicensingInfo: 'Metapodatki (informacije o skupni rabi in licenciranju)',
-  fillInTheFieldsBelow : 'Izpolnite spodnja polja',
+  fillInTheFieldsBelow: 'Izpolnite spodnja polja',
   saveMetadata: 'Shrani metapodatke',
   addAuthor: 'Shrani avtorja',
   confirmRemoveAuthor: 'Ali ste prepričani, da želite odstraniti tega avtorja?',
   addNewChange: 'Dodaj nov zapis',
   confirmDeleteChangeLog: 'Ali ste prepričani, da želite odstraniti zapis te spremembe?',
-  changelogDescription: 'Nekatere licence zahtevajo, da se beležijo in prikazujejo vse spremembe prvotnega dela ali njegovih različic. Na tem mestu lahko beležite svoje spremembe iz razloga izdaje licence ali zgolj zato, da sebi in drugim omogočite lažje sledenje sprememb v tej vsebini.',
+  changelogDescription:
+    'Nekatere licence zahtevajo, da se beležijo in prikazujejo vse spremembe prvotnega dela ali njegovih različic. Na tem mestu lahko beležite svoje spremembe iz razloga izdaje licence ali zgolj zato, da sebi in drugim omogočite lažje sledenje sprememb v tej vsebini.',
   logThisChange: 'Shrani spremembo',
   newChangeHasBeenLogged: 'Nova sprememba je zabeležena',
   loggedChanges: 'Zabeležene spremembe',
   noChangesHaveBeenLogged: 'Ni zabeleženih sprememb',
   errorHeader: 'Prišlo je do napake',
-  errorCalculatingMaxScore: 'Najboljšega rezultata za to vsebino ni mogoče izračunati. Predvideva se, da je število enako 0. Če to ni pravilno, se obrnite na svojega administratorja.',
+  errorCalculatingMaxScore:
+    'Najboljšega rezultata za to vsebino ni mogoče izračunati. Predvideva se, da je število enako 0. Če to ni pravilno, se obrnite na svojega administratorja.',
   maxScoreSemanticsMissing: 'V vsebini ni bilo mogoče najti pričakovane semantike.',
   copyButton: 'Kopiraj',
   copiedButton: 'Kopirano',
   pasteButton: 'Prilepi',
   pasteAndReplaceButton: 'Prilepi & Zamenjaj',
   pasteContent: 'Zamenjaj vsebino',
-  confirmPasteContent: 'S tem boste trenutno vsebino nadomestili z vsebino iz odložišča in bo izgubljena. Ali ste prepričani, da želite nadaljevati?',
+  confirmPasteContent:
+    'S tem boste trenutno vsebino nadomestili z vsebino iz odložišča in bo izgubljena. Ali ste prepričani, da želite nadaljevati?',
   confirmPasteButtonText: 'Zamenjaj vsebino',
   copyToClipboard: 'Kopiraj vsebino H5P v odložišče',
   copiedToClipboard: 'Vsebina H5P je kopirana v odložišče',
@@ -195,10 +206,12 @@ H5PEditor.language.core = {
   pasteError: 'Iz odložišča ni mogoče prilepiti vsebine H5P',
   pasteContentNotSupported: 'Vsebina H5P v odložišču v tej različici ni podprta',
   pasteContentRestricted: 'Vsebina H5P v odložišču ima na tej strani omejen dostop',
-  pasteTooOld: 'Vsebina H5P v odložišču je pripravljena v nižji različici (:clip), kot je podprta tukaj (:local). Če je mogoče, poskušajte najprej vsebino nadgraditi in jo šele nato prilepiti sem.',
-  pasteTooNew: 'Vsebina H5P v odložišču je pripravljena v višji različici (:clip), kot je podprta tukaj (:local). Če je mogoče, poskušajte najprej nadgraditi to vsebino in šele nato prilepiti sem želeno.',
+  pasteTooOld:
+    'Vsebina H5P v odložišču je pripravljena v nižji različici (:clip), kot je podprta tukaj (:local). Če je mogoče, poskušajte najprej vsebino nadgraditi in jo šele nato prilepiti sem.',
+  pasteTooNew:
+    'Vsebina H5P v odložišču je pripravljena v višji različici (:clip), kot je podprta tukaj (:local). Če je mogoče, poskušajte najprej nadgraditi to vsebino in šele nato prilepiti sem želeno.',
   ok: 'V redu',
-  avTablistLabel: 'Insert using',
+  avTablistLabel: 'Vstavljanje',
   tabTitleBasicFileUpload: 'Naloži datoteko',
   tabTitleInputLinkURL: 'Povezava/URL',
   errorTooHighVersion: 'Parametri vsebujejo %used, medtem ko so podprti samo %supported ali predhodni.',
@@ -209,12 +222,86 @@ H5PEditor.language.core = {
   language: 'Jezik',
   noLanguagesSupported: 'Jeziki niso podprti',
   changeLanguage: 'Spremenim jezik v :language?',
-  thisWillPotentially: "To bi lahko vplivalo na ponastavitev vseh preglasitev besed in prevodov, medtem ko bo sama vsebina ostala nespremenjena. Tega ne morete razveljaviti. Ali želite nadaljevati?",
+  thisWillPotentially:
+    'To bi lahko vplivalo na ponastavitev vseh preglasitev besed in prevodov, medtem ko bo sama vsebina ostala nespremenjena. Tega ne morete razveljaviti. Ali želite nadaljevati?',
   notAllTextsChanged: 'Vse besede niso bile spremenjene, saj manjkajo prevodi za jezik :language.',
-  contributeTranslations: 'Več o prispevanju prevodov za jezik :language najdete na <a href=":url" target="_new">spletni strani H5P</a>',
-  unknownLibrary: 'Na žalost izbran tip vsebine \'%lib\' ni nameščen na tem spletnem mestu',
+  contributeTranslations:
+    'Več o prispevanju prevodov za jezik :language najdete na <a href=":url" target="_new">spletni strani H5P</a>',
+  unknownLibrary: "Na žalost izbran tip vsebine '%lib' ni nameščen na tem spletnem mestu",
   proceedButtonLabel: 'Shrani',
   enterFullscreenButtonLabel: 'Vklopi celozaslonski način',
   exitFullscreenButtonLabel: 'Izklopi celozaslonski način',
-  reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  reuseSuccess: ':title uspešno uvožen iz središča H5P Hub.',
+  noContentHeader: 'Ni ustrezne vsebine?',
+  noContentSuggestion: 'Ustvarite svojo!',
+  tutorials: 'Vodiči',
+  contentSectionAll: 'Vsa deljena vsebina',
+  popularContent: 'Priljubljena vsebina',
+  allPopular: 'Vse priljubljene',
+  newOnTheHub: 'Novo na H5P Hub',
+  allNew: 'Vse novo',
+  filterBy: 'Filtriraj po',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Stopnja',
+      dialogHeader: 'Izberi stopnjo izobrazbe',
+      dialogButtonLabel: 'Filtriraj po stopnji izobrazbe',
+    },
+    language: {
+      dropdownLabel: 'Jezik',
+      dialogHeader: 'Izberi jezik',
+      dialogButtonLabel: 'Filtriraj po jeziku',
+      searchPlaceholder: 'Vnos jezika za iskanje',
+    },
+    reviewed: {
+      dropdownLabel: 'Pregledano',
+      dialogHeader: 'Pregledana vsebina',
+      dialogButtonLabel: 'Filtriraj pregledano',
+      optionLabel: 'Prikaži samo pregledano vsebino',
+    },
+    contentTypes: {
+      dropdownLabel: 'Tip vsebine',
+      dialogHeader: 'Izberi tip vsebine',
+      dialogButtonLabel: 'Filtriraj po tipu vsebine',
+      searchPlaceholder: 'Vnos tipa vsebine za iskanje',
+    },
+    disciplines: {
+      dropdownLabel: 'Področje',
+      dialogHeader: 'Izberi področje',
+      dialogButtonLabel: 'Filtriraj po področju',
+      searchPlaceholder: 'Vnos področja za iskanje',
+    },
+    licenses: {
+      dropdownLabel: 'Licence',
+      dialogHeader: 'Izberi licenco uporabe',
+      dialogButtonLabel: 'Filtriraj po licenci',
+      options: { modified: 'Dovoljene spremembe', commercial: 'Dovoljena komercialna raba' },
+    },
+  },
+  clearFilters: 'Počisti vse filtre',
+  contentSearchFieldPlaceholder: 'Iskanje vsebine',
+  loadingContentTitle: 'Nalagamo vsebino ...',
+  loadingContentSubtitle: 'Prosim, počakajte',
+  by: 'Od',
+  dropdownButton: 'Odpri spustni seznam',
+  paginationNavigation: 'Navigacija po straneh',
+  page: 'Stran',
+  currentPage: 'Trenutna stran',
+  nextPage: 'Pojdi na naslednjo stran',
+  previousPage: 'Pojdi na prejšnjo stran',
+  contentPreviewButtonLabel: 'Predogled',
+  contentDownloadButtonLabel: 'Pridobi vsebino',
+  reuseContentTabLabel: 'Pridobite vsebino v skupni rabi',
+  contentPublisherPanelHeader: 'Informacije o založniku',
+  noContentFoundDesc: 'Ni vsebine, ki bi ustrezala iskalnim kriterijem.',
+  h5pType: 'Tip H5P',
+  level: 'Stopnja',
+  size: 'Velikost',
+  failedFetchingData: 'Pridobivanje podatkov ni uspelo',
+  filterErrorMessage: 'Nekaj je šlo narobe. Ponovno naložite stran.',
+  in: 'v',
+  navigateToParent: 'Pomakni se do starša',
+  a11yTitleShowLabel: 'Pokaži za bralnik zaslona',
+  a11yTitleHideLabel: 'Skrij za bralnik zaslona',
 };
