@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: 'Lägg till :entity',
   tooLong: 'Fältvärdet är för långt, får max innehålla :max tecken eller färre.',
   invalidFormat: 'Fältvärdet innehåller ett ogiltigt format eller tecken som är förbjudna.',
-  confirmChangeLibrary: 'Om du gör detta så kommer du förlora allt du gjort med nuvarande innehållstyp. Är du säker på att du vill byta innehållstyp?',
+  confirmChangeLibrary:
+    'Om du gör detta så kommer du förlora allt du gjort med nuvarande innehållstyp. Är du säker på att du vill byta innehållstyp?',
   commonFields: 'Textersättningar och översättningar',
   commonFieldsDescription: 'Här kan du ändra eller översätta texter som används i detta innehåll.',
   uploading: 'Laddar upp, vänta ...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Klistra in YouTube-länk eller annan URL till videokälla',
   uploadAudioTitle: 'Ladda upp ljudfil',
   uploadVideoTitle: 'Ladda upp videofil',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Infoga',
   cancel: 'Avbryt',
   height: 'höjd',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Ladda upp',
   uploadPlaceholder: 'Ingen fil vald',
   uploadInstructionsTitle: 'Ladda upp en H5P-fil.',
-  uploadInstructionsContent: 'Du kan börja med exempel från <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Du kan börja med exempel från <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Ladda upp en fil',
   uploadFileButtonChangeLabel: 'Byt ut fil',
   uploadingThrobber: 'Laddar upp ...',
@@ -84,11 +87,10 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Den valda filen kunde inte laddas upp',
   h5pFileWrongExtensionContent: 'Endast filer med filändelsen .h5p tillåts.',
   h5pFileValidationFailedTitle: 'Kunde inte validera H5P-fil.',
-  h5pFileValidationFailedContent: 'Kontrollera att uppladdad H5P innehåller giltigt H5P innehåll. H5P' +
-  ' filer som enbart innehåller bibliotek ska laddas upp via sidan H5P-bibliotek.',
+  h5pFileValidationFailedContent:
+    'Kontrollera att uppladdad H5P innehåller giltigt H5P innehåll. H5P filer som enbart innehåller bibliotek ska laddas upp via sidan H5P-bibliotek.',
   h5pFileUploadServerErrorTitle: 'H5P-filen kunde inte laddas upp',
-  h5pFileUploadServerErrorContent: 'Ett fel har uppstått. Kontrollera din fellogg på servern för' +
-  ' mer information.',
+  h5pFileUploadServerErrorContent: 'Ett fel har uppstått. Kontrollera din fellogg på servern för mer information.',
   contentTypeSectionAll: 'Alla innehållstyper',
   searchResults: 'Sökrsultat',
   contentTypeSearchFieldPlaceholder: 'Sök innehållstyper',
@@ -118,7 +120,8 @@ H5PEditor.language.core = {
   readLess: 'Läs mindre',
   contentTypeOwner: 'Av :owner',
   contentTypeUnsupportedApiVersionTitle: 'Denna innehållstyp kräver en nyare core-version',
-  contentTypeUnsupportedApiVersionContent: 'Kontakta din systemadministratör för att få tillgång till nödvändiga uppdateringar',
+  contentTypeUnsupportedApiVersionContent:
+    'Kontakta din systemadministratör för att få tillgång till nödvändiga uppdateringar',
   contentTypeUpdateAvailable: 'Uppdatering tillgänglig',
   contentTypeRestricted: 'Begränsad innehållstyp',
   contentTypeRestrictedDesc: 'Användningen av denna innehållstyp har begränsats av en administratör.',
@@ -128,11 +131,13 @@ H5PEditor.language.core = {
   currentMenuSelected: 'nuvarande urval',
   errorCommunicatingHubTitle: 'Kan inte kommunicera med hubben.',
   errorCommunicatingHubContent: 'Ett fel uppstod. Försök igen.',
-  warningNoContentTypesInstalled: "Du har inte någon innehållstyp installerad.",
-  warningChangeBrowsingToSeeResults: 'Klicka <em>Alla</em> för att få listan med alla innehållstyper som du kan installera.',
+  warningNoContentTypesInstalled: 'Du har inte någon innehållstyp installerad.',
+  warningChangeBrowsingToSeeResults:
+    'Klicka <em>Alla</em> för att få listan med alla innehållstyper som du kan installera.',
   warningUpdateAvailableTitle: 'En ny version av :contentType är tillgänglig.',
   warningUpdateAvailableBody: 'Uppdatera till senaste version för en förbättrad upplevelse.',
-  licenseDescription: 'Vissa egenskaper för denna licens framgår nedan. Klicka på info-ikonen ovanför för att läsa licenstexten i original.',
+  licenseDescription:
+    'Vissa egenskaper för denna licens framgår nedan. Klicka på info-ikonen ovanför för att läsa licenstexten i original.',
   licenseModalTitle: 'Licensuppgifter',
   licenseModalSubtitle: 'Välj en licens för att ta del av information om korrekt användning',
   licenseUnspecified: 'Ospecificerad',
@@ -152,12 +157,15 @@ H5PEditor.language.core = {
   screenshots: 'Skärmbilder',
   reloadButtonLabel: 'Ladda om',
   videoQuality: 'Etikett för videokvalitet',
-  videoQualityDescription: 'Denna etikett underlättar för användaren att identifiera nuvarande kvalitet på videon. T ex 1080p, 720p, HD eller mobil',
+  videoQualityDescription:
+    'Denna etikett underlättar för användaren att identifiera nuvarande kvalitet på videon. T ex 1080p, 720p, HD eller mobil',
   videoQualityDefaultLabel: 'Kvalitet :index',
   noContentTypesAvailable: 'Inga innehållstyper är tillgängliga',
-  noContentTypesAvailableDesc: 'Din webbplats har problem med att ansluta till H5P.org och listan över tillgängliga innehållstyper.',
+  noContentTypesAvailableDesc:
+    'Din webbplats har problem med att ansluta till H5P.org och listan över tillgängliga innehållstyper.',
   contentTypeCacheOutdated: 'Lista över innehållstyper behöver uppdateras.',
-  contentTypeCacheOutdatedDesc: 'Din webbplats har problem med att ansluta till H5P.org för att leta efter uppdateringar av innehållstyper. Det kan hända att du inte kan uppdatera eller installera nya innehållstyper.',
+  contentTypeCacheOutdatedDesc:
+    'Din webbplats har problem med att ansluta till H5P.org för att leta efter uppdateringar av innehållstyper. Det kan hända att du inte kan uppdatera eller installera nya innehållstyper.',
   tryAgain: 'Försök igen',
   getHelp: 'Få hjälp',
   untitled: 'Utan titel :libraryTitle',
@@ -166,26 +174,29 @@ H5PEditor.language.core = {
   addTitle: 'Lägg till titel',
   usedForSearchingReportsAndCopyrightInformation: 'Används för sökning, rapporter och copyright-information',
   metadataSharingAndLicensingInfo: 'Metadata (info om delning och licensiering)',
-  fillInTheFieldsBelow : 'Fyll i fälten nedan',
+  fillInTheFieldsBelow: 'Fyll i fälten nedan',
   saveMetadata: 'Spara metadata',
   addAuthor: 'Spara författare',
   confirmRemoveAuthor: 'Är du säker på att du vill ta bort denna författare?',
   addNewChange: 'Lägg till ny ändring',
   confirmDeleteChangeLog: 'Är du säker på att du vill radera denna post i ändringsloggen?',
-  changelogDescription: 'Vissa licenser kräver att ändringar till originalet, eller derivat, loggas och visas. Du kan logga dina ändringar här av licensskäl eller bara för din egen eller andras skull.',
+  changelogDescription:
+    'Vissa licenser kräver att ändringar till originalet, eller derivat, loggas och visas. Du kan logga dina ändringar här av licensskäl eller bara för din egen eller andras skull.',
   logThisChange: 'Logga denna ändring',
   newChangeHasBeenLogged: 'Ny ändring har loggats',
   loggedChanges: 'Loggade ändringar',
   noChangesHaveBeenLogged: 'Inga ändringar har loggats',
   errorHeader: 'Ett fel uppstod',
-  errorCalculatingMaxScore: 'Kunde inte beräkna maxpoäng för detta innehåll. Maxpoäng förväntas vara 0. Kontakta din administratör om detta inte är korrekt.',
+  errorCalculatingMaxScore:
+    'Kunde inte beräkna maxpoäng för detta innehåll. Maxpoäng förväntas vara 0. Kontakta din administratör om detta inte är korrekt.',
   maxScoreSemanticsMissing: 'Kunde inte hitta förväntad semantik i detta innehåll.',
   copyButton: 'Kopiera',
   copiedButton: 'Kopierad',
   pasteButton: 'Klistra in',
   pasteAndReplaceButton: 'Klistra in & ersätt',
   pasteContent: 'Ersätt innehåll',
-  confirmPasteContent: 'Om du gör detta så kommer du ersätta befintligt innehåll med ditt kopierade innehåll från urklipp. Nuvarande innehåll kommer att försvinna. Är du säker på att du vill fortsätta?',
+  confirmPasteContent:
+    'Om du gör detta så kommer du ersätta befintligt innehåll med ditt kopierade innehåll från urklipp. Nuvarande innehåll kommer att försvinna. Är du säker på att du vill fortsätta?',
   confirmPasteButtonText: 'Ersätt innehåll',
   copyToClipboard: 'Kopiera H5P-innehåll till urklipp',
   copiedToClipboard: 'Innehåll är kopierat till urklipp',
@@ -195,8 +206,10 @@ H5PEditor.language.core = {
   pasteError: 'Kan inte klistra in från urklipp',
   pasteContentNotSupported: 'Innehållet i urklipp stöds inte i detta sammanhang',
   pasteContentRestricted: 'Innehållet i urklipp har begränsats på denna webbplats',
-  pasteTooOld: 'Innehållet i urklipp har en lägre version (:clip) än vad som stöds i detta sammanhang (:local), om möjligt, uppgradera innehållet du vill klistra in och försök sen igen.',
-  pasteTooNew: 'Innehållet i urklipp har en högre version (:clip) än vad som stöds i detta sammanhang (:local), om möjligt, uppgradera innehållet här först och försök sen igen.',
+  pasteTooOld:
+    'Innehållet i urklipp har en lägre version (:clip) än vad som stöds i detta sammanhang (:local), om möjligt, uppgradera innehållet du vill klistra in och försök sen igen.',
+  pasteTooNew:
+    'Innehållet i urklipp har en högre version (:clip) än vad som stöds i detta sammanhang (:local), om möjligt, uppgradera innehållet här först och försök sen igen.',
   ok: 'OK',
   avTablistLabel: 'Infoga genom',
   tabTitleBasicFileUpload: 'Filuppladdning',
@@ -209,12 +222,86 @@ H5PEditor.language.core = {
   language: 'Språk',
   noLanguagesSupported: 'Inga språk stöd',
   changeLanguage: 'Ändra språk till :language?',
-  thisWillPotentially: "Detta kan nollställa alla textändringar och översättningar här. Du kan inte ångra detta. Själva innehållet kommer däremot inte att påverkas. Vill du fortsätta?",
+  thisWillPotentially:
+    'Detta kan nollställa alla textändringar och översättningar här. Du kan inte ångra detta. Själva innehållet kommer däremot inte att påverkas. Vill du fortsätta?',
   notAllTextsChanged: 'Inte alla texter ändrades, det finns endast delvis stöd för :language.',
-  contributeTranslations: 'Om du vill bidra till översättningen av :language så kan du läsa mer här <a href=":url" target="_new">contributing translations to H5P</a>',
-  unknownLibrary: 'Tyvärr, innehållstypen \'%lib\' är inte installerad på denna webbplats.',
+  contributeTranslations:
+    'Om du vill bidra till översättningen av :language så kan du läsa mer här <a href=":url" target="_new">contributing translations to H5P</a>',
+  unknownLibrary: "Tyvärr, innehållstypen '%lib' är inte installerad på denna webbplats.",
   proceedButtonLabel: 'Fortsätt för att spara',
   enterFullscreenButtonLabel: 'Helskärmsläge',
   exitFullscreenButtonLabel: 'Avsluta helskärm',
   reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
+  a11yTitleShowLabel: 'Show label for AT',
+  a11yTitleHideLabel: 'Hide label for AT',
 };
