@@ -27,7 +27,8 @@ H5PEditor.language.core = {
   addEntity: ':entity ekle',
   tooLong: 'Alan değeri çok uzun; en fazla :max harf ya da daha azı olmalı.',
   invalidFormat: 'Alan değerinde izin verilmeyen geçersiz bir format ya da harf var.',
-  confirmChangeLibrary: 'Bunu yaptığınızda mevcut içerik tipiyle yapılmış tüm çalışmanızı yitireceksiniz. İçerik tipini değiştirmek istediğinize emin misiniz?',
+  confirmChangeLibrary:
+    'Bunu yaptığınızda mevcut içerik tipiyle yapılmış tüm çalışmanızı yitireceksiniz. İçerik tipini değiştirmek istediğinize emin misiniz?',
   commonFields: 'Ayarlar ve metinler',
   commonFieldsDescription: 'Burada, bu içerikte kullanılan ayarları düzenleyebilir ya da metinleri çevirebilirsiniz.',
   uploading: 'Yükleniyor, bekleyin ...',
@@ -54,7 +55,8 @@ H5PEditor.language.core = {
   enterVideoTitle: 'Youtube linkini ya da diğer video kaynağının URLsini yapıştırın',
   uploadAudioTitle: 'Ses dosyası yükle',
   uploadVideoTitle: 'Video dosyası yükle',
-  addVideoDescription: 'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
+  addVideoDescription:
+    'H5P supports all external video sources formatted as mp4, webm or ogv, like Vimeo Pro, and has support for YouTube and Panopto links.',
   insert: 'Ekle',
   cancel: 'Vazgeç',
   height: 'yükseklik',
@@ -74,7 +76,8 @@ H5PEditor.language.core = {
   uploadTabLabel: 'Yükle',
   uploadPlaceholder: 'Herhangi bir dosya seçilmedi',
   uploadInstructionsTitle: 'Bir H5P dosyası yükle.',
-  uploadInstructionsContent: 'Şu adresteki örneklerle başlayabilirsiniz <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
+  uploadInstructionsContent:
+    'Şu adresteki örneklerle başlayabilirsiniz <a href="https://h5p.org/content-types-and-applications" target="blank">H5P.org</a>.',
   uploadFileButtonLabel: 'Bir dosya yükle',
   uploadFileButtonChangeLabel: 'Dosyayı değiştir',
   uploadingThrobber: 'Şu anda yükleniyor...',
@@ -84,9 +87,11 @@ H5PEditor.language.core = {
   h5pFileWrongExtensionTitle: 'Seçilen dosya yüklenemedi',
   h5pFileWrongExtensionContent: 'Sadece .h5p uzantılı dosyalar kabul edilmektedir.',
   h5pFileValidationFailedTitle: 'H5P dosyası doğrulanamamıştır.',
-  h5pFileValidationFailedContent: 'Yüklenen H5P\'nin geçerli H5P içeriklerini içerdiğinden emin olun. H5P dosyaları sadece H5P Kütüphaneleri (H5P Libraries) sayfası aracılığıyla yüklenen kütüphaneleri içerebilir.',
+  h5pFileValidationFailedContent:
+    "Yüklenen H5P'nin geçerli H5P içeriklerini içerdiğinden emin olun. H5P dosyaları sadece H5P Kütüphaneleri (H5P Libraries) sayfası aracılığıyla yüklenen kütüphaneleri içerebilir.",
   h5pFileUploadServerErrorTitle: 'H5P dosyası yüklenemedi',
-  h5pFileUploadServerErrorContent: 'Beklenmeyen bir hata oluştu. Daha detaylı bilgi için sunucunuzun hata log kayıtlarını inceleyin.',
+  h5pFileUploadServerErrorContent:
+    'Beklenmeyen bir hata oluştu. Daha detaylı bilgi için sunucunuzun hata log kayıtlarını inceleyin.',
   contentTypeSectionAll: 'Tüm İçerik Türleri',
   searchResults: 'Arama Sonuçları',
   contentTypeSearchFieldPlaceholder: 'İçerik Türü Ara',
@@ -109,7 +114,7 @@ H5PEditor.language.core = {
   recentlyUsedFirst: 'Son Kullanılanlar',
   popularFirst: 'Popülerler',
   newestFirst: 'Yeniler',
-  aToZ: 'A\'dan Z\'ye',
+  aToZ: "A'dan Z'ye",
   noResultsFound: 'Hiçbir sonuç bulunamadı',
   noResultsFoundDesc: 'Arama kriterlerinizle eşleşen içerik türü bulunamadı.',
   readMore: 'Daha fazla oku',
@@ -130,7 +135,8 @@ H5PEditor.language.core = {
   warningChangeBrowsingToSeeResults: 'Yükleyebileceğiniz tüm içerik türlerinin listesi için <em>Hepsi</em>ne tıklayın.',
   warningUpdateAvailableTitle: ':contentType yeni bir versiyonu mevcut.',
   warningUpdateAvailableBody: 'Gelişmiş deneyim için son versiyona güncelleyin.',
-  licenseDescription: 'Bu lisansın bazı özellikleri aşağıda belirtilmiştir. Orijinal lisans metnini okumak için yukarıdaki bilgi ikonuna tıklayın.',
+  licenseDescription:
+    'Bu lisansın bazı özellikleri aşağıda belirtilmiştir. Orijinal lisans metnini okumak için yukarıdaki bilgi ikonuna tıklayın.',
   licenseModalTitle: 'Lisans Detayları',
   licenseModalSubtitle: 'Kurallara uygun kullanım hakkında bilgiyi görüntülemek için bir lisans seçin',
   licenseUnspecified: 'Tanımlanmamış',
@@ -144,18 +150,21 @@ H5PEditor.language.core = {
   licenseMustIncludeLicense: 'Lisans içermelidir',
   licenseFetchDetailsFailed: 'Lisans ayrıntıları alınamadı',
   imageLightboxTitle: 'Resimler',
-  imageLightBoxProgress: ':total\'dan :num of',
+  imageLightBoxProgress: ":total'dan :num of",
   nextImage: 'Sonraki resim',
   previousImage: 'Önceki resim',
   screenshots: 'Ekran görüntüleri',
   reloadButtonLabel: 'Tekrar yükle',
   videoQuality: 'Video kalite etiketi',
-  videoQualityDescription: 'Bu etiket kullanıcının mevcut videonun kalitesini tanımlamasına yardımcı olur. Örn: 1080p, 720p, HD veya Mobil',
+  videoQualityDescription:
+    'Bu etiket kullanıcının mevcut videonun kalitesini tanımlamasına yardımcı olur. Örn: 1080p, 720p, HD veya Mobil',
   videoQualityDefaultLabel: 'Kalite :index',
   noContentTypesAvailable: 'Hiçbir içerik mevcut değil',
-  noContentTypesAvailableDesc: 'Web siteniz H5P.org sitesine bağlanmada ve mevcut içerik türlerini listelemede zorluk yaşıyor.',
+  noContentTypesAvailableDesc:
+    'Web siteniz H5P.org sitesine bağlanmada ve mevcut içerik türlerini listelemede zorluk yaşıyor.',
   contentTypeCacheOutdated: 'İçerik türü listesi güncel değil',
-  contentTypeCacheOutdatedDesc: 'Web siteniz mevcut içerik türü güncellemelerini kontrol etmek için H5P.org sitesine bağlanmada zorluk yaşıyor. Yeni içerik türlerini yükleme ve güncellemek için yetkiniz olmayabilir.',
+  contentTypeCacheOutdatedDesc:
+    'Web siteniz mevcut içerik türü güncellemelerini kontrol etmek için H5P.org sitesine bağlanmada zorluk yaşıyor. Yeni içerik türlerini yükleme ve güncellemek için yetkiniz olmayabilir.',
   tryAgain: 'Yeniden dene',
   getHelp: 'Yardım al',
   untitled: 'Başlıksız :libraryTitle',
@@ -170,20 +179,23 @@ H5PEditor.language.core = {
   confirmRemoveAuthor: 'Bu yazarı kaldırmak istediğinizden emin misiniz?',
   addNewChange: 'Yeni değişiklik ekle',
   confirmDeleteChangeLog: 'Bu değişim log kaydını silmek istediğinizden emin misiniz?',
-  changelogDescription: 'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
+  changelogDescription:
+    'Some licenses require that changes made to the original work, or derivatives are logged and displayed. You may log your changes here for licensing reasons or just to allow yourself and others to keep track of the changes made to this content.',
   logThisChange: 'Bu değişimi logla',
   newChangeHasBeenLogged: 'Yeni değişiklik loglandı',
   loggedChanges: 'Loglanmış değişiklikler',
   noChangesHaveBeenLogged: 'Hiçbir değişiklik loglanmadı',
   errorHeader: 'Bir hata oluştu',
-  errorCalculatingMaxScore: 'Bu içerik için maksimum puan hesaplanamadı. Maksimum puan 0 olarak kabul edilir. Doğru değilse yöneticinize başvurun.',
+  errorCalculatingMaxScore:
+    'Bu içerik için maksimum puan hesaplanamadı. Maksimum puan 0 olarak kabul edilir. Doğru değilse yöneticinize başvurun.',
   maxScoreSemanticsMissing: 'İçerikte beklenen anlam bulunamadı.',
   copyButton: 'Kopyala',
   copiedButton: 'Kopyalandı',
   pasteButton: 'Yapıştır',
   pasteAndReplaceButton: 'Yapıştır & Değiştir',
   pasteContent: 'İçeriği Değiştir',
-  confirmPasteContent: 'Bunu yaparak, mevcut içeriği panodaki içerikle değiştirirsiniz. Mevcut içerik kaybedilecek. Devam etmek istediğinize emin misiniz?',
+  confirmPasteContent:
+    'Bunu yaparak, mevcut içeriği panodaki içerikle değiştirirsiniz. Mevcut içerik kaybedilecek. Devam etmek istediğinize emin misiniz?',
   confirmPasteButtonText: 'İçeriği değiştir',
   copyToClipboard: 'H5P içeriğini panoya kopyalayın',
   copiedToClipboard: 'İçerik panoya kopyalandı',
@@ -193,8 +205,10 @@ H5PEditor.language.core = {
   pasteError: 'Panodan H5P içeriği yapıştırılamadı',
   pasteContentNotSupported: 'H5P panosundaki içerik bu bağlamda desteklenmiyor',
   pasteContentRestricted: 'The content in the clipboard has been restricted on this site',
-  pasteTooOld: 'H5P panosundaki içerik, bu bağlamda (:local) desteklenenden daha düşük bir sürümdür (:clip) mümkünse yapıştırmak istediğiniz içeriğin yükseltilmesini sağlayın , tekrar kopyalayın ve buraya yapıştırmayı deneyin.',
-  pasteTooNew: 'H5P panosundaki içerik, bu bağlamda (:local) desteklenenden daha yüksek bir versiyona (:clip) sahiptir, mümkünse önce bu içeriğin yükseltilmesini sağlayın ve ardından içeriği buraya yapıştırmayı deneyin.',
+  pasteTooOld:
+    'H5P panosundaki içerik, bu bağlamda (:local) desteklenenden daha düşük bir sürümdür (:clip) mümkünse yapıştırmak istediğiniz içeriğin yükseltilmesini sağlayın , tekrar kopyalayın ve buraya yapıştırmayı deneyin.',
+  pasteTooNew:
+    'H5P panosundaki içerik, bu bağlamda (:local) desteklenenden daha yüksek bir versiyona (:clip) sahiptir, mümkünse önce bu içeriğin yükseltilmesini sağlayın ve ardından içeriği buraya yapıştırmayı deneyin.',
   ok: 'TAMAM',
   avTablistLabel: 'Kullanarak ekle',
   tabTitleBasicFileUpload: 'Dosya Yükleme',
@@ -207,14 +221,86 @@ H5PEditor.language.core = {
   language: 'Dil',
   noLanguagesSupported: 'Desteklenen Dil Yok',
   changeLanguage: 'Dili değiştir :language?',
-  thisWillPotentially: 'Bu, tüm metni ve çevirileri potansiyel olarak sıfırlayacaktır. Bu değişikliği sıfırlayamazsınız. İçeriğin kendisi değişmeyecek. Devam etmek istiyor musunuz?',
+  thisWillPotentially:
+    'Bu, tüm metni ve çevirileri potansiyel olarak sıfırlayacaktır. Bu değişikliği sıfırlayamazsınız. İçeriğin kendisi değişmeyecek. Devam etmek istiyor musunuz?',
   notAllTextsChanged: 'Metinlerin tümü değişmedi, yalnızca kısmi kapsama alanı var :language.',
-  contributeTranslations: 'Çeviriyi tamamlamak istiyorsanız :language hakkında bilgi edinebilirsiniz <a href=":url" target="_new">H5P’ye çevirilerde katkıda bulunmak</a>',
-  unknownLibrary: 'Ne yazık ki, seçilen içerik türü \'%lib \' bu sisteme yüklenmedi.',
+  contributeTranslations:
+    'Çeviriyi tamamlamak istiyorsanız :language hakkında bilgi edinebilirsiniz <a href=":url" target="_new">H5P’ye çevirilerde katkıda bulunmak</a>',
+  unknownLibrary: "Ne yazık ki, seçilen içerik türü '%lib ' bu sisteme yüklenmedi.",
   proceedButtonLabel: 'Kaydetmeye devam et',
   enterFullscreenButtonLabel: 'Tam ekran yap',
   exitFullscreenButtonLabel: 'Tam ekrandan çık',
   a11yTitleShowLabel: 'Show label for AT',
   a11yTitleHideLabel: 'Hide label for AT',
-  reuseSuccess: ':title was successfully imported from the H5P Hub.'
+  reuseSuccess: ':title was successfully imported from the H5P Hub.',
+  noContentHeader: 'No Fitting Content?',
+  noContentSuggestion: 'Create one yourself!',
+  tutorials: 'Tutorials',
+  contentSectionAll: 'All shared content',
+  popularContent: 'Popular Content',
+  allPopular: 'All Popular',
+  newOnTheHub: 'New on the Hub',
+  allNew: 'All New',
+  filterBy: 'Filter by',
+  filter: 'Filter',
+  filters: {
+    level: {
+      dropdownLabel: 'Level',
+      dialogHeader: 'Select level of education',
+      dialogButtonLabel: 'Filter level of education',
+    },
+    language: {
+      dropdownLabel: 'Language',
+      dialogHeader: 'Select language(s)',
+      dialogButtonLabel: 'Filter languages',
+      searchPlaceholder: 'Type to search for languages',
+    },
+    reviewed: {
+      dropdownLabel: 'Reviewed',
+      dialogHeader: 'Reviewed Content',
+      dialogButtonLabel: 'Filter',
+      optionLabel: 'Show only reviewed content',
+    },
+    contentTypes: {
+      dropdownLabel: 'Content types',
+      dialogHeader: 'Select Content type(s)',
+      dialogButtonLabel: 'Filter Content Types',
+      searchPlaceholder: 'Type to search for content types',
+    },
+    disciplines: {
+      dropdownLabel: 'Discipline',
+      dialogHeader: 'Select Your Discipline',
+      dialogButtonLabel: 'Filter Disciplines',
+      searchPlaceholder: 'Type to search for disciplines',
+    },
+    licenses: {
+      dropdownLabel: 'License',
+      dialogHeader: 'Select preferred rights of use',
+      dialogButtonLabel: 'Filter licenses',
+      options: { modified: 'Can be modified', commercial: 'Allows commercial use' },
+    },
+  },
+  clearFilters: 'Clear all filters',
+  contentSearchFieldPlaceholder: 'Search for Content',
+  loadingContentTitle: 'We are loading content for you...',
+  loadingContentSubtitle: 'Please wait',
+  by: 'By',
+  dropdownButton: 'Open dropdown',
+  paginationNavigation: 'Pagination navigation',
+  page: 'Page',
+  currentPage: 'Current page',
+  nextPage: 'Go to next page',
+  previousPage: 'Go to previous page',
+  contentPreviewButtonLabel: 'Preview',
+  contentDownloadButtonLabel: 'Get Content',
+  reuseContentTabLabel: 'Get Shared Content',
+  contentPublisherPanelHeader: 'Publisher Info',
+  noContentFoundDesc: 'There is no content that matches your search criteria.',
+  h5pType: 'H5P Type',
+  level: 'Level',
+  size: 'Size',
+  failedFetchingData: 'Failed fetching data',
+  filterErrorMessage: 'Something went wrong. Please reload the page.',
+  in: 'in',
+  navigateToParent: 'Navigate to parent',
 };
